% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_load_maxquant.R
\name{add_maxquant_sdata}
\alias{add_maxquant_sdata}
\title{Add maxquant sdata}
\usage{
add_maxquant_sdata(object, design_file)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{design_file}{sample file}
}
\value{
sumexp with sdata added
}
\description{
Add maxquant sdata
}
