% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rnaseq.R
\name{compute_precision_weights}
\alias{compute_precision_weights}
\alias{add_precision_weights}
\title{Compute (and add) precision weights}
\usage{
compute_precision_weights(object,
  design = autonomics.import::create_design_matrix(object), plot = TRUE)

add_precision_weights(object,
  design = autonomics.import::create_design_matrix(object), plot = TRUE)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{design}{design matrix}

\item{plot}{whether to plot mean-var trend (logical)}

\item{...}{passed to limma::lmFit}
}
\value{
weights vector
}
\description{
Compute (and add) precision weights for SummarizedExperiment with log2cpm values.
}
\details{
Refactored version of limma::voom() which operates on a SummarizedExperiment with log2cpm values.
Created to allow a separation of the two steps in limma::voom():
   1. raw counts -> log2cpm (autonomics.import::counts_to_cpm)
   2. log2cpm    -> weights (autonomics.import::compute_precision_weights)
If a block factor is present, precision weights are updated using the block correlation information,
as suggested by Gordon Brown on BioC support (https://support.bioconductor.org/p/59700/)
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   object <- 'extdata/rnaseq/gene_counts.txt'           \%>\%
              system.file(package = 'subramanian.2016') \%>\%
              autonomics.import::load_rnaseq(infer_design_from_sampleids = TRUE)
   object \%>\% autonomics.import::compute_precision_weights() \%>\% str()
   object \%>\% autonomics.import::add_precision_weights()
   object \%>\% autonomics.import::load_rnaseq()
}

}
\author{
Charity Law, Gordon Smyth, Aditya Bhagwat
}
