% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sample_file.R
\name{add_replicate_values}
\alias{add_replicate_values}
\title{Add replicate values}
\usage{
add_replicate_values(sample_df)
}
\arguments{
\item{sample_df}{dataframe with columns 'sample_id' and 'subgroup'}
}
\description{
Replicate values are set to the unique sampleid tails within each subgroup
}
\examples{
require(magrittr)
sample_df <- data.frame(sample_id = c('E_1', 'E_2', 'E_3', 'EM_1', 'EM_2', 'EM_3'),
                        subgroup  = c('E',   'E',   'E',   'EM',   'EM',   'EM')) \%>\%
             magrittr::set_rownames(.$sample_id)
sample_df
sample_df \%>\% add_replicate_values()
}
