% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_sdata.R
\name{add_sdata}
\alias{add_sdata}
\title{Add sdata}
\usage{
add_sdata(object, sample_file)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{sample_file}{sample file}
}
\value{
sumexp with sdata added
}
\description{
Add sdata
}
