% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{add_smiles_to_fdata}
\alias{add_smiles_to_fdata}
\title{Map pubchem IDs to canonical smiles}
\usage{
add_smiles_to_fdata(object)
}
\arguments{
\item{object}{SummarizedExperiment}
}
\value{
named character vector (elements = SMILES, names = PUBCHEMIDS)
}
\description{
Uses \href{https://pubchemdocs.ncbi.nlm.nih.gov/pug-rest-tutorial}{PUBCHEM's REST interface}
}
\examples{
\dontrun{
require(magrittr)
if (require(subramanian.2016)){
   subramanian.2016::metabolon \%>\%
      autonomics.import::add_smiles_to_fdata()
}
}

}
