% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_svalues_available.R
\name{all_are_missing_or_empty_character}
\alias{all_are_missing_or_empty_character}
\title{Are all missing or empty character}
\usage{
all_are_missing_or_empty_character(x)
}
\arguments{
\item{x}{character vector}
}
\value{
logical(1)
}
\description{
Are all missing or empty character
}
\examples{
require(magrittr)
x <- c('a', 'b', NA_character_, 'd', '')
x \%>\% autonomics.support::all_are_missing_or_empty_character()
}
