% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_omics.R
\name{annotate_proteingroups}
\alias{annotate_proteingroups}
\title{Annotate proteingroups through uniprot.ws}
\usage{
annotate_proteingroups(object, connection = NULL,
  fields = c("SUBCELLULAR-LOCATIONS", "INTERPRO", "GO-ID"))
}
\arguments{
\item{object}{SummarizedExperiment}

\item{connection}{uniprot.ws connection}

\item{fields}{uniprot.ws fields to query}
}
\value{
Annotated SummarizedExperiment
}
\description{
Annotate proteingroups through uniprot.ws
}
\examples{
require(magrittr)
\dontrun{
   if (require(autonomics.data)){
      object <- autonomics.data::stemcomp.proteinratios[1:10, ]
      connection <- autonomics.annotate::connect_to_uniprot(c("A0AV96", "A0AVF1", "A0AVT1"))
      object \%>\% annotate_proteingroups(c('SUBCELLULAR-LOCATIONS', 'GO-ID'))
   }
}
}
