% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_maxquant2.R
\name{annotate_proteingroups2}
\alias{annotate_proteingroups2}
\title{Annotate and deconvolute proteingroups SumExp}
\usage{
annotate_proteingroups2(object, fasta_file = NULL)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{fasta_file}{path to fasta file}
}
\description{
Steps:
\enumerate{
   \item Separates proteingroups into uniprot accessions.
   \item Fetches up-to-date annotations for each from uniprot
   \item Keeps best annotated entries (per proteingroup)
   \item Collapses isoforms (per canonical accession)
   \item Keeps first of redundant uniprot entries
   \item Collapses paralogs (per protein group)
   \item Merges annotations into fdata(object) and returns
}
}
\examples{
\dontrun{
require(magrittr)
if (require(autonomics.data)){
   object <- system.file('extdata/billing2016/proteinGroups.txt',
                          package = 'autonomics.data') \%>\%
             autonomics.import::load_proteingroups() \%>\%
             magrittr::extract(1:10, )
   object \%>\% autonomics.import::fdata() \%>\% str()
   object \%>\% autonomics.import::annotate_proteingroups() \%>\%
              autonomics.import::fdata() \%>\% str()
}
}
}
