% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters.R
\docType{methods}
\name{annotation}
\alias{annotation}
\alias{annotation,SummarizedExperiment-method}
\alias{annotation,eSet-method}
\alias{annotation,EList-method}
\alias{annotation<-}
\alias{annotation<-,SummarizedExperiment,character-method}
\alias{annotation<-,eSet,character-method}
\alias{annotation<-,EList,character-method}
\title{Get /set annotation}
\usage{
annotation(object)

\S4method{annotation}{SummarizedExperiment}(object)

\S4method{annotation}{eSet}(object)

\S4method{annotation}{EList}(object)

annotation(object) <- value

\S4method{annotation}{SummarizedExperiment,character}(object) <- value

\S4method{annotation}{eSet,character}(object) <- value

\S4method{annotation}{EList,character}(object) <- value
}
\arguments{
\item{object}{SummarizedExperiment, eSet, or EList}

\item{value}{list}
}
\value{
annotation details (get) or updated eset (set)
}
\description{
Get /set annotation
}
