% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange.R
\name{arrange_features_}
\alias{arrange_features_}
\alias{arrange_features}
\title{Arrange features of eset}
\usage{
arrange_features_(x, fvars)

arrange_features(x, ...)
}
\arguments{
\item{x}{eset}

\item{fvars}{fvars (SE)}

\item{...}{fvars (NSE)}
}
\value{
arranged eset
}
\description{
Arrange eset on specified fvars
}
\examples{
if(require(autonomics.data)){
   head(autonomics.import::fdata(ALL)[, -2])
   head(autonomics.import::fdata(arrange_features(ALL, gene_symbols))[, -2])
   head(autonomics.import::fdata(arrange_features_(ALL, "gene_symbols"))[, -2])
}
}
