% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_filter.R
\name{arrange_on_magnitude}
\alias{arrange_on_magnitude}
\title{Order features in eSet on magnitude of given contrast}
\usage{
arrange_on_magnitude(object, contrast_name)
}
\arguments{
\item{object}{eset}

\item{contrast_name}{contrast_name}
}
\description{
Order features in eSet on magnitude of given contrast
}
