% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_filter.R
\name{arrange_samples_}
\alias{arrange_samples_}
\alias{arrange_samples}
\title{Arrange samples of eset}
\usage{
arrange_samples_(x, svars)

arrange_samples(x, ...)
}
\arguments{
\item{x}{eset}

\item{svars}{svars (SE)}

\item{...}{svars (NSE)}
}
\value{
arranged eset
}
\description{
Arrange eset on specified svars
}
\examples{
if(require(autonomics.data)){
   head(autonomics.import::sdata(ALL)[, -2])
   head(autonomics.import::sdata(arrange_samples(ALL, age)))
   head(autonomics.import::sdata(arrange_samples_(ALL, "age")))
}
}
