% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_has_is.R
\name{assert_all_are_valid_features}
\alias{assert_all_are_valid_features}
\title{Assert that features are valid}
\usage{
assert_all_are_valid_features(features, object)
}
\arguments{
\item{features}{\code{\link{numeric}} index or \code{\link{character}} names
of \code{\link[autonomics.import]{fdata}} entries}

\item{object}{eSet}
}
\description{
Assert that features are valid
}
