% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_has_is.R
\name{assert_is_valid_object}
\alias{assert_is_valid_object}
\title{Assert that x is a valid eSet}
\usage{
assert_is_valid_object(x)
}
\arguments{
\item{x}{eset}
}
\value{
error if not true
}
\description{
Assert that x is a valid eSet
}
