% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rnaseq.R
\name{compute_precision_weights}
\alias{compute_precision_weights}
\title{Compute voom precision weights}
\usage{
compute_precision_weights(object,
  design = autonomics.import::create_design_matrix(object), plot = TRUE)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{design}{design matrix}

\item{plot}{logical}

\item{...}{passed to limma::voom() -> limma::lmFit()}
}
\description{
Compute voom precision weights
}
\examples{
require(magrittr)
if (require(subramanian.2016)){
   object <- 'extdata/rnaseq/gene_counts.txt' \%>\%
              system.file(package = 'subramanian.2016') \%>\%
              autonomics.import::load_rnaseq(infer_design_from_sampleids = TRUE)
   object \%>\% compute_precision_weights()
}
}
