% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_has_is.R
\name{contains_block}
\alias{contains_block}
\title{Does object contain block values?}
\usage{
contains_block(object)
}
\arguments{
\item{object}{SummarizedExperiment}
}
\value{
logical(1)
}
\description{
Does object contain block values?
}
