% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_has_is.R
\name{contains_ratios}
\alias{contains_ratios}
\title{Does object contain ratio values?}
\usage{
contains_ratios(object)
}
\arguments{
\item{object}{SummarizedExperiment}
}
\value{
logical
}
\description{
Does object contain ratio values?
}
\examples{
require(magrittr)
if (require(billing.differentiation.data)){
   billing.differentiation.data::protein.ratios \%>\%
      autonomics.import::contains_ratios()
   billing.differentiation.data::rna.voomcounts \%>\%
      autonomics.import::contains_ratios()
}
}
