% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters.R
\docType{methods}
\name{contrastdefs}
\alias{contrastdefs}
\alias{contrastdefs,SummarizedExperiment-method}
\alias{contrastdefs<-}
\alias{contrastdefs<-,SummarizedExperiment,character-method}
\alias{contrastdefs<-,SummarizedExperiment,NULL-method}
\alias{set_contrastdefs}
\title{Get/Set contrast definitions}
\usage{
contrastdefs(object)

\S4method{contrastdefs}{SummarizedExperiment}(object)

contrastdefs(object) <- value

\S4method{contrastdefs}{SummarizedExperiment,character}(object) <- value

\S4method{contrastdefs}{SummarizedExperiment,`NULL`}(object) <- value

set_contrastdefs(object, value)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{value}{named contrast vector}
}
\value{
updated object
}
\description{
Get/Set contrast definitions
}
\examples{
require(magrittr)
if (require(subramanian.2016)){
   object <- subramanian.2016::metabolon
   contrastdefs <- subramanian.2016::contrasts.metabolon
   autonomics.import::contrastdefs(object)                      # getter
   autonomics.import::contrastdefs(object) <- contrastdefs    # setter (conventional)
   autonomics.import::set_contrastdefs(object, contrastdefs)  # setter (piping)
}
}
