% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters.R
\docType{methods}
\name{contrasts<-}
\alias{contrasts<-}
\alias{contrasts<-,SummarizedExperiment,character-method}
\alias{contrasts<-,SummarizedExperiment,NULL-method}
\alias{contrasts}
\alias{contrasts,SummarizedExperiment-method}
\title{Get/Set contrasts}
\usage{
contrasts(object) <- value

\S4method{contrasts}{SummarizedExperiment,character}(object) <- value

\S4method{contrasts}{SummarizedExperiment,`NULL`}(object) <- value

contrasts(object)

\S4method{contrasts}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{value}{named contrast vector}
}
\value{
updated object
}
\description{
Get/Set contrasts
}
\examples{
require(magrittr)
if (require(subramanian.2016)){
   object <- subramanian.2016::metabolon
   contrasts <- subramanian.2016::contrasts.metabolon
   autonomics.import::contrasts(object)
   autonomics.import::contrasts(object) <- contrasts
}
}
