% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xxx_components.R
\name{count_components}
\alias{count_components}
\alias{count_components.SummarizedExperiment}
\alias{count_components.character}
\title{Count no of components}
\usage{
count_components(x, ...)

\method{count_components}{SummarizedExperiment}(x, svar = "subgroup",
  sep = x \%>\% autonomics.import::slevels(svar) \%>\%
  autonomics.import::guess_component_sep(.))

\method{count_components}{character}(x, sep = x \%>\% as.character() \%>\%
  unique() \%>\% autonomics.import::guess_component_sep(.))
}
\arguments{
\item{x}{character(n) or SummarizedExperiment}

\item{svar}{character(1)}

\item{sep}{character(1)}
}
\value{
integer(1)
}
\description{
Count no of components
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   autonomics.data::glutaminase            \%>\% count_components()
   autonomics.data::glutaminase$subgroup   \%>\% count_components()
}
}
