% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters.R
\name{counts}
\alias{counts}
\alias{counts<-}
\title{Get/Set counts}
\usage{
counts(object)

counts(object) <- value
}
\arguments{
\item{object}{SummarizedExperiment}

\item{value}{count matrix (features x samples)}
}
\value{
exprs matrix (get) or updated object (set)
}
\description{
Get / Set counts matrix
}
