% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rnaseq.R
\name{counts_to_cpm}
\alias{counts_to_cpm}
\title{Convert counts into cpm (counts per million reads)}
\usage{
counts_to_cpm(counts, lib.size = libsizes(counts))
}
\arguments{
\item{counts}{count matrix}

\item{lib.size}{scaled library sizes (vector)}
}
\value{
cpm matrix
}
\description{
Convert counts into cpm (counts per million reads)
}
