% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rnaseq.R
\name{cpm_to_counts}
\alias{cpm_to_counts}
\title{Invert cpm (counts per million reads) into counts}
\usage{
cpm_to_counts(cpm, lib.size)
}
\arguments{
\item{cpm}{cpm matrix}

\item{lib.size}{scaled library sizes (vector)}
}
\value{
count matrix
}
\description{
Invert cpm (counts per million reads) into counts
}
