% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_design_file.R
\name{create_exiqon_design_file}
\alias{create_exiqon_design_file}
\title{Create exiqon design file}
\usage{
create_exiqon_design_file(exiqon_file,
  sample_file = paste0(dirname(proteingroups_file), "/sample_design.txt"),
  infer_design = FALSE)
}
\arguments{
\item{exiqon_file}{string: path to exiqon file}

\item{sample_file}{string: path to sample file}

\item{infer_design}{logical: whether to infer design from sampleids}
}
\value{
string: path to design file
}
\description{
Create exiqon design file
}
\examples{
require(magrittr)
if (require(subramanian.2016){
   exiqon_file <- system.file('extdata/exiqon/subramanian.2016.exiqon.xlsx',
                               package = 'subramanian.2016')
   exiqon_file \%>\% autonomics.import::create_exiqon_design_file(sample_file = tempfile())

}
}
