% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sample_file.R
\name{create_exiqon_sample_df}
\alias{create_exiqon_sample_df}
\title{Create exiqon sample dataframe}
\usage{
create_exiqon_sample_df(exiqon_file, infer_design = FALSE)
}
\arguments{
\item{exiqon_file}{string}

\item{sample_file}{string}

\item{value_type}{string: any value in autonomics.import::MAXQUANT_VALUE_TYPES}
}
\value{
string: path to design file
}
\description{
Create exiqon sample dataframe
}
\examples{
require(magrittr)
if (require(subramanian.2016)){
   exiqon_file <- system.file('extdata/exiqon/subramanian.2016.exiqon.xlsx',
                               package = 'subramanian.2016')
   exiqon_file \%>\% autonomics.import::create_exiqon_sample_df()
   exiqon_file \%>\% autonomics.import::create_exiqon_sample_df(infer_design = TRUE)
}
}
