% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_maxquant.R
\name{create_maxquant_design}
\alias{create_maxquant_design}
\title{Create maxquant design}
\usage{
create_maxquant_design(proteingroups_file,
  value_type = autonomics.import::infer_maxquant_value_type(proteingroups_file))
}
\arguments{
\item{proteingroups_file}{full path to protein groups file}

\item{value_type}{any value in autonomics.import::MAXQUANT_VALUE_TYPES}

\item{sample_file}{full path to sample design file}

\item{...}{backward compatibility to deprecated functions}
}
\value{
sample design dataframe
}
\description{
Create maxquant design
}
\examples{
require(magrittr)

# NORMALIZED RATIOS
if (require(billing.differentiation.data)){
   system.file('extdata/maxquant/proteinGroups.txt',
                package = 'billing.differentiation.data') \%>\%
   autonomics.import::create_maxquant_design()
}

# REPORTER INTENSITIES
if (require(billing.vesicles)){
   system.file('extdata/proteinGroups.txt', package = 'billing.vesicles') \%>\%
   autonomics.import::create_maxquant_design()
}

# LFQ INTENSITIES
if (require(graumann.lfq)){
   system.file('extdata/proteinGroups.txt', package = 'graumann.lfq') \%>\%
   autonomics.import::create_maxquant_design()
}

# LABELED INTENSITIES
if (require(billing.differentiation.data)){
   system.file('extdata/maxquant/proteinGroups.txt',
                package = 'billing.differentiation.data') \%>\%
   autonomics.import::create_maxquant_design(value_type = 'raw.intensity')
}

# UNLABELED INTENSITIES
if (require(alnoubi.2017)){
   system.file('extdata/proteinGroups.txt', package = 'alnoubi.2017') \%>\%
   autonomics.import::create_maxquant_design()
}
}
