% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_maxquant.R
\name{create_sample_design_df}
\alias{create_sample_design_df}
\alias{create_sample_design_file}
\alias{create_maxquant_design_df}
\alias{create_maxquant_design_file}
\title{Create maxquant sample design}
\usage{
create_sample_design_df(...)

create_sample_design_file(...)

create_maxquant_design_df(proteingroups_file,
  value_type = autonomics.import::infer_maxquant_value_type(proteingroups_file))

create_maxquant_design_file(proteingroups_file,
  sample_design_file = paste0(dirname(proteingroups_file),
  "/sample_design.txt"),
  value_type = infer_maxquant_value_type(proteingroups_file))
}
\arguments{
\item{...}{backward compatibility to deprecated functions create_sample_design_df and create_sample_design_file}

\item{proteingroups_file}{full path to protein groups file}

\item{value_type}{any value in autonomics.import::MAXQUANT_VALUE_TYPES}

\item{sample_design_file}{full path to sample design file}
}
\value{
sample design dataframe (create_maxquant_design_df) or file (create_maxquant_design_file)
}
\description{
Create maxquant sample design
}
\examples{
require(magrittr)

# normalized ratios
#------------------
if (require(billing.differentiation.data)){
   proteingroups_file <- system.file('extdata/maxquant/proteinGroups.txt',
                                     package = 'billing.differentiation.data')
   autonomics.import::create_maxquant_design_df(proteingroups_file)
   sample_design_file <- paste0(tempdir(), '/billing.differentiation/sample_design.txt')
   dir.create(dirname(sample_design_file))
   create_maxquant_design_file(proteingroups_file, sample_design_file)
}

# reporter intensities
#---------------------
if (require(billing.vesicles)){
   proteingroups_file <- system.file('extdata/proteinGroups.txt', package = 'billing.vesicles')
   autonomics.import::create_maxquant_design_df(proteingroups_file)
   sample_design_file <- paste0(tempdir(), '/billing.vesicles/sample_design.txt')
   dir.create(dirname(sample_design_file))
   create_maxquant_design_file(proteingroups_file, sample_design_file)
}

# lfq intensities
#----------------
if (require(graumann.lfq)){
   proteingroups_file <- system.file('extdata/proteinGroups.txt', package = 'graumann.lfq')
   create_maxquant_design_df(proteingroups_file)
   sample_design_file <- paste0(tempdir(), '/graumann.lfq/sample_design.txt')
   dir.create(dirname(sample_design_file))
   create_maxquant_design_file(proteingroups_file, sample_design_file)
}

# raw labeled intensities
#---------------------
if (require(billing.differentiation.data)){
   proteingroups_file <- system.file('extdata/maxquant/proteinGroups.txt',
                                     package = 'billing.differentiation.data')
   create_maxquant_design_df(proteingroups_file, value_type = 'raw.intensity')
}


# raw unlabeled intensities
#--------------------------
if (require(alnoubi.2017)){
   proteingroups_file <- system.file('extdata/proteinGroups.txt', package = 'alnoubi.2017')
   create_maxquant_design_df(proteingroups_file)
}
}
