% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sample_file.R
\name{create_maxquant_sample_file}
\alias{create_maxquant_sample_file}
\alias{create_maxquant_design_file}
\alias{create_sample_design_file}
\title{Create maxquant design file}
\usage{
create_maxquant_sample_file(proteingroups_file,
  sample_file = paste0(dirname(proteingroups_file), "/sample_design.txt"),
  value_type = autonomics.import::infer_maxquant_value_type(proteingroups_file),
  infer_design = FALSE)

create_maxquant_design_file(...)

create_sample_design_file(...)
}
\arguments{
\item{proteingroups_file}{string: full path to protein groups file}

\item{sample_file}{string: full path to sample design file}

\item{value_type}{string: any value in autonomics.import::MAXQUANT_VALUE_TYPES}

\item{infer_design}{logical: should design be infered from sample ids (see details)}

\item{...}{backward compatibility to deprecated functions}
}
\value{
sample design dataframe
}
\description{
For automated infer of design, use the following sample naming
scheme (before running max quant): WT(L).KD(M).OE(H).R1
}
\examples{
require(magrittr)

# LABELED RATIOS and INTENSITIES
if (require(autonomics.data)){
   file <- system.file(
             'extdata/stemcell.differentiation/maxquant/proteinGroups.txt',
              package = 'autonomics.data')
   file \%>\% create_maxquant_sample_df()
   file \%>\% create_maxquant_sample_file(sample_file = tempfile())
}

}
