% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sample_file.R
\name{create_metabolon_sample_df}
\alias{create_metabolon_sample_df}
\title{Create metabolon sample dataframe}
\usage{
create_metabolon_sample_df(metabolon_file, infer_design = FALSE)
}
\arguments{
\item{metabolon_file}{string: path to metabolon file}

\item{infer_design}{logical: whether to infer design from CLIENT_IDENTIFIER}

\item{sample_file}{string: path to sample file}
}
\value{
design dataframe
}
\description{
Create metabolon sample dataframe
}
\examples{
if (require(autonomics.data)){
   metabolon_file <- system.file('extdata/glutaminase/glutaminase.xlsx',
                                  package = 'autonomics.data')
   metabolon_file \%>\% autonomics.import::create_metabolon_sample_df()
   metabolon_file \%>\% autonomics.import::create_metabolon_sample_df(infer_design = TRUE)
}
}
