% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sample_file.R
\name{create_metabolon_sample_file}
\alias{create_metabolon_sample_file}
\title{Create metabolon sample file}
\usage{
create_metabolon_sample_file(metabolon_file,
  sample_file = paste0(dirname(proteingroups_file), "/sample_design.txt"),
  infer_design = FALSE)
}
\arguments{
\item{metabolon_file}{string: path to metabolon file}

\item{sample_file}{string: path to sample file}

\item{infer_design}{logical: whether to infer design from sampleids}
}
\value{
path to metabolon file
}
\description{
Create metabolon sample file
}
\examples{
if (require(autonomics.data)){
   metabolon_file <- system.file('extdata/glutaminase/glutaminase.xlsx',
                                  package = 'autonomics.data')
   metabolon_file \%>\% autonomics.import::create_metabolon_sample_file(
                         sample_file = tempfile())
}
}
