% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_prepro_list.R
\name{create_prepro_list}
\alias{create_prepro_list}
\title{Create list with preprocessing details}
\usage{
create_prepro_list(assay, entity, quantity, software, parameters = list())
}
\arguments{
\item{assay}{value in SUPPORTED_ASSAYS}

\item{entity}{value in SUPPORTED_ENTITIES}

\item{quantity}{value in SUPPORTED_QUANTITIES}

\item{software}{value in SUPPORTED_SOFTWARES}

\item{parameters}{preprocessing parameters}
}
\value{
list(assay, entity, quantity, software, parameters)
}
\description{
Create list with preprocessing details
}
