% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sample_file.R
\name{create_soma_sample_df}
\alias{create_soma_sample_df}
\title{Create soma design dataframe}
\usage{
create_soma_sample_df(soma_file, infer_design = FALSE)
}
\arguments{
\item{soma_file}{string: path to soma file}

\item{infer_design}{logical: whether to infer design from sampleids}
}
\value{
sample design dataframe
}
\description{
Create soma design dataframe
}
\examples{
if (require(autonomics.data)){
   soma_file <- system.file('extdata/stemcell.comparison/stemcell.comparison.adat',
                             package = 'autonomics.data')
   soma_file \%>\% autonomics.import::create_soma_sample_df()
   soma_file \%>\% autonomics.import::create_soma_sample_df(infer_design = TRUE)
}
}
