% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sample_file.R
\name{create_soma_sample_file}
\alias{create_soma_sample_file}
\title{Create soma sample file}
\usage{
create_soma_sample_file(soma_file, sample_file, infer_design = FALSE)
}
\arguments{
\item{soma_file}{string: path to soma file}

\item{sample_file}{string: path to sample file}

\item{infer_design}{logical: whether to infer design from sampleids}
}
\value{
path to soma design file
}
\description{
Create soma sample file
}
\examples{
   soma_file <- system.file('extdata/stemcell.comparison/stemcell.comparison.adat',
                             package = 'autonomics.data')
   soma_file \%>\% autonomics.import::create_soma_sample_file(sample_file = tempfile())
}
