% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_omics_asis.R
\name{create_voom_design}
\alias{create_voom_design}
\title{Create design for voom transformation}
\usage{
create_voom_design(object, verbose = TRUE)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{verbose}{logical(1)}
}
\value{
NULL (if no replicates) or design matrix
}
\description{
Create design for voom transformation
}
