% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scomponents.R
\name{dcast_subgroups}
\alias{dcast_subgroups}
\title{Split, arrange, count subgroups}
\usage{
dcast_subgroups(subgroup_values, fill, fun.aggregate)
}
\arguments{
\item{subgroup_values}{vector with subgroup values}

\item{sep}{subgroup separator}
}
\value{
integer
}
\description{
Split, arrange, count subgroups
}
\examples{
require(magrittr)
if (require(atkin.2014)){
   object <- 'extdata/soma/WCQ-18-007.hybNorm.plateScale.medNorm.calibrate.20181004.adat' \%>\%
              system.file(package='atkin.2014') \%>\%
              autonomics.import::load_soma()
   subgroup_values <- object \%>\% autonomics.import::subgroup_values()
   subgroup_values \%>\% autonomics.import::sep
   subgroup_values \%>\% count_components()
   subgroup_values \%>\% split_components()
   subgroup_values \%>\% layout_subgroups()
   subgroup_values \%>\% count_subgroups()
}
}
