% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_omics.R
\name{deconvolute_proteingroups}
\alias{deconvolute_proteingroups}
\title{Deconvolute proteingroups}
\usage{
deconvolute_proteingroups(object, fastafile, fastafields = c("GENES",
  "PROTEIN-NAMES", "EXISTENCE", "REVIEWED"))
}
\arguments{
\item{object}{SummerizedExperiment with proteinGroups data}

\item{fastafile}{path to fastafile}

\item{fastafields}{character vector: fields to load from fastafile}
}
\value{
deconvoluted and annotated SummarizedExperiment
}
\description{
Deconvolute proteingroups
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   object <- autonomics.data::stemcomp.proteinratios
   fastafile <- '../data/uniprot_hsa_20140515.fasta'
   if (file.exists(fastafile)){
      object \%>\% deconvolute_proteingroups(fastafile = fastafile)
   }
}
}
