% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_omics.R
\name{demultiplex_snames}
\alias{demultiplex_snames}
\alias{demultiplex_snames.character}
\alias{demultiplex_snames.SummarizedExperiment}
\title{Demultiplex snames}
\usage{
demultiplex_snames(x, ...)

\method{demultiplex_snames}{character}(x, verbose = FALSE, ...)

\method{demultiplex_snames}{SummarizedExperiment}(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{character vector or SummarizedExperiment}

\item{...}{allow for S3 dispatch}

\item{verbose}{logical}
}
\description{
For charactervector or SummarizedExperiment
}
\examples{
require(magrittr)

# character vector
    # Alternate multiplexing forms supported
     demultiplex_snames("STD(L)_EM00(M)_EM01(H)_R1{M/L}")     # Label Ratio
     demultiplex_snames('A(0)_B(1)_C(2)_D(3)_R1{0}'     )     # Reporter intensity
     demultiplex_snames('STD(L)_EM00(M)_EM01(H)_R1{L}')       # Label Intensity

   # Alternate separators supported
     demultiplex_snames('STD(L)_EM00(M)_EM01(H)_R1{L}')       # underscore
     demultiplex_snames('STD(L).EM00(M).EM01(H).R1{L}')       # dot
     demultiplex_snames('STD(L)EM00(M)EM01(H).R1{L}')         # no separator

   # Composite snames supported
     demultiplex_snames("WT.t0(L)_WT.t1(M)_WT.t2(H)_R1{H/L}") # composite snames

   # Uniqueness ensured by appending labels when necessary
     demultiplex_snames(c("STD(L).BM00(M).BM00(H).R10{M/L}",  # implicit uniquification
                                         "STD(L).BM00(M).BM00(H).R10{H/L}"))
   # Uniplexed snames are returned unchanged
     demultiplex_snames(c('STD_R1', 'EM0_R1'))

# SummarizedExperiment
  if (require(autonomics.data)){
     x <- 'extdata/stemdiff/maxquant/proteinGroups.txt' \%>\%
           system.file(package = 'autonomics.data')     \%>\%
           read_proteingroups(standardize_snames = FALSE, demultiplex_snames = FALSE) \%>\%
           standardize_maxquant_snames()
     x \%>\% demultiplex_snames(verbose = TRUE)
  }

}
