% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load2_sdata.R
\name{designify_maxquant_sampleids}
\alias{designify_maxquant_sampleids}
\title{Designify maxquant sampleids}
\usage{
designify_maxquant_sampleids(sampleids)
}
\arguments{
\item{sampleids}{character vector}
}
\description{
Designify maxquant sampleids
}
\examples{
require(magrittr)
sampleids <- c("STD(L).EM00(M).EM01(H).R1[H/L]",
               "STD(L).EM00(M).EM01(H).R1[M/L]")
sampleids \%>\% autonomics.import::designify_maxquant_sampleids()

sampleids <- c("Gel 11 1", "Gel 11 2", "Gel Ctrl 1")
sampleids \%>\% autonomics.import::designify_maxquant_sampleids()

sampleids <- c("ESC(0).NCM(1).CM(2).MV(3).EX(4).KIT(5).R1[0]",
               "ESC(0).NCM(1).CM(2).MV(3).EX(4).KIT(5).R1[1]")
sampleids \%>\% autonomics.import::designify_maxquant_sampleids()
}
