% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eset_specific_var.R
\name{ensg_var}
\alias{ensg_var}
\title{Get ensg var}
\usage{
ensg_var(object)
}
\arguments{
\item{object}{SummarizedExperiment}
}
\description{
Get ensg var
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   object <- 'extdata/stemdiff/rnaseq/gene_counts.txt' \%>\%
              system.file(package = 'autonomics.data') \%>\%
              read_rnaseq(fid_var = 'gene_id')
   fvars(object)
}
}
