% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eset_to_table.R
\name{eset_to_long_table}
\alias{eset_to_long_table}
\title{Convert eset into long datatable}
\usage{
eset_to_long_table(object, fid = autonomics.import::fid_var(object),
  fvars = character(0), sid = "sample_id", svars = character(0))
}
\arguments{
\item{object}{eset}

\item{fid}{fvar carrying feature id}

\item{fvars}{additional fvars to include in table}

\item{sid}{svar carrying sample id}

\item{svars}{additional svars to include in table}
}
\value{
datatable
}
\description{
Convert eset into long datatable
}
