% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eset_to_table.R
\name{eset_to_wide_table}
\alias{eset_to_wide_table}
\title{Convert eset into wide datatable}
\usage{
eset_to_wide_table(object, fid = autonomics.import::fid_var(object),
  fvars = character(0))
}
\arguments{
\item{object}{eset}

\item{fid}{fvar carrying feature id}

\item{fvars}{additional fvars to include in table}
}
\value{
datatable
}
\description{
Convert eset into wide datatable
}
