% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_maxquant.R
\name{esetise_maxquant_dt}
\alias{esetise_maxquant_dt}
\title{Convert long data table into eset}
\usage{
esetise_maxquant_dt(DT, entity, quantity, sample_file, parameter_file)
}
\arguments{
\item{DT}{long datatable}

\item{entity}{entity}

\item{quantity}{quantity}

\item{sample_file}{sample design file}

\item{parameter_file}{parameter file}
}
\description{
Convert long data table into eset
}
\examples{
require(magrittr)
if (require(billing.differentiation.data)){
   maxquant_dir       <- system.file('extdata/maxquant', package = 'billing.differentiation.data')
   sample_file        <- paste0(maxquant_dir, '/sample_design.txt')
   proteingroups_file <- paste0(maxquant_dir, '/proteinGroups.txt')
   phosphosites_file  <- paste0(maxquant_dir, '/phospho (STY)Sites.txt')
   parameter_file     <- paste0(maxquant_dir, '/parameters.txt')

   autonomics.import::load_proteingroups_to_long_dt(proteingroups_file) \%>\%
   autonomics.import::esetise_maxquant_dt(entity = 'proteingroup', quantity = 'normalized.ratio',
                          sample_file, parameter_file)
   autonomics.import::load_phosphosite_ratios_to_long_dt(phosphosites_file) \%>\%
   autonomics.import::esetise_maxquant_dt(entity = 'phosphosite', quantity = 'normalized.ratio',
                          sample_file, parameter_file)
   autonomics.import::load_phosphosite_occupancies_to_long_dt(phosphosites_file) \%>\%
   autonomics.import::esetise_maxquant_dt(entity = 'phosphosite', quantity = 'occupancy',
                          sample_file, parameter_file)
}
if (require(graumann.lfq)){
   maxquant_dir <- system.file('extdata', package = 'graumann.lfq')
   proteingroups_file <- paste0(maxquant_dir, '/proteinGroups.txt')
   sample_file <- paste0(maxquant_dir,    '/sample_design.txt')
   parameter_file <- paste0(maxquant_dir, '/parameters.txt')
   DT <- autonomics.import::load_proteingroups_to_long_dt(proteingroups_file)
   DT \%>\% autonomics.import::esetise_maxquant_dt(
             entity = 'proteingroup', quantity = 'lfq.intensity', sample_file, parameter_file)
}
}
