% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_maxquant_evidences.R
\name{evidences_to_proteingroups}
\alias{evidences_to_proteingroups}
\title{Aggregate evidences into protein groups}
\usage{
evidences_to_proteingroups(evidence_file)
}
\arguments{
\item{evidence_file}{path to evidence file}
}
\description{
Reads evidences, filters, deconvolutes razor sequences, melts labels, and summarizes per protein group.
}
\details{
Reads the evidence.txt file. \cr \cr
Then filters features by removing:
\itemize{
   \item sequences with variable modifications
   \item contaminant and reverse sequences
   \item features with no quantifications
}
Deconvolutes razor sequences by assigning them to the protein group with most other sequences. \cr
Melts the (labels of the) evidence table. \cr
Summarizes eviences per protein group:
\itemize{
   \item 1-3 obs: geomedian of all obs
   \item 3+ direct peptides: geomedian of direct obs
   \item cor(LR,LI)<0.005: max(lm(LR,LI))
   \item remaining obs: geomedian of all obs
}
}
\examples{
require(magrittr)
\dontrun{
if (require(billing.differentiation.data)){
   evidence_file <- system.file('extdata/maxquant/evidence.txt',
                                 package = 'billing.differentiation.data')
   evidences_to_proteingroups(evidence_file)
}
}
}
