% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_exiqon.R
\docType{data}
\name{EXIQON_FEATURE_ROWS}
\alias{EXIQON_FEATURE_ROWS}
\title{EXIQON FEATURE ROWS}
\format{An object of class \code{character} of length 3.}
\usage{
EXIQON_FEATURE_ROWS
}
\description{
EXIQON FEATURE ROWS
}
\keyword{datasets}
