% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_exiqon.R
\docType{data}
\name{EXIQON_PATTERN}
\alias{EXIQON_PATTERN}
\title{EXIQON PATTERN}
\format{An object of class \code{character} of length 1.}
\usage{
EXIQON_PATTERN
}
\description{
EXIQON PATTERN
}
\keyword{datasets}
