% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_exiqon.R
\docType{data}
\name{EXIQON_SAMPLE_COLS}
\alias{EXIQON_SAMPLE_COLS}
\title{EXIQON SAMPLE COLS}
\format{An object of class \code{character} of length 3.}
\usage{
EXIQON_SAMPLE_COLS
}
\description{
EXIQON SAMPLE COLS
}
\keyword{datasets}
