% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_features}
\alias{extract_features}
\title{Extract features}
\usage{
extract_features(object, extractor)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{extractor}{logical/numeric vector}
}
\description{
Extract features
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   object <- autonomics.data::stemcomp.proteinratios
   (object \%<>\% extract_features(c(5,4)))
   object \%>\% limma()
}
}
