% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_limma.R
\name{extract_limma_dt}
\alias{extract_limma_dt}
\title{Extract limma datatable}
\usage{
extract_limma_dt(object)
}
\arguments{
\item{object}{SummarizedExperiment}
}
\value{
melted data.table
}
\description{
Extract limma datatable
}
\examples{
\dontrun{
require(magrittr)
if (require(subramanian.2016)){
   subramanian.2016::metabolon \%>\% autonomics.import::extract_limma_dt()
}
}
}
