% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_maxquant2.R
\name{extract_maxquant_channels}
\alias{extract_maxquant_channels}
\title{Extract channel values from maxquant file}
\usage{
extract_maxquant_channels(file)
}
\arguments{
\item{file}{string: path to proteinGroups.txt file (or other maxquant file)}
}
\value{
character vector: maxquant channel values
}
\description{
Extract channel values from maxquant file
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   file <- system.file('extdata/stemcell.comparison/maxquant/proteinGroups.txt',
                        package = 'autonomics.data')
   file \%>\% autonomics.import::extract_maxquant_channels()
}
if (require(graumann.lfq)){
   file <- system.file('extdata/proteinGroups.txt',package = 'graumann.lfq')
   file \%>\% autonomics.import::extract_maxquant_channels()
}
if (require(billing.vesicles)){
   file <- system.file('extdata/proteinGroups.txt', package = 'billing.vesicles')
   file \%>\% autonomics.import::extract_maxquant_channels()
}
}
