% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_maxquant2.R
\name{extract_maxquant_fdata}
\alias{extract_maxquant_fdata}
\title{Extract maxquant fdata}
\usage{
extract_maxquant_fdata(file)
}
\arguments{
\item{file}{path to proteinGroups.txt file}
}
\value{
dataframe
}
\description{
Extract maxquant fdata
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   file <- 'extdata/stemcell.comparison/maxquant/proteinGroups.txt' \%>\%
            system.file(package = 'autonomics.data')
   file \%>\% autonomics.import::extract_maxquant_fdata() \%>\% head()
}
}
