% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load___sdata.R
\name{extract_maxquant_injections}
\alias{extract_maxquant_injections}
\title{Extract injection values from maxquant file}
\usage{
extract_maxquant_injections(file)
}
\arguments{
\item{file}{'proteinGroups.txt'}
}
\value{
character vector: injection values
}
\description{
Extract injection values from maxquant file
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   file <- system.file('extdata/stemcomp/maxquant/proteinGroups.txt',
                        package = 'autonomics.data')
   file \%>\% autonomics.import::extract_maxquant_injections()
}
if (require(graumann.lfq)){
   file <- system.file('extdata/proteinGroups.txt',package = 'graumann.lfq')
   file \%>\% autonomics.import::extract_maxquant_injections()
}
if (require(billing.vesicles)){
   file <- system.file('extdata/proteinGroups.txt', package = 'billing.vesicles')
   file \%>\% autonomics.import::extract_maxquant_injections()
}
}
