% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_maxquant2.R
\name{extract_maxquant_intensity_colnames}
\alias{extract_maxquant_intensity_colnames}
\title{Extract maxquant intensity colnames}
\usage{
extract_maxquant_intensity_colnames(file)
}
\arguments{
\item{file}{full path to proteinGroups.txt}
}
\value{
character vector with intensity column names
}
\description{
Extract maxquant intensity colnames
}
\examples{
 require(magrittr)
 if (require(autonomics.data)){
   file <- 'extdata/stemcell.comparison/maxquant/proteinGroups.txt' \%>\%
            system.file(package = 'autonomics.data')
   file \%>\% autonomics.import::extract_maxquant_intensity_colnames() \%>\% head()
}
}
