% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_maxquant2.R
\name{extract_maxquant_ratio_colnames}
\alias{extract_maxquant_ratio_colnames}
\title{Extract maxquant ratio colnames}
\usage{
extract_maxquant_ratio_colnames(file, normalized = FALSE)
}
\arguments{
\item{file}{full path to proteinGroups.txt}

\item{normalized}{logical}
}
\value{
character vector with intensity column names
}
\description{
Extract maxquant ratio colnames
}
\examples{
 require(magrittr)
 if (require(autonomics.data)){
   file <- 'extdata/stemcell.comparison/maxquant/proteinGroups.txt' \%>\%
            system.file(package = 'autonomics.data')
   file \%>\% autonomics.import::extract_maxquant_ratio_colnames() \%>\% head()
}
}
