% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_maxquant2.R
\name{extract_maxquant_snames}
\alias{extract_maxquant_snames}
\title{Extract maxquant snames}
\usage{
extract_maxquant_snames(file)
}
\arguments{
\item{file}{full path to proteinGroups.txt}
}
\value{
character vector with sample names
}
\description{
Extract maxquant snames
}
\examples{
 require(magrittr)
 if (require(autonomics.data)){
   file <- 'extdata/stemcell.comparison/maxquant/proteinGroups.txt' \%>\%
            system.file(package = 'autonomics.data')
   file \%>\% autonomics.import::extract_maxquant_snames() \%>\% head()
}
}
