% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_omics.R
\name{extract_rectangle}
\alias{extract_rectangle}
\alias{extract_rectangle.character}
\alias{extract_rectangle.data.table}
\alias{extract_rectangle.matrix}
\title{Extract rectangle from omics file, data.table, or matrix}
\usage{
extract_rectangle(x, ...)

\method{extract_rectangle}{character}(
  x,
  rows = 1:nrows(x, sheet = sheet),
  cols = 1:ncols(x, sheet = sheet),
  verbose = FALSE,
  transpose = FALSE,
  drop = FALSE,
  sheet = 1,
  ...
)

\method{extract_rectangle}{data.table}(
  x,
  rows = 1:nrow(x),
  cols = 1:ncol(x),
  transpose = FALSE,
  drop = FALSE,
  ...
)

\method{extract_rectangle}{matrix}(
  x,
  rows = 1:nrow(x),
  cols = 1:ncol(x),
  transpose = FALSE,
  drop = FALSE,
  ...
)
}
\arguments{
\item{x}{omics datafile or datatable}

\item{...}{allow for S3 method dispatch}

\item{rows}{numeric vector}

\item{cols}{numeric vector}

\item{verbose}{logical}

\item{transpose}{logical}

\item{drop}{logical}

\item{sheet}{numeric or string}
}
\value{
matrix
}
\description{
Extract rectangle from omics file, data.table, or matrix
}
\examples{
require(magrittr)
if (require('autonomics.data')){

   # filepath
      x <- system.file('extdata/glutaminase/glutaminase.xlsx', package = 'autonomics.data')
      x \%>\% extract_rectangle(rows = 11:401, cols = 15:86, sheet = 2) \%>\% extract(1:3, 1:3) # exprs
      x \%>\% extract_rectangle(rows = 11:401, cols = 5,     sheet = 2) \%>\% extract(1:3, )    # fids
      x \%>\% extract_rectangle(rows = 2,      cols = 15:86, sheet = 2) \%>\% extract(,1:3)     # sids
      x \%>\% extract_rectangle(rows = 10:401, cols = 1:14,  sheet = 2) \%>\% extract(1:3, 1:3) # fdata
      x \%>\% extract_rectangle(rows = 1:10,   cols = 14:86, sheet = 2,
                              transpose = TRUE)                       \%>\% extract(1:3, 1:3) # sdata
   # matrix
      x <- system.file('extdata/glutaminase/glutaminase.xlsx', package = 'autonomics.data') \%>\%
           extract_rectangle(sheet = 2)
      x \%>\% extract_rectangle(rows = 11:401, cols = 15:86, sheet = 2) \%>\% extract(1:3, 1:3) # exprs
      x \%>\% extract_rectangle(rows = 11:401, cols = 5,     sheet = 2) \%>\% extract(1:3, )    # fids
      x \%>\% extract_rectangle(rows = 2,      cols = 15:86, sheet = 2) \%>\% extract(,1:3)     # sids
      x \%>\% extract_rectangle(rows = 10:401, cols = 1:14,  sheet = 2) \%>\% extract(1:3, 1:3) # fdata
      x \%>\% extract_rectangle(rows = 1:10,   cols = 14:86, sheet = 2,
                              transpose = TRUE)                       \%>\% extract(1:3, 1:3) # sdata
}
}
