% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters.R
\docType{methods}
\name{fdata}
\alias{fdata}
\alias{fdata,SummarizedExperiment-method}
\alias{fdata,eSet-method}
\alias{fdata,EList-method}
\alias{fdata<-}
\alias{fdata<-,SummarizedExperiment,data.frame-method}
\alias{fdata<-,eSet,data.frame-method}
\alias{fdata<-,EList,data.frame-method}
\title{Get/Set fdata}
\usage{
fdata(object)

\S4method{fdata}{SummarizedExperiment}(object)

\S4method{fdata}{eSet}(object)

\S4method{fdata}{EList}(object)

fdata(object) <- value

\S4method{fdata}{SummarizedExperiment,data.frame}(object) <- value

\S4method{fdata}{eSet,data.frame}(object) <- value

\S4method{fdata}{EList,data.frame}(object) <- value
}
\arguments{
\item{object}{SummarizedExperiment, eSet, or EList}

\item{value}{data.frame}
}
\value{
feature dataframe (get) or updated object (set)
}
\description{
Get/Set feature data
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   object <- autonomics.data::stemcomp.proteinratios
   object \%>\% fdata() \%>\% head()
   fdata(object) \%<>\% cbind(z=1)
   object
}
}
