% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_exiqon.R
\name{filter_exiqon_features}
\alias{filter_exiqon_features}
\title{Filter mir features}
\usage{
filter_exiqon_features(object, rm_ref_genes = TRUE, rm_spike = TRUE,
  verbose = TRUE)
}
\arguments{
\item{object}{eset}

\item{rm_ref_genes}{whether to remove ref genes (logical)}

\item{rm_spike}{whether to remove spike-ins (logical)}

\item{verbose}{whether to report progress}
}
\value{
filtered eset
}
\description{
Filter mir features
}
