% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_exprs_replicated_in_some_subgroup}
\alias{filter_exprs_replicated_in_some_subgroup}
\title{Filter features with replicated expression in some subgroup}
\usage{
filter_exprs_replicated_in_some_subgroup(object, comparator = if
  (contains_ratios(object)) "!=" else ">", lod = 0)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{comparator}{'>' or '!='}

\item{lod}{numeric(1)}
}
\value{
Filtered SummarizedExperiment
}
\description{
Filter features with replicated expression in some subgroup
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   object <- autonomics.data::stemcomp.proteinratios
   object \%>\% filter_exprs_replicated_in_some_subgroup()
   object <- autonomics.data::glutaminase
   object \%>\% filter_exprs_replicated_in_some_subgroup()
}
}
