% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_features_on_direction}
\alias{filter_features_on_direction}
\title{Select portion of eSet with only up (or down) features}
\usage{
filter_features_on_direction(object, contrast, direction)
}
\arguments{
\item{object}{eSet}

\item{contrast}{contrast (named string)}

\item{direction}{'neg' or 'pos'}
}
\description{
Select portion of eSet with only up (or down) features
}
