% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_features_on_fvalues}
\alias{filter_features_on_fvalues}
\title{Filter subet for reference features}
\usage{
filter_features_on_fvalues(object, fvar, split, fvalues)
}
\arguments{
\item{object}{eSet}

\item{fvar}{fvar}

\item{split}{split}

\item{fvalues}{fvalues}
}
\value{
subset of object with reference features
}
\description{
Filter subet for reference features
}
\examples{
if (require(autonomics.data)){
   require(magrittr)
   fvalues <- c("A0MZ66", "A0JLT2", "Q8NG66", "Q9NPI5")
   object <- autonomics.data::stemcomp.proteinratios
   object \%>\% filter_features_on_fvalues(fvar    = 'Uniprot accessions',
                                         split   = ';',
                                         fvalues = fvalues)
}
}
