% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_filter.R
\name{filter_samples_}
\alias{filter_samples_}
\alias{filter_samples}
\title{Filter samples on condition}
\usage{
filter_samples_(x, condition)

filter_samples(x, condition)
}
\arguments{
\item{x}{eSet}

\item{condition}{filter condition}
}
\value{
filtered eSet
}
\description{
Filter samples on condition
}
\examples{
if (require(autonomics.data)){
   filter_samples( ALL,  sex == 'M')
   filter_samples_(ALL, "sex == 'M'")
}
}
