% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_consistent_nas.R
\name{is_available_for_some_feature}
\alias{is_available_for_some_feature}
\alias{filter_samples_available_for_some_feature}
\title{Filter samples available for some feature}
\usage{
is_available_for_some_feature(object)

filter_samples_available_for_some_feature(object, verbose = FALSE)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{verbose}{TRUE or FALSE}
}
\value{
SummarizedExperiment
}
\description{
Filter samples available for some feature
}
