% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{flatten}
\alias{flatten}
\title{Write sumexp to file}
\usage{
flatten(object, limma_quantities = c("effect", "p", "fdr"))
}
\arguments{
\item{object}{SummarizedExperiment}

\item{limma_quantities}{character vector: which limma quantities to include}
}
\value{
data.table       data.table
}
\description{
Write sumexp to file
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   object <- autonomics.data::stemcomp.proteinratios
   object \%>\% autonomics.import::flatten() \%>\% head()
}
}
