% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters.R
\name{flevels}
\alias{flevels}
\title{Get fvar levels}
\usage{
flevels(object, fvar)
}
\arguments{
\item{object}{SummarizedExperiment, eSet, or EList}

\item{fvar}{feature variable}
}
\value{
fvar values
}
\description{
Get fvar levels
}
\examples{
require(magrittr)
if (require(subramanian.2016)){
   subramanian.2016::metabolon \%>\% flevels('BIOCHEMICAL')
}
}
