% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eset_specific_var.R
\name{fname_var}
\alias{fname_var}
\alias{fname_values}
\title{Get fname var/values}
\usage{
fname_var(object)

fname_values(object)
}
\arguments{
\item{object}{SummarizedExperiment}
}
\value{
fvar with feature_name values
}
\description{
Get fname var/values
}
\examples{
library(magrittr)
if (require(autonomics.data)){
   object <- 'extdata/stemcomp/soma/stemcomp.adat' \%>\%
              system.file(package = 'autonomics.data') \%>\%
              autonomics.import::read_somascan()
   object \%>\% autonomics.import::fname_var()
   object \%>\% autonomics.import::fname_values()
}
}
