% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_maxquant_evidences.R
\name{geomedian}
\alias{geomedian}
\title{Compute geometric median}
\usage{
geomedian(x)
}
\arguments{
\item{x}{numeric vector for which to compute geometric median}
}
\description{
Compute geometric median
}
