% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_load_maxquant.R
\name{get_maxquant_lfq_intensities}
\alias{get_maxquant_lfq_intensities}
\title{Extract lfq intensity columns}
\usage{
get_maxquant_lfq_intensities(DT)
}
\arguments{
\item{DT}{maxquant data.table}
}
\value{
LFQ intensity data.table
}
\description{
Extract lfq intensity columns
}
\examples{
require(magrittr)
if (require(alnoubi.2017)){
   system.file('extdata/proteinGroups.txt',
                package = 'alnoubi.2017') \%>\%
   data.table::fread() \%>\%
   autonomics.import::get_maxquant_lfq_intensities()
}
if (require(graumann.lfq)){
   system.file('extdata/proteinGroups.txt',
                package = 'graumann.lfq') \%>\%
   data.table::fread() \%>\%
   get_maxquant_lfq_intensities()
}
}
