% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_maxquant.R
\name{get_maxquant_npeptides}
\alias{get_maxquant_npeptides}
\title{Get npeptides matrix}
\usage{
get_maxquant_npeptides(DT, value_type)
}
\arguments{
\item{DT}{data.table}

\item{value_type}{maxquant value type}
}
\description{
Get feature x sample matrix with number of peptides
}
\examples{
require(magrittr)
file <- 'extdata/stemcomp/maxquant/proteinGroups.txt' \%>\%
         system.file(package = 'autonomics.data')
DT <- file \%>\% data.table::fread()
value_type <- autonomics.import::infer_maxquant_value_type(file)
autonomics.import::get_maxquant_npeptides(DT, value_type) \%>\% str()
}
