% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_maxquant.R
\name{get_maxquant_raw_intensities}
\alias{get_maxquant_raw_intensities}
\title{Extract raw intensities}
\usage{
get_maxquant_raw_intensities(DT)
}
\arguments{
\item{DT}{maxquant data.table}
}
\value{
raw intensity data.table
}
\description{
Extract raw intensities
}
\examples{
require(magrittr)
if (require(billing.differentiation.data)){
   system.file('extdata/maxquant/proteinGroups.txt',
                package = 'billing.differentiation.data') \%>\%
   data.table::fread() \%>\%
   autonomics.import::get_maxquant_raw_intensities()
}
if (require(graumann.lfq)){
   system.file('extdata/proteinGroups.txt',
                package = 'graumann.lfq') \%>\%
   data.table::fread() \%>\%
   autonomics.import::get_maxquant_raw_intensities()
}
if (require(alnoubi.2017)){
   system.file('extdata/proteinGroups.txt',
                package = 'alnoubi.2017') \%>\%
   data.table::fread() \%>\%
   autonomics.import::get_maxquant_raw_intensities()
}
}
