% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_load_maxquant.R
\name{get_maxquant_raw_ratios}
\alias{get_maxquant_raw_ratios}
\title{Extract raw ratio columns}
\usage{
get_maxquant_raw_ratios(DT)
}
\arguments{
\item{DT}{maxquant data.table}
}
\value{
raw ratio data.table
}
\description{
Extract raw ratio columns
}
\examples{
require(magrittr)
if (require(billing.differentiation.data)){
   system.file('extdata/maxquant/proteinGroups.txt',
                package = 'billing.differentiation.data') \%>\%
      data.table::fread() \%>\%
      get_maxquant_raw_ratios()
}
}
