% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_load_maxquant.R
\name{get_maxquant_reporter_intensities}
\alias{get_maxquant_reporter_intensities}
\title{Extract reporter intensities}
\usage{
get_maxquant_reporter_intensities(DT)
}
\arguments{
\item{DT}{maxquant data.table}
}
\value{
reporter intensity data.table
}
\description{
Extract reporter intensities
}
\examples{
require(magrittr)
if (require(billing.vesicles)){
   system.file('extdata/proteinGroups.txt',
                package = 'billing.vesicles') \%>\%
   data.table::fread() \%>\%
   get_maxquant_reporter_intensities()
}
}
