% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eset_specific_var.R
\name{get_sample_id_var}
\alias{get_sample_id_var}
\alias{get_sample_id_values}
\title{Get sample id var}
\usage{
get_sample_id_var(object)

get_sample_id_values(object)
}
\arguments{
\item{object}{SummarizedExperiment}
}
\value{
sample id var (character)
}
\description{
Get sample id var
}
