% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_maxquant.R
\name{get_summary_attr}
\alias{get_summary_attr}
\title{Get autonomics summary attributes}
\usage{
get_summary_attr(x)
}
\arguments{
\item{x}{object}
}
\value{
autonomics summary attribute of object
}
\description{
Get autonomics summary attributes
}
