% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xxx_components.R
\name{guess_component_sep}
\alias{guess_component_sep}
\alias{guess_component_sep.character}
\alias{guess_component_sep.SummarizedExperiment}
\title{Guess component separator}
\usage{
guess_component_sep(x, ...)

\method{guess_component_sep}{character}(x, possible_separators = c(".", " ",
  "_"), verbose = FALSE)

\method{guess_component_sep}{SummarizedExperiment}(x, svar = "subgroup")
}
\arguments{
\item{x}{character vector with sampleids}

\item{possible_separators}{character vector with possible separators to look for}

\item{verbose}{logical}
}
\value{
separator (string) or NULL (if no separator could be identified)
}
\description{
Guess component separator
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   autonomics.data::glutaminase \%>\% subgroup_values() \%>\% guess_component_sep()
   autonomics.data::glutaminase \%>\%                       guess_component_sep()
}

x <- c('PERM_NON.R1[H/L]', 'PERM_NON.R2[H/L]', 'PERM_NON.R3[H/L]', 'PERM_NON.R4[H/L]')
x \%>\% guess_component_sep()

x <- c('WT untreated 1', 'WT untreated 2', 'WT treated 1')
x \%>\% guess_component_sep()

x <- c('group1', 'group2', 'group3.R1')
x \%>\% guess_component_sep()
}
