% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_omics.R
\name{guess_maxquant_quantity}
\alias{guess_maxquant_quantity}
\alias{guess_maxquant_quantity.character}
\alias{guess_maxquant_quantity.data.frame}
\alias{guess_maxquant_quantity.SummarizedExperiment}
\title{Guess maxquant quantity from snames}
\usage{
guess_maxquant_quantity(x, ...)

\method{guess_maxquant_quantity}{character}(x, ...)

\method{guess_maxquant_quantity}{data.frame}(x, ...)

\method{guess_maxquant_quantity}{SummarizedExperiment}(x, ...)
}
\arguments{
\item{x}{charactervector, dataframe, or SummarizedExperiment}

\item{...}{used for proper S3 method dispatch}
}
\value{
character(1): value from names(maxquant_patterns)
}
\description{
charactervector, dataframe, or SummarizedExperiment.
}
\examples{
if (require(autonomics.data)){
   require(magrittr)

   # file
      x <- 'extdata/stemcomp/maxquant/proteinGroups.txt' \%>\%
            system.file(package = 'autonomics.data')
      guess_maxquant_quantity(x)

   # charactervector
       guess_maxquant_quantity("Ratio M/L normalized STD(L)_EM00(M)_EM01(H)_R1")
       guess_maxquant_quantity("Ratio M/L STD(L)_EM00(M)_EM01(H)_R1")
       guess_maxquant_quantity("LFQ intensity EM00.R1")
       guess_maxquant_quantity("Reporter intensity corrected 0 STD(0)EM00(1)EM01(2)_R1")
       guess_maxquant_quantity("Reporter intensity 0 STD(0)EM00(1)EM01(2)_R1")
       guess_maxquant_quantity("Intensity H STD(L)_EM00(M)_EM01(H)_R1")

   # dataframe
       x <- 'extdata/stemcomp/maxquant/proteinGroups.txt' \%>\%
             system.file(package = 'autonomics.data') \%>\%
             data.table::fread()
       guess_maxquant_quantity(x)

   # SummarizedExperiment
      x <- 'extdata/stemcomp/maxquant/proteinGroups.txt' \%>\%
            system.file(package = 'autonomics.data') \%>\%
            read_proteingroups(standardize_snames = FALSE,
                               demultiplex_snames = FALSE)
      guess_maxquant_quantity(x)
}
}
