% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess.R
\name{guess_sep}
\alias{guess_sep}
\alias{guess_sep.character}
\alias{guess_sep.factor}
\alias{guess_sep.SummarizedExperiment}
\alias{ssep}
\alias{subgroup_sep}
\title{Guess separator}
\usage{
guess_sep(x, ...)

\method{guess_sep}{character}(x, possible_separators = c(".", " ", "_"),
  verbose = FALSE, ...)

\method{guess_sep}{factor}(x, ...)

\method{guess_sep}{SummarizedExperiment}(x, var = "sample_id",
  possible_separators = c(".", "_", " "), verbose = FALSE, ...)

ssep(...)

subgroup_sep(...)
}
\arguments{
\item{x}{character vector or SummarizedExperiment}

\item{...}{used for proper S3 method dispatch}

\item{possible_separators}{character vector with possible separators to look for}

\item{verbose}{logical}

\item{var}{svar or fvar}
}
\value{
separator (string) or NULL (if no separator could be identified)
}
\description{
Guess separator
}
\examples{
require(magrittr)

# charactervector
   x <- c('PERM_NON.R1[H/L]', 'PERM_NON.R2[H/L]', 'PERM_NON.R3[H/L]', 'PERM_NON.R4[H/L]')
   x \%>\% autonomics.import::guess_sep()

   x <- c('WT untreated 1', 'WT untreated 2', 'WT treated 1')
   x \%>\% autonomics.import::guess_sep()

   x <- c('group1', 'group2', 'group3.R1')
   x \%>\% autonomics.import::guess_sep()

# SummarizedExperiment
   if (require(autonomics.data))   autonomics.data::glutaminase \%>\%
                                   autonomics.import::guess_sep()

   if (require(autonomics.data))   autonomics.data::stemcomp.proteinratios \%>\%
                                   autonomics.import::guess_sep()

   if (require(graumann.lfq))      graumann.lfq::lfq.intensities \%>\%
                                   autonomics.import::guess_sep()
}
