% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_has_is.R
\name{has_complete_svalues}
\alias{has_complete_svalues}
\alias{has_complete_subgroup_values}
\alias{has_complete_block_values}
\title{Does object have complete svalues}
\usage{
has_complete_svalues(object, svar)

has_complete_subgroup_values(object)

has_complete_block_values(object)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{svar}{sample var}
}
\value{
logical
}
\description{
Does object have complete svalues
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   object <- autonomics.data::glutaminase
   object \%>\% autonomics.import::has_complete_subgroup_values()
   object \%>\% autonomics.import::has_complete_block_values()
}
}
