% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svar_has_two_components.R
\name{has_two_components}
\alias{has_two_components}
\title{Do values consist of two components?}
\usage{
has_two_components(x)
}
\arguments{
\item{x}{character vector}
}
\value{
logical(1)
}
\description{
Do values consist of two components?
}
