% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_soma.R
\name{identify_soma_structure}
\alias{identify_soma_structure}
\title{Identify soma structure}
\usage{
identify_soma_structure(file)
}
\arguments{
\item{file}{adat file}
}
\value{
list(row, col, fvars, svars)
}
\description{
Identify structure of soma file
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   file <- system.file('extdata/stemcomp/soma/stemcomp.adat',
                        package = 'autonomics.data')
   file \%>\% autonomics.import::identify_soma_structure()
}
if (require(atkin.2014)){
   file <- system.file('extdata/soma/WCQ-14-130_Set_A_RPT.HybMedNormCal_20140925.adat',
                        package = 'atkin.2014')
   file \%>\% autonomics.import::identify_soma_structure()
}
}
\author{
Aditya Bhagwat
}
