% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_design.R
\name{infer_add_design}
\alias{infer_add_design}
\title{Infer and add design to sdata}
\usage{
infer_add_design(sdata)
}
\arguments{
\item{sdata}{sample dataframe}
}
\value{
augmented sample dataframe
}
\description{
Infer and add design to sdata
}
