% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_design.R
\name{infer_design_from_default_sampleids}
\alias{infer_design_from_default_sampleids}
\title{Infer design from (non-maxquant) sample ids}
\usage{
infer_design_from_default_sampleids(sample_ids, sep = sample_ids \%>\%
  autonomics.import::guess_component_sep(c(".", " ", "_")))
}
\arguments{
\item{sample_ids}{character vector}

\item{sep}{string: design separator}
}
\description{
Infer design from (non-maxquant) sample ids
}
\examples{
require(magrittr)
sample_ids <- c("UT_10h_R1", "UT_10h_R2", "UT_10h_R3", "UT_10h_R4")
sample_ids \%>\% autonomics.import::infer_design_from_default_sampleids()
}
