% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_design.R
\name{infer_design_from_maxquant_sampleids}
\alias{infer_design_from_maxquant_sampleids}
\title{Infer design from maxquant sample ids}
\usage{
infer_design_from_maxquant_sampleids(sample_ids, sep = sample_ids \%>\%
  autonomics.import::guess_component_sep(c(".", " ", "_")))
}
\arguments{
\item{sample_ids}{character vector}

\item{sep}{string: design separator}
}
\description{
Infer design from maxquant sample ids
}
\examples{
require(magrittr)
sample_ids <- c("E(L).EM(M).BM(H).R1[M/L]", "E(L).EM(M).BM(H).R1[H/L]")
sample_ids \%>\% infer_design_from_maxquant_sampleids()
}
