% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_design.R
\name{infer_design_from_sampleids}
\alias{infer_design_from_sampleids}
\title{Convert sample IDs into sample design}
\usage{
infer_design_from_sampleids(sample_ids, sep = sample_ids \%>\%
  autonomics.import::infer_design_sep(c(".", " ", "_")),
  drop_common_parts = FALSE)
}
\arguments{
\item{sample_ids}{character vector with sampleids}

\item{sep}{string separator}

\item{drop_common_parts}{logical}
}
\description{
Convert sample IDs into sample design
}
\examples{
require(magrittr)
sample_ids <- c('PERM_NON.R1[H/L]', 'PERM_NON.R2[H/L]', 'PERM_NON.R3[H/L]', 'PERM_NON.R4[H/L]')
sample_ids \%>\% infer_design_from_sampleids()
if (require(subramanian.2016)){
   subramanian.2016::metabolon$sample_id \%>\%
   autonomics.import::infer_design_from_sampleids()
}
sample_ids <- c("UT_10h_R1", "UT_10h_R2", "UT_10h_R3", "UT_10h_R4")
sample_ids \%>\% autonomics.import::infer_design_from_sampleids()
}
