% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load___exprs.R
\name{infer_maxquant_quantity}
\alias{infer_maxquant_quantity}
\title{Infer maxquant quantity}
\usage{
infer_maxquant_quantity(file)
}
\arguments{
\item{file}{path to maxquant file}
}
\value{
string
}
\description{
Infer maxquant quantity
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   file <- 'extdata/stemcomp/maxquant/proteinGroups.txt' \%>\%
            system.file(package = 'autonomics.data')
   file \%>\% autonomics.import::infer_maxquant_quantity()
}
if (require(graumann.lfq)){
   file <- system.file('extdata/proteinGroups.txt', package = 'graumann.lfq')
   file \%>\% autonomics.import::infer_maxquant_quantity()
}
}
