% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_maxquant2.R
\name{infer_maxquant_value_pattern}
\alias{infer_maxquant_value_pattern}
\title{Load proteingroups snames}
\usage{
infer_maxquant_value_pattern(file)
}
\arguments{
\item{file}{full path to proteinGroups.txt}
}
\description{
Load proteingroups snames
}
\examples{
 require(magrittr)
 if (require(autonomics.data)){
    file <- 'extdata/stemcell.comparison/maxquant/proteinGroups.txt'  \%>\%
             system.file(package = 'autonomics.data')
    file \%>\% autonomics.import::infer_maxquant_type() \%>\% head()
}
if (require(graumann.lfq)){
    file <- 'extdata/proteinGroups.txt'  \%>\%
             system.file(package = 'graumann.lfq')
}
}
