% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_maxquant.R
\name{infer_maxquant_value_type}
\alias{infer_maxquant_value_type}
\title{Infer value type}
\usage{
infer_maxquant_value_type(file)
}
\arguments{
\item{file}{maxquant file}
}
\value{
value type
}
\description{
Infer value type
}
\examples{
require(magrittr)

# normalized.ratios
#------------------
if (require(billing.differentiation.data)){
   file <- system.file('extdata/maxquant/proteinGroups.txt',
                      package = 'billing.differentiation.data')
   file \%>\% infer_maxquant_value_type()
}

if (require(alnoubi.2017)){
   file <- system.file('extdata/proteinGroups.txt',
                      package = 'alnoubi.2017')
   file \%>\% infer_maxquant_value_type()
}

# reporter.intensities
#---------------------
if (require(billing.vesicles)){
   file <- system.file('extdata/proteinGroups.txt',
                        package = 'billing.vesicles')
   file \%>\% infer_maxquant_value_type()
}

# lfq intensities
#----------------
if (require(graumann.lfq)){
   file <- system.file('extdata/proteinGroups.txt',
                        package = 'graumann.lfq')
   file \%>\% infer_maxquant_value_type()
}

# raw intensities
#----------------
if (require(alnoubi.2017)){
   file <- system.file('extdata/proteinGroups.txt',
                        package = 'alnoubi.2017')
   file \%>\% infer_maxquant_value_type()
}
}
