% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert.R
\name{invert}
\alias{invert}
\alias{invert.character}
\alias{invert.SummarizedExperiment}
\title{Invert}
\usage{
invert(x, ...)

\method{invert}{character}(x, sep = guess_sep(x), ...)

\method{invert}{SummarizedExperiment}(x, subgroups = x \%>\%
  slevels("subgroup"), sep = x \%>\% guess_sep("subgroup"), ...)
}
\arguments{
\item{x}{character(.) or SummarizedExperiment}

\item{...}{to enable S3 method dispatch}

\item{sep}{character(1): collapsed string separator}

\item{subgroups}{character(.): subgroup levels for which inversion is required}
}
\value{
character(.) or SummarizedExperiment
}
\description{
For character input: invert collapsed strings.
For SummarizedExperiment input: invert expression values, invert subgroup definitions, update sample id values.
}
\examples{
require(magrittr)

# character
#----------
   x <- c('Ctrl_A', 'Ctrl_B')
   x \%>\% invert()

# SummarizedExperiment
#---------------------
if (require(autonomics.data)){
   x <- system.file('extdata/stemcomp/maxquant/proteinGroups.txt',
                     package = 'autonomics.data') \%>\%
        read_proteingroups()
   x \%>\% invert(subgroups = c('E_EM', 'E_BM', 'EM_BM'))
}
}
