% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_xxx_eset.R
\name{is_exiqon_eset}
\alias{is_exiqon_eset}
\title{Is object an exiqon eset?}
\usage{
is_exiqon_eset(object)
}
\arguments{
\item{object}{eset}
}
\value{
logical
}
\description{
Is object an exiqon eset?
}
\examples{
require(magrittr)
if (require(subramanian.2016)){
   object <- subramanian.2016::exiqon
   object \%>\% autonomics.import::is_exiqon_eset()
}
}
