% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_omics.R
\name{is_fixed_column_file}
\alias{is_fixed_column_file}
\title{Is this a fixed column file?}
\usage{
is_fixed_column_file(file)
}
\arguments{
\item{file}{string: file name}
}
\value{
TRUE or FALSE
}
\description{
Is this a fixed column file?
}
\examples{
# SOMASCAN
#---------
#file <- file.path(
#    '~/wcmq/atkin.hypo/atkin.hypoglycemia/extdata/soma/',
#    'WCQ-18-007.hybNorm.plateScale.medNorm.calibrate.20181004.original.adat')
#is_fixed_column_file(file)

# METABOLON
#----------
file <- system.file('extdata/glutaminase/glutaminase.xlsx',
                     package = 'autonomics.data')
is_fixed_column_file(file)
}
