% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_filter.R
\name{is_fvalue_feature}
\alias{is_fvalue_feature}
\title{Identify reference features}
\usage{
is_fvalue_feature(object, fvar, split, fvalues)
}
\arguments{
\item{object}{eset}

\item{fvar}{fvar}

\item{split}{split}

\item{fvalues}{fvalues}
}
\description{
Identify reference features
}
\examples{
if (require(billing.differentiation.data)){
   require(magrittr)
   fvalues <- c("A0MZ66", "A0JLT2", "Q8NG66", "Q9NPI5")
   object <- billing.differentiation.data::protein.ratios
   object \%>\% is_fvalue_feature(fvar = 'Uniprot accessions',
                                  split = ';', fvalues = fvalues) \%>\%
                sum()
}
}
