% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters.R
\name{is_imputed}
\alias{is_imputed}
\alias{is_imputed,SummarizedExperiment-method}
\alias{is_imputed<-}
\alias{is_imputed<-,SummarizedExperiment,matrix-method}
\alias{is_imputed<-,SummarizedExperiment,NULL-method}
\title{Get/set is_imputed}
\usage{
is_imputed(object)

\S4method{is_imputed}{SummarizedExperiment}(object)

is_imputed(object) <- value

\S4method{is_imputed}{SummarizedExperiment,matrix}(object) <- value

\S4method{is_imputed}{SummarizedExperiment,`NULL`}(object) <- value
}
\arguments{
\item{object}{SummarizedExperiment}

\item{value}{matrix}
}
\value{
matrix (get) or updated object (set)
}
\description{
Get/Set is_imputed
}
\examples{
if (require(autonomics.data)){
  require(magrittr)
  object <- autonomics.data::stemcomp.proteinratios
  object \%>\% is_imputed() \%>\% sum()
}
}
