% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_xxx_eset.R
\name{is_maxquant_eset}
\alias{is_maxquant_eset}
\title{Is max quant eSet?}
\usage{
is_maxquant_eset(object)
}
\arguments{
\item{object}{eSet}
}
\value{
logical
}
\description{
Is max quant eSet?
}
\examples{
library(magrittr)
if (require(autonomics.data)){
   object <- autonomics.data::billing2016
   object \%>\% is_maxquant_eset()
}
}
