% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_xxx_eset.R
\name{is_metabolon_eset}
\alias{is_metabolon_eset}
\title{Is object a metabolon eset?}
\usage{
is_metabolon_eset(object)
}
\arguments{
\item{object}{eset}
}
\value{
logical
}
\description{
Is object a metabolon eset?
}
\examples{
require(magrittr)
if (require(subramanian.2016)){
   object <- subramanian.2016::metabolon
   object \%>\% autonomics.import::is_metabolon_eset()
}
}
