% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_svalues_available.R
\name{is_missing_or_empty_character}
\alias{is_missing_or_empty_character}
\title{Is missing or empty character}
\usage{
is_missing_or_empty_character(x)
}
\arguments{
\item{x}{character vector}
}
\value{
logical vector
}
\description{
Is missing or empty character
}
