% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters.R
\name{is_na}
\alias{is_na}
\alias{is_na,SummarizedExperiment-method}
\title{Is NA?}
\usage{
is_na(object)

\S4method{is_na}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{SummarizedExperiment}
}
\value{
matrix
}
\description{
is.na equivalent for SummarizedExperiment
}
