% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_xxx_eset.R
\name{is_soma_eset}
\alias{is_soma_eset}
\title{Is object a soma eset?}
\usage{
is_soma_eset(object)
}
\arguments{
\item{object}{eset}
}
\value{
logical
}
\description{
Is object a soma eset?
}
\examples{
if (require(autonomics.data)){
require(magrittr)
   object <- autonomics.data::stemcomp.soma
   object \%>\% is_soma_eset()
}
}
