% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_has_is.R
\name{is_valid_eset}
\alias{is_valid_eset}
\title{Is valid eset?}
\usage{
is_valid_eset(x, .xname = assertive.base::get_name_in_parent(x))
}
\arguments{
\item{x}{eset}

\item{.xname}{see assertive.base::get_name_in_parent}
}
\value{
logical
}
\description{
Is valid eset?
}
