% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_has_is.R
\name{is_valid_object}
\alias{is_valid_object}
\title{Is valid SummarizedExperiment}
\usage{
is_valid_object(x, .xname = assertive.base::get_name_in_parent(x))
}
\arguments{
\item{x}{SummarizedExperiment}

\item{.xname}{see assertive.base::get_name_in_parent}
}
\value{
logical
}
\description{
Is valid SummarizedExperiment
}
