% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xxx_composite_values.R
\name{layout_composite_values}
\alias{layout_composite_values}
\alias{layout_composite_values.character}
\alias{layout_composite_values.factor}
\alias{layout_composite_values.SummarizedExperiment}
\title{Layout composite (s)values
Reshapes a vector with composite values into a 2d matrix:
   + rownames = values of component 1:(n-1)
   + colnames = values of component n
   + values = composite levels}
\usage{
layout_composite_values(x, ...)

\method{layout_composite_values}{character}(x,
  sep = autonomics.import::guess_sep(x))

\method{layout_composite_values}{factor}(x,
  sep = autonomics.import::guess_sep(x))

\method{layout_composite_values}{SummarizedExperiment}(x, svar = "subgroup",
  sep = autonomics.import::guess_sep(x))
}
\arguments{
\item{x}{character(n), factor(n), or SummarizedExperiment with composite (s)values}

\item{sep}{character(1)}

\item{svar}{character(1)}
}
\value{
integer
}
\description{
Layout composite (s)values
Reshapes a vector with composite values into a 2d matrix:
   + rownames = values of component 1:(n-1)
   + colnames = values of component n
   + values = composite levels
}
\examples{
require(magrittr)

# STEM CELL DIFFERENTIATION
if (require(autonomics.data)){
   x <- autonomics.data::stemdiff.proteinratios
   x \%>\% autonomics.import::subgroup_values() \%>\% autonomics.import::layout_composite_values()
   x                                          \%>\% autonomics.import::layout_composite_values()
}

# ATKIN
if (require(atkin.2014)){
   x <- atkin.2014::soma.2018
   x \%>\% autonomics.import::subgroup_values() \%>\% autonomics.import::layout_composite_values()
   x \%>\%                                          autonomics.import::layout_composite_values()
}
}
