% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scomponents.R
\name{layout_subgroups}
\alias{layout_subgroups}
\alias{count_subgroups}
\title{Layout/Count subgroups}
\usage{
layout_subgroups(subgroup_values)

count_subgroups(subgroup_values)
}
\arguments{
\item{subgroup_values}{vector with subgroup values}

\item{sep}{subgroup separator}
}
\value{
integer
}
\description{
Layout/Count subgroups
}
\examples{
require(magrittr)
if (require(atkin.2014)){
   object <- 'extdata/soma/WCQ-18-007.hybNorm.plateScale.medNorm.calibrate.20181004.adat' \%>\%
              system.file(package='atkin.2014') \%>\%
              autonomics.import::load_soma()
   subgroup_values <- object \%>\% autonomics.import::subgroup_values()
   subgroup_values \%>\% layout_subgroups()
   subgroup_values \%>\% count_subgroups()
}
}
