% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xxx_values.R
\name{layout_values}
\alias{layout_values}
\alias{layout_values.character}
\alias{layout_values.factor}
\alias{layout_values.SummarizedExperiment}
\title{Layout composite (s)values
Reshapes a vector with composite values into a 2d matrix:
   + rownames = values of component 1:(n-1)
   + colnames = values of component n
   + values = composite levels}
\usage{
layout_values(x, ...)

\method{layout_values}{character}(x, sep = guess_sep(x), ...)

\method{layout_values}{factor}(x, sep = guess_sep(x), ...)

\method{layout_values}{SummarizedExperiment}(x, svar = "subgroup",
  sep = guess_sep(x), ...)
}
\arguments{
\item{x}{character(n), factor(n), or SummarizedExperiment with composite (s)values}

\item{...}{used for proper S3 dispatching}

\item{sep}{character(1)}

\item{svar}{character(1)}
}
\value{
integer
}
\description{
Layout composite (s)values
Reshapes a vector with composite values into a 2d matrix:
   + rownames = values of component 1:(n-1)
   + colnames = values of component n
   + values = composite levels
}
\examples{
if (require(autonomics.data)){
   # GLUTAMINASE
   require(magrittr)
   x <- autonomics.data::glutaminase
   x \%>\% subgroup_values() \%>\% layout_values()
   x                       \%>\% layout_values()
}

}
