% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters.R
\name{limma}
\alias{limma}
\alias{limma,SummarizedExperiment-method}
\alias{limma<-}
\alias{limma<-,SummarizedExperiment,array-method}
\alias{limma<-,SummarizedExperiment,NULL-method}
\title{Get/set limma results}
\usage{
limma(object)

\S4method{limma}{SummarizedExperiment}(object)

limma(object) <- value

\S4method{limma}{SummarizedExperiment,array}(object) <- value

\S4method{limma}{SummarizedExperiment,`NULL`}(object) <- value
}
\arguments{
\item{object}{SummarizedExperiment}

\item{value}{list}
}
\value{
limma results (get) or updated object (set)
}
\description{
Get/Set limma results
}
