% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_exiqon.R
\name{load_exiqon_sdata}
\alias{load_exiqon_sdata}
\alias{load_exiqon_fdata}
\alias{load_exiqon}
\title{Load exiqon file}
\usage{
load_exiqon_sdata(file)

load_exiqon_fdata(file)

load_exiqon(file, sample_file = NULL, infer_design_from_sampleids = TRUE,
  rm_ref_genes = TRUE, rm_spike = TRUE)
}
\arguments{
\item{file}{string: path to exiqon txt file}

\item{sample_file}{NULL or string (path to sample file)}

\item{infer_design_from_sampleids}{logical: whether to infer design from samplesids}

\item{rm_ref_genes}{logical: whether to rm reference genes}

\item{rm_spike}{logical: whether to rm spike-ins}
}
\value{
eset
}
\description{
Load exiqon file
}
\examples{
require(magrittr)
if (require(subramanian.2016)){
   file <- system.file('extdata/exiqon/subramanian.2016.exiqon.xlsx',
                               package = 'subramanian.2016')
   file \%>\% autonomics.import::load_exiqon()
   file \%>\% autonomics.import::load_exiqon(infer_design = FALSE)
   file \%>\% autonomics.import::load_exiqon_sdata() \%>\% head(1)
   file \%>\% autonomics.import::load_exiqon_fdata() \%>\% head(1)
}
}
