% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load2_exprs.R
\name{load_exprs}
\alias{load_exprs}
\title{Load exprs}
\usage{
load_exprs(file, sheet = NULL, quantity = NULL, platform)
}
\arguments{
\item{file}{path to omics data file}

\item{sheet}{excel sheet number or name (if applicable)}

\item{quantity}{string: quantity to be loaded as exprs (if applicable)}

\item{platform}{'metabolon', 'metabolonlipids'}
}
\value{
sample dataframe
}
\description{
Load exprs
}
\examples{
 require(magrittr)
 if (require(autonomics.data)){
    file <- system.file('extdata/glutaminase/glutaminase.xlsx', package = 'autonomics.data')
    file \%>\% load_exprs(2, 'metabolon') \%>\% extract(1:3, 1:3)
 }
file <- '../../datasets/WCQA-01-18MLCLP-1/WCQA-01-18MLCLP CLP  6-TAB FILE (180710).XLSX'
if (file.exists(file)){
   file \%>\% load_exprs('Lipid Class Concentrations', 'metabolonlipids') \%>\% extract(1:3, 1:3)
}
}
