% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load2_exprs.R
\name{load_exprs_exiqon}
\alias{load_exprs_exiqon}
\title{Load exiqon exprs}
\usage{
load_exprs_exiqon(file)
}
\arguments{
\item{file}{path to exiqon data file}
}
\value{
numeric matrix
}
\description{
Load exiqon exprs
}
\examples{
require(magrittr)
if (require(subramanian.2016)){
   file <- system.file('extdata/exiqon/subramanian.2016.exiqon.xlsx',
                        package = 'subramanian.2016')
   file \%>\% autonomics.import::load_exprs_exiqon()
}
}
