% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load___exprs.R
\name{load_exprs_maxquant}
\alias{load_exprs_maxquant}
\title{Load maxquant exprs}
\usage{
load_exprs_maxquant(file, quantity)
}
\arguments{
\item{file}{full path to proteinGroups.txt}

\item{quantity}{'Ratio', 'Ratio normalized', Intensity', 'LFQ intensity', 'Reporter intensity'}
}
\value{
character vector with sample names
}
\description{
Load maxquant exprs
}
\examples{
 require(magrittr)
 if (require(autonomics.data)){
   file <- 'extdata/stemcomp/maxquant/proteinGroups.txt' \%>\%
            system.file(package = 'autonomics.data')
   file \%>\% load_exprs_maxquant('Ratio')            \%>\% extract(1:3, 1:3)
   file \%>\% load_exprs_maxquant('Ratio normalized') \%>\% extract(1:3, 1:3)
   file \%>\% load_exprs_maxquant('Intensity')        \%>\% extract(1:3, 1:3)
            extract(1:3, 1:3)
}
if (require(graumann.lfq)){
   file <- system.file('extdata/proteinGroups.txt', package = 'graumann.lfq')
   file \%>\% load_exprs_maxquant('LFQ intensity') \%>\% extract(1:3, 1:3)
   file \%>\% load_exprs_maxquant('Intensity')     \%>\% extract(1:3, 1:3)
}
if (require(billing.vesicles)){
   file <- system.file('extdata/proteinGroups.txt', package = 'billing.vesicles')
   file \%>\% load_exprs_maxquant('Reporter intensity') \%>\% extract(1:3, 1:3)
}
}
