% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load2_exprs.R
\name{load_exprs_metabolon}
\alias{load_exprs_metabolon}
\title{Load metabolon exprs}
\usage{
load_exprs_metabolon(file, sheet)
}
\arguments{
\item{file}{path to metabolon file}

\item{sheet}{excell sheet number or name}
}
\value{
sample dataframe
}
\description{
Load metabolon exprs
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   file <- system.file('extdata/glutaminase/glutaminase.xlsx', package = 'autonomics.data')
   file \%>\% load_exprs_metabolon(2) \%>\% extract(1:3, 1:3)
}
}
