% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load2_exprs.R
\name{load_exprs_rnaseq}
\alias{load_exprs_rnaseq}
\title{Load rnaseq exprs}
\usage{
load_exprs_rnaseq(file)
}
\arguments{
\item{file}{path to exiqon xls file}
}
\value{
sample dataframe
}
\description{
Load rnaseq exprs
}
\examples{
 require(magrittr)
 if (require(subramanian.2016)){
    file <- system.file('extdata/rnaseq/gene_counts.txt', package = 'subramanian.2016')
    file \%>\% autonomics.import::load_exprs_rnaseq()
 }
}
