% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load2_exprs.R
\name{load_exprs_soma}
\alias{load_exprs_soma}
\title{Load soma exprs}
\usage{
load_exprs_soma(file)
}
\arguments{
\item{file}{string: path to adat file}
}
\value{
sample dataframe
}
\description{
Load soma exprs
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   file <- system.file('extdata/stemcomp/soma/stemcomp.adat',
                        package = 'autonomics.data')
   file \%>\% autonomics.import::load_exprs_soma() \%>\% extract(1:3, 1:3)
}
if (require(atkin.2014)){
   file <- system.file('extdata/soma/WCQ-14-130_Set_A_RPT.HybMedNormCal_20140925.adat',
                        package = 'atkin.2014')
   file \%>\% autonomics.import::load_exprs_soma() \%>\% extract(1:3, 1:3)
}
}
\author{
Aditya Bhagwat
}
