% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load___fdata.R
\name{load_fdata_exiqon}
\alias{load_fdata_exiqon}
\title{Load exiqon fdata}
\usage{
load_fdata_exiqon(file)
}
\arguments{
\item{file}{path to exiqon data file}
}
\value{
feature dataframe
}
\description{
Load exiqon fdata
}
\examples{
require(magrittr)
if (require(subramanian.2016)){
   file <- system.file('extdata/exiqon/subramanian.2016.exiqon.xlsx',
                        package = 'subramanian.2016')
   file \%>\% autonomics.import::load_fdata_exiqon()
}
}
