% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load___fdata.R
\name{load_fdata_maxquant}
\alias{load_fdata_maxquant}
\title{Load maxquant fdata}
\usage{
load_fdata_maxquant(file)
}
\arguments{
\item{file}{path to proteinGroups.txt file}
}
\value{
dataframe
}
\description{
Load maxquant fdata
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   file <- 'extdata/stemcomp/maxquant/proteinGroups.txt' \%>\%
            system.file(package = 'autonomics.data')
   file \%>\% autonomics.import::load_fdata_maxquant() \%>\% head()
}
}
