% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load2_fdata.R
\name{load_fdata_soma}
\alias{load_fdata_soma}
\title{Load soma fdata}
\usage{
load_fdata_soma(file)
}
\arguments{
\item{file}{string: path to adat file}
}
\value{
feature dataframe
}
\description{
Load soma fdata
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   file <- system.file('extdata/stemcomp/soma/stemcomp.adat',
                        package = 'autonomics.data')
   file \%>\% autonomics.import::load_fdata_soma() \%>\% head()
}
if (require(atkin.2014)){
   file <- system.file('extdata/soma/WCQ-14-130_Set_A_RPT.HybMedNormCal_20140925.adat',
                        package = 'atkin.2014')
   file \%>\% autonomics.import::load_fdata_soma() \%>\% head()
}
}
\author{
Aditya Bhagwat
}
