% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load2_fdata.R
\name{load_fnames_maxquant}
\alias{load_fnames_maxquant}
\title{Load maxquant fnames}
\usage{
load_fnames_maxquant(file)
}
\arguments{
\item{file}{full path to proteinGroups.txt}
}
\value{
character vector with sample names
}
\description{
Load maxquant fnames
}
\examples{
 require(magrittr)
 if (require(autonomics.data)){
   file <- 'extdata/stemcell.comparison/maxquant/proteinGroups.txt' \%>\%
            system.file(package = 'autonomics.data')
   file \%>\% autonomics.import::load_fnames_maxquant() \%>\% head()
}
}
