% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_metabolon.R
\name{load_metabolon_sdata}
\alias{load_metabolon_sdata}
\alias{load_metabolon_fdata}
\alias{load_metabolon}
\alias{load_metabolon_file}
\title{Load metabolon data}
\usage{
load_metabolon_sdata(file, sheet = 2)

load_metabolon_fdata(file, sheet = 2)

load_metabolon(file, sheet = 2, verbose = TRUE,
  add_kegg_pathways = FALSE, add_smiles = FALSE)

load_metabolon_file(...)
}
\arguments{
\item{file}{metabolon xlsx file (character)}

\item{sheet}{number or name}

\item{verbose}{logical}

\item{add_kegg_pathways}{logical}

\item{add_smiles}{logical}

\item{...}{(backward compatibility)}
}
\value{
SummarizedExperiment (load_metabolon) or dataframe (load_metabolon_sdata, load_metabolon_fdata)
}
\description{
Load metabolon data
}
\examples{
require(magrittr)
if (require(subramanian.2016)){
   file <- system.file('extdata/metabolon/subramanian.2016.metabolon.xlsx',
                        package = 'subramanian.2016')
   file \%>\% autonomics.import::load_metabolon(sheet=5)
   file \%>\% autonomics.import::load_metabolon_sdata(sheet=5) \%>\% head(1)
   file \%>\% autonomics.import::load_metabolon_fdata(sheet=5) \%>\% head(1)
}
if (require(halama.2016)){
   file <- system.file('extdata/halama.2016.xlsx', package = 'halama.2016')
   file \%>\% autonomics.import::load_metabolon(sheet=4)
   file \%>\% autonomics.import::load_metabolon_sdata(sheet=4) \%>\% head(1)
   file \%>\% autonomics.import::load_metabolon_fdata(sheet=4) \%>\% head(1)
}
}
