% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_metabolon.R
\name{load_metabolon_exprs}
\alias{load_metabolon_exprs}
\title{Load metabolon exprs}
\usage{
load_metabolon_exprs(file, sheet)
}
\arguments{
\item{file}{path to metabolon file}

\item{sheet}{excell sheet number or name}
}
\value{
sample dataframe
}
\description{
Load metabolon exprs
}
\examples{
if (require(autonomics.data)){
   file <- system.file('extdata/glutaminase/glutaminase.xlsx', package = 'autonomics.data')
   file \%>\% load_metabolon_exprs(2) \%>\% extract(1:3, 1:3)
}
}
