% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_metabolon_lipidomics.R
\name{load_metabolon_lipidomics}
\alias{load_metabolon_lipidomics}
\title{Load metabolon lipidomics}
\usage{
load_metabolon_lipidomics(file, entity = "Species",
  quantity = "Concentrations")
}
\arguments{
\item{file}{path to metabolon lipidomics file}

\item{entity}{'Lipid Class', 'Species', or 'Fatty Acid'}

\item{quantity}{'Concentrations' or 'Compositions'}
}
\value{
SummarizedExperiment
}
\description{
Load metabolon lipidomics
}
\examples{
require(magrittr)
file <- '../../datasets/WCQA-01-18MLCLP-1/WCQA-01-18MLCLP CLP  6-TAB FILE (180710).XLSX'
if (file.exists(file)){
   file \%>\% autonomics.import::load_metabolon_lipidomics(entity = 'Lipid Class', quantity = 'Concentrations')
   file \%>\% autonomics.import::load_metabolon_lipidomics('Lipid Class', 'Compositions')
   file \%>\% autonomics.import::load_metabolon_lipidomics('Species',     'Concentrations')
   file \%>\% autonomics.import::load_metabolon_lipidomics('Species',     'Compositions')
   file \%>\% autonomics.import::load_metabolon_lipidomics('Fatty Acid',  'Concentrations')
   file \%>\% autonomics.import::load_metabolon_lipidomics('Fatty Acid',  'Compositions')
}
}
