% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_omics.R
\name{load_metabolonlipids}
\alias{load_metabolonlipids}
\title{Load metabolonlipids}
\usage{
load_metabolonlipids(file, sheet, log2_transform = TRUE, design_file = NULL,
  infer_design_from_sampleids = FALSE)
}
\arguments{
\item{file}{path to metabolon lipids file}

\item{sheet}{name of excel sheet (any value in METABOLONLIPIDS_SHEETS)}

\item{design_file}{path to sample design file}

\item{infer_from_sampleids}{logical}
}
\value{
SummarizedExperiment
}
\description{
Load data from metabolon complex lipid panel (clp) file
}
\examples{
require(magrittr)
file <- '../../datasets/WCQA-01-18MLCLP-1/WCQA-01-18MLCLP CLP  6-TAB FILE (180710).XLSX'
if (file.exists(file)){
   file \%>\% autonomics.import::load_metabolonlipids('Lipid Class Concentrations')
   file \%>\% autonomics.import::load_metabolonlipids('Lipid Class Compositions')
   file \%>\% autonomics.import::load_metabolonlipids(    'Species Concentrations')
   file \%>\% autonomics.import::load_metabolonlipids(    'Species Compositions')
   file \%>\% autonomics.import::load_metabolonlipids( 'Fatty Acid Concentrations')
   file \%>\% autonomics.import::load_metabolonlipids( 'Fatty Acid Compositions')
}
}
