% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_omics.R
\name{load_metabolonlipids_exprs}
\alias{load_metabolonlipids_exprs}
\title{Load metabolonlipids exprs}
\usage{
load_metabolonlipids_exprs(file, sheet)
}
\arguments{
\item{file}{path to metabolon clp (complex lipid panel) file}

\item{sheet}{name of excel sheet (any value in METABOLONLIPIDS_SHEETS)}
}
\value{
exprs matrix
}
\description{
Load exprs from metabolon clp (complex lipid panel) file
}
\examples{
require(magrittr)
file <- '../../datasets/WCQA-01-18MLCLP-1/WCQA-01-18MLCLP CLP  6-TAB FILE (180710).XLSX'
if (file.exists(file)){
   file \%>\% load_metabolonlipids_exprs('Lipid Class Concentrations') \%>\% extract(1:3,1:3)
   file \%>\% load_metabolonlipids_exprs(    'Species Concentrations') \%>\% extract(1:3,1:3)
   file \%>\% load_metabolonlipids_exprs( 'Fatty Acid Concentrations') \%>\% extract(1:3,1:3)
   file \%>\% load_metabolonlipids_exprs('Lipid Class Compositions')   \%>\% extract(1:3,1:3)
   file \%>\% load_metabolonlipids_exprs(    'Species Compositions')   \%>\% extract(1:3,1:3)
   file \%>\% load_metabolonlipids_exprs( 'Fatty Acid Compositions')   \%>\% extract(1:3,1:3)
}
}
