% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_omics.R
\name{load_metabolonlipids_fdata}
\alias{load_metabolonlipids_fdata}
\title{Load metabolonlipids fdata}
\usage{
load_metabolonlipids_fdata(file, sheet)
}
\arguments{
\item{file}{path to metabolon clp (complex lipid panel) file}

\item{sheet}{name of excel sheet (any value in METABOLONLIPIDS_SHEETS)}
}
\value{
feature dataframe
}
\description{
Load fdata from metabolon clp (complex lipid panel) file
}
\examples{
require(magrittr)
file <- '../../datasets/WCQA-01-18MLCLP-1/WCQA-01-18MLCLP CLP  6-TAB FILE (180710).XLSX'
if (file.exists(file)){
   file \%>\% load_metabolonlipids_fdata('Lipid Class Concentrations') \%>\% head()
   file \%>\% load_metabolonlipids_fdata(    'Species Concentrations') \%>\% head()
   file \%>\% load_metabolonlipids_fdata( 'Fatty Acid Concentrations') \%>\% head()
   file \%>\% load_metabolonlipids_fdata('Lipid Class Compositions')   \%>\% head()
   file \%>\% load_metabolonlipids_fdata(    'Species Compositions')   \%>\% head()
   file \%>\% load_metabolonlipids_fdata( 'Fatty Acid Compositions')   \%>\% head()
}
}
