% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_maxquant.R
\name{load_phosphosite_occupancies}
\alias{load_phosphosite_occupancies}
\title{Load phosphosite occupancies}
\usage{
load_phosphosite_occupancies(phosphosites_file,
  proteingroups_file = paste0(dirname(proteingroups_file),
  "/proteinGroups.txt"), sample_file = paste0(dirname(proteingroups_file),
  "/sample_design.txt"), parameter_file = paste0(dirname(proteingroups_file),
  "/parameters.txt"),
  value_type = infer_maxquant_value_type(proteingroups_file))
}
\arguments{
\item{phosphosites_file}{full path to "Phospho (STY)Sites.txt"}

\item{proteingroups_file}{full path to "proteinGroups.txt"}

\item{sample_file}{full path to "sample_design.txt"}

\item{parameter_file}{full path to "parameters.txt"}

\item{value_type}{any value in autonomics.import::MAXQUANT_VALUE_TYPES}
}
\description{
Load phosphosite occupancies
}
\examples{
require(magrittr)
if (require(billing.differentiation.data)){
   'extdata/maxquant/Phospho (STY)Sites.txt' \%>\%
    system.file(package = 'billing.differentiation.data') \%>\%
    load_phosphosite_occupanciess()
}
}
