% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_maxquant.R
\name{load_phosphosite_occupancies_to_long_dt}
\alias{load_phosphosite_occupancies_to_long_dt}
\title{Load datatable with phosphosite occupancies}
\usage{
load_phosphosite_occupancies_to_long_dt(phosphosites_file,
  proteingroups_file = paste0(dirname(phosphosites_file),
  "/proteinGroups.txt"),
  value_type = infer_maxquant_value_type(proteingroups_file),
  log2_transform = TRUE, min_loc_prob = 0.75)
}
\arguments{
\item{phosphosites_file}{phosphosites  file}

\item{proteingroups_file}{proteingroups file}

\item{value_type}{any value in names(VALUE_COLNAME_PATTERNS)}

\item{log2_transform}{whether to log2 transform}

\item{min_loc_prob}{minimum localization probability (numeric)}
}
\value{
datatable with phosphosite occupancies
}
\description{
Load datatable with phosphosite occupancies
}
\examples{
require(magrittr)
if (require(billing.differentiation.data)){
  system.file('extdata/maxquant/phospho (STY)Sites.txt',
               package = 'billing.differentiation.data') \%>\%
  load_phosphosite_occupancies_to_long_dt() \%>\%
  str()
}
}
