% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_maxquant.R
\name{load_phosphosite_ratios_to_long_dt}
\alias{load_phosphosite_ratios_to_long_dt}
\title{Load datatable with phosphosite ratios}
\usage{
load_phosphosite_ratios_to_long_dt(phosphosites_file,
  value_type = infer_maxquant_value_type(phosphosites_file),
  log2_transform = TRUE, min_loc_prob = 0.75, remove_multiple_pg = FALSE)
}
\arguments{
\item{phosphosites_file}{phosphosites file}

\item{value_type}{any value in names(VALUE_COLNAME_PATTERNS)}

\item{log2_transform}{whether to log2 transform}

\item{min_loc_prob}{mimum localization probability}

\item{remove_multiple_pg}{whether to remove phosphosites mapping to multiple protein groups (logical)}
}
\value{
phosophosites data.table
}
\description{
Load datatable with phosphosite ratios
}
\examples{
# require(magrittr)
# if (require(billing.differentiation.data)){
#  system.file('extdata/maxquant/phospho (STY)Sites.txt',
#               package = 'billing.differentiation.data') \%>\%
#  load_phosphosite_ratios_to_long_dt() \%>\%
#  str()
#}
}
