% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_maxquant.R
\name{load_phosphosites}
\alias{load_phosphosites}
\title{Load phosphosite ratios}
\usage{
load_phosphosites(phosphosites_file,
  sample_file = paste0(dirname(proteingroups_file), "/sample_design.txt"),
  parameter_file = paste0(dirname(proteingroups_file), "/parameters.txt"),
  value_type = autonomics.import::infer_maxquant_value_type(phosphosites_file))
}
\arguments{
\item{phosphosites_file}{full path to "Phospho (STY)Sites.txt"}

\item{sample_file}{full path to "sample_design.txt"}

\item{parameter_file}{full path to "parameters.txt"}

\item{value_type}{any value in autonomics.import::MAXQUANT_VALUE_TYPES}
}
\description{
Load phosphosite ratios
}
\examples{
require(magrittr)
if (require(billing.differentiation.data)){
   'extdata/maxquant/Phospho (STY)Sites.txt' \%>\%
   system.file(package = 'billing.differentiation.data') \%>\%
   load_phosphosites()
}
}
