% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_maxquant.R
\name{load_phosphosites_to_wide_dt}
\alias{load_phosphosites_to_wide_dt}
\title{Load phosphosites file into wide data.table}
\usage{
load_phosphosites_to_wide_dt(phosphosites_file,
  value_type = infer_maxquant_value_type(phosphosites_file),
  min_loc_prob = 0.75, remove_multiple_pg = FALSE)
}
\arguments{
\item{phosphosites_file}{path to phosphoSites file}

\item{value_type}{any value in names(VALUE_COLNAME_PATTERNS)}

\item{min_loc_prob}{minimum localization probability}

\item{remove_multiple_pg}{whether to phosphosites mapping to multiple protein groups (logical)}
}
\description{
Load phosphosites file into wide data.table
}
\examples{
require(magrittr)
if (require(billing.differentiation.data)){
   system.file('extdata/maxquant/phospho (STY)Sites.txt',
                package = 'billing.differentiation.data') \%>\%
      load_phosphosites_to_wide_dt() \%>\%
      str()
}
}
