% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_omics.R
\name{load_proteingroups}
\alias{load_proteingroups}
\title{Load proteingroups}
\usage{
load_proteingroups(file,
  quantity = autonomics.import::infer_maxquant_quantity(file),
  infer_design_from_sampleids = FALSE, design_sep = NULL,
  design_file = NULL, fasta_file = NULL, log2_transform = TRUE,
  log2_offset = 0)
}
\arguments{
\item{file}{path to proteinGroups.txt}

\item{quantity}{'Ratio normalized', 'Ratio', 'Intensity', 'LFQ intensity', 'Reporter intensity'}

\item{infer_design_from_sampleids}{logical: whether to infer design from sampleids}

\item{design_sep}{string: design separator}

\item{design_file}{path to design file (created with write_maxquant_design)}

\item{fasta_file}{path to uniprot fasta database}

\item{log2_transform}{logical: whether to log2 transform}

\item{log2_offset}{numeric: offset used in mapping: x -> log2(offset + x)}
}
\description{
Load proteingroups
}
\examples{
require(magrittr)

# STEM CELL COMPARISON
if (require(autonomics.data)){
   file <- system.file('extdata/stemcomp/maxquant/proteinGroups.txt',
                        package = 'autonomics.data')

   # The sample design can be inferred from the sample ids
      object <- file \%>\% load_proteingroups(infer_design_from_sampleids = TRUE)
      object \%>\% sdata()

   # Or it can be loaded through a sample file
      design_file <- tempfile()
      write_design(file, platform = 'maxquant', infer_design_from_sampleids = TRUE,
                   design_file = design_file)
      object <- load_proteingroups(file, design_file = design_file)
      sdata(object)
}

# STEM CELL DIFFERENTIATION
if (require(autonomics.data)){
   file <- 'extdata/stemdiff/maxquant/proteinGroups.txt' \%>\%
            system.file(package = 'autonomics.data')
   object <- file \%>\% autonomics.import::load_proteingroups(
                         infer_design_from_sampleids = TRUE)
   sdata(object) \%>\% str()
}

}
