% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_maxquant2.R
\name{load_proteingroups_exprs}
\alias{load_proteingroups_exprs}
\title{Load proteingroups exprs}
\usage{
load_proteingroups_exprs(file, pattern)
}
\arguments{
\item{file}{full path to proteinGroups.txt}

\item{value_type}{value in autonomics.import::MAXQUANT_VALUE_TYPES}
}
\description{
Load proteingroups exprs
}
\examples{
 require(magrittr)
 if (require(autonomics.data)){

   # Normalized Ratios
   pattern <- 'Ratio (.+) normalized (.+)'
   file <- 'extdata/stemcell.comparison/maxquant/proteinGroups.txt'  \%>\%
            system.file(package = 'autonomics.data')
   file \%>\% autonomics.import::load_proteingroups_exprs(pattern) \%>\% str()

   # Ratios
   pattern <- 'Ratio .+ .'
}
}
