% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_maxquant2.R
\name{load_proteingroups_fnames}
\alias{load_proteingroups_fnames}
\title{Load proteingroups fnames}
\usage{
load_proteingroups_fnames(file)
}
\arguments{
\item{file}{full path to proteinGroups.txt}
}
\description{
Load proteingroups fnames
}
\examples{
 require(magrittr)
 if (require(autonomics.data)){
   'extdata/stemcell.comparison/maxquant/proteinGroups.txt'  \%>\%
    system.file(package = 'autonomics.data')                 \%>\%
    autonomics.import::load_proteingroups_fnames()           \%>\%
    head()
}
}
