% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_maxquant.R
\name{load_proteingroups_to_long_dt}
\alias{load_proteingroups_to_long_dt}
\title{Load protein table}
\usage{
load_proteingroups_to_long_dt(proteingroups_file,
  value_type = autonomics.import::infer_maxquant_value_type(proteingroups_file),
  log2_transform = TRUE)
}
\arguments{
\item{proteingroups_file}{proteingroups file}

\item{value_type}{any value in names(VALUE_COLNAME_PATTERNS)}

\item{log2_transform}{whether to log2 transform}
}
\value{
proteingroups data.table
}
\description{
Load protein table
}
\examples{
# require(magrittr)
# if (require(autonomics.data)){
#  proteingroups_file <- system.file(
#     'extdata/stemdiff/maxquant/proteinGroups.txt',
#     package = 'autonomics.data')
#  proteingroups_file \%>\% autonomics.import::load_proteingroups_to_long_dt() \%>\% str()
#}
#if (require(graumann.lfq)){
#  system.file('extdata/proteinGroups.txt', package = 'graumann.lfq') \%>\%
#     autonomics.import::load_proteingroups_to_long_dt() \%>\% str()
#}
#if (require(billing.vesicles)){
#  system.file('extdata/proteinGroups.txt', package = 'billing.vesicles') \%>\%
#     autonomics.import::load_proteingroups_to_long_dt() \%>\% str()
#}
}
