% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_load_maxquant.R
\name{load_proteingroups_to_wide_dt}
\alias{load_proteingroups_to_wide_dt}
\title{Load proteingroups file into wide data.table}
\usage{
load_proteingroups_to_wide_dt(proteingroups_file,
  value_type = infer_maxquant_value_type(proteingroups_file))
}
\arguments{
\item{proteingroups_file}{path to proteinGroups file}

\item{value_type}{any value in names(VALUE_COLNAME_PATTERNS)}
}
\description{
Load proteingroups file into wide data.table
}
\examples{
#require(magrittr)
#
# NORMALIZED RATIOS
#if (require(autonomics.data)){
#   proteingroups_file <- system.file(
#      'extdata/stemdiff/maxquant/proteinGroups.txt',
#       package = 'autonomics.data')
#   proteingroups_file \%>\% load_proteingroups_to_wide_dt() \%>\% dim()
#}

# REPORTER INTENSITIES
#if (require(billing.vesicles)){
#   proteingroups_file <- system.file(
#      'extdata/proteinGroups.txt', package = 'billing.vesicles')
#   proteingroups_file \%>\% load_proteingroups_to_wide_dt() \%>\% dim()
#}

# RAW RATIOS
#if (require(billing.differentiation.data)){
#   proteingroups_file <- system.file(
#      'extdata/maxquant/proteinGroups.txt', package = 'billing.differentiation.data')
#   proteingroups_file \%>\% load_proteingroups_to_wide_dt(value_type = 'raw.ratio') \%>\% dim()
#}

# LFQ INTENSITIES
#if (require(graumann.lfq)){
#   proteingroups_file <- system.file(
#      'extdata/proteinGroups.txt', package = 'graumann.lfq')
#   proteingroups_file \%>\% load_proteingroups_to_wide_dt() \%>\% dim()
#}

# RAW INTENSITIES
#if (require(alnoubi.2017)){
#   proteingroups_file <- system.file(
#      'extdata/proteinGroups.txt', package = 'alnoubi.2017')
#   proteingroups_file \%>\% load_proteingroups_to_wide_dt(value_type = 'raw.intensity') \%>\% dim()

#    proteingroups_file <- system.file(
#      'extdata/proteinGroups.txt', package = 'alnoubi.2017')
#   proteingroups_file \%>\% load_proteingroups_to_wide_dt() \%>\% dim()
#}
}
