% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rnaseq.R
\name{load_rnaseq}
\alias{load_rnaseq}
\title{Load RNAseq cpm}
\usage{
load_rnaseq(file, design_file = NULL, infer_design_from_sampleids = FALSE,
  design_sep = NULL)
}
\arguments{
\item{file}{rnaseq counts file}

\item{design_file}{sample design file}

\item{infer_design_from_sampleids}{logical}

\item{design_sep}{character}
}
\description{
Load RNAseq cpm
}
\examples{
require(magrittr)
if (require(autonomics.data){
   file <- system.file('extdata/stemdiff/rnaseq/gene_counts.txt', package = 'autonomics.data')
})
if (require(subramanian.2016)){
   file <- system.file('extdata/rnaseq/gene_counts.txt', package = 'subramanian.2016')
   load_rnaseq(file, infer_design_from_sampleids = TRUE)
}
}
