% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rnaseq.R
\name{load_rnaseq_counts}
\alias{load_rnaseq_counts}
\title{Load RNA seq counts}
\usage{
load_rnaseq_counts(dir = ".", count_file = paste0(dir, "/gene_counts.txt"),
  sample_file = if (file.exists(paste0(dir, "/sample_design.txt")))
  paste0(dir, "/sample_design.txt") else NULL, map_entrezg = FALSE)
}
\arguments{
\item{dir}{directory with count file and sample design file}

\item{count_file}{gene count file}

\item{sample_file}{sample design file}

\item{map_entrezg}{logical}
}
\value{
eset
}
\description{
Load RNA seq counts
}
\examples{
require(magrittr)
if (require(billing.differentiation.data)){
   dir <- system.file('extdata/rnaseq', package = 'billing.differentiation.data')
   autonomics.import::load_rnaseq_counts(dir)
}
if (require(subramanian.2016)){
   dir <- system.file('extdata/rnaseq', package = 'subramanian.2016')
   autonomics.import::load_rnaseq_counts(dir, sample_file = NULL)
}
}
