% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rnaseq.R
\name{load_rnaseq_exprs}
\alias{load_rnaseq_exprs}
\title{Load rnaseq exprs}
\usage{
load_rnaseq_exprs(count_file)
}
\arguments{
\item{count_file}{count file}
}
\value{
count matrix
}
\description{
Load rnaseq exprs
}
