% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_load_rnaseq.R
\name{load_rnaseq_sdata}
\alias{load_rnaseq_sdata}
\title{Required variables in sample design file
Load sample design file}
\usage{
load_rnaseq_sdata(sample_design_file)
}
\arguments{
\item{sample_design_file}{sample design file}
}
\value{
sample design dataframe
}
\description{
Required variables in sample design file
Load sample design file
}
\examples{
if (require(billing.differentiation.data)){
   sample_design_file <- system.file('extdata/rnaseq/sample_design.txt',
                            package = 'billing.differentiation.data')
   autonomics.import::load_rnaseq_sdata(sample_design_file)
}
}
