% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_sdata.R
\name{load_sample_file}
\alias{load_sample_file}
\title{Load sample design}
\usage{
load_sample_file(sample_file)
}
\arguments{
\item{sample_file}{sample design file}
}
\value{
sample design datatable
}
\description{
Load sample design
}
\examples{
if (require(autonomics.data)){
   sample_file <- system.file('extdata/billing2016/sample_design.txt',
                                      package = 'autonomics.data')
   autonomics.import::load_sample_file(sample_file)
}
if (require(billing.differentiation.data)){
   sample_file <- system.file('extdata/maxquant/sample_design.txt',
                                      package = 'billing.differentiation.data')
   load_sample_file(sample_file)
}
}
