% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load___sdata.R
\name{load_sdata_exiqon}
\alias{load_sdata_exiqon}
\title{Load exiqon sdata}
\usage{
load_sdata_exiqon(file)
}
\arguments{
\item{file}{path to exiqon xls file}
}
\value{
sample dataframe
}
\description{
Load exiqon sdata
}
\examples{
 require(magrittr)
 if (require(subramanian.2016)){
    file <- system.file('extdata/exiqon/subramanian.2016.exiqon.xlsx',
                         package = 'subramanian.2016')
    file \%>\% autonomics.import::load_sdata_exiqon()
 }
}
