% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load___sdata.R
\name{load_sdata_maxquant}
\alias{load_sdata_maxquant}
\title{Load maxquant sdata}
\usage{
load_sdata_maxquant(file,
  quantity = autonomics.import::infer_maxquant_quantity(file))
}
\arguments{
\item{file}{path to maxquant file}

\item{quantity}{string: quantity to be used for exprs}
}
\value{
sample dataframe
}
\description{
Load maxquant sdata
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   file <- 'extdata/stemcomp/maxquant/proteinGroups.txt' \%>\%
            system.file(package = 'autonomics.data')
   file \%>\% load_sdata_maxquant()
}
if (require(graumann.lfq)){
   file <- system.file('extdata/proteinGroups.txt', package = 'graumann.lfq')
   file \%>\% load_sdata_maxquant()
}
}
