% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_omics.R
\name{load_sdata_metabolon}
\alias{load_sdata_metabolon}
\title{Load metabolon sdata}
\usage{
load_sdata_metabolon(file, sheet)
}
\arguments{
\item{file}{path to metabolon file}

\item{sheet}{excell sheet number or name}
}
\value{
sample dataframe
}
\description{
Load metabolon sdata
}
\examples{
if (require(autonomics.data)){
   file <- system.file('extdata/glutaminase/glutaminase.xlsx', package = 'autonomics.data')
   file \%>\% load_sdata_metabolon(2) \%>\% extract(1:3, 1:3)
}
}
