% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load___sdata.R
\name{load_sdata_metabolonlipids}
\alias{load_sdata_metabolonlipids}
\title{Load metabolonlipids sdata}
\usage{
load_sdata_metabolonlipids(file, sheet)
}
\arguments{
\item{file}{path to metabolon clp (complex lipid panel) file}

\item{sheet}{name of excel sheet (any value in METABOLONLIPIDS_SHEETS)}
}
\value{
sample dataframe
}
\description{
Load sdata from metabolon clp (complex lipid panel) file
}
\examples{
require(magrittr)
file <- '../../datasets/WCQA-01-18MLCLP-1/WCQA-01-18MLCLP CLP  6-TAB FILE (180710).XLSX'
if (file.exists(file)){
   file \%>\% load_sdata_metabolonlipids('Lipid Class Concentrations') \%>\% head()
   file \%>\% load_sdata_metabolonlipids(    'Species Concentrations') \%>\% head()
   file \%>\% load_sdata_metabolonlipids( 'Fatty Acid Concentrations') \%>\% head()
   file \%>\% load_sdata_metabolonlipids('Lipid Class Compositions')   \%>\% head()
   file \%>\% load_sdata_metabolonlipids(    'Species Compositions')   \%>\% head()
   file \%>\% load_sdata_metabolonlipids( 'Fatty Acid Compositions')   \%>\% head()
}
}
