% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rnaseq.R
\name{load_sdata_rnaseq}
\alias{load_sdata_rnaseq}
\title{Load sample design file}
\usage{
load_sdata_rnaseq(sample_design_file)
}
\arguments{
\item{sample_design_file}{sample design file}
}
\value{
sample design dataframe
}
\description{
Load sample design file
}
\examples{
if (require(billing.differentiation.data)){
   sample_design_file <- system.file('extdata/rnaseq/sample_design.txt',
                            package = 'billing.differentiation.data')
   autonomics.import::load_sdata_rnaseq(sample_design_file)
}
}
