% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_omics.R
\name{load_sdata_soma}
\alias{load_sdata_soma}
\title{Load soma sdata}
\usage{
load_sdata_soma(file)
}
\arguments{
\item{file}{string: path to adat file}
}
\value{
sample dataframe
}
\description{
Load soma sdata
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   file <- system.file('extdata/stemcell.comparison/stemcell.comparison.adat',
                        package = 'autonomics.data')
   file \%>\% autonomics.import::load_sdata_soma() \%>\% head()
}
if (require(atkin.2014)){
   file <- system.file('extdata/soma/WCQ-14-130_Set_A_RPT.HybMedNormCal_20140925.adat',
                        package = 'atkin.2014')
   file \%>\% autonomics.import::load_sdata_soma() \%>\% head()
}
}
\author{
Aditya Bhagwat
}
