% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load2_sdata.R
\name{load_snames_maxquant}
\alias{load_snames_maxquant}
\title{Load maxquant snames}
\usage{
load_snames_maxquant(file,
  quantity = autonomics.import::infer_maxquant_quantity(file),
  clean = FALSE, designify = FALSE)
}
\arguments{
\item{file}{path to maxquant file}
}
\description{
Load maxquant snames
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   file <- 'extdata/stemcell.comparison/maxquant/proteinGroups.txt' \%>\%
            system.file(package = 'autonomics.data')

   file \%>\% load_snames_maxquant('Ratio')
   file \%>\% load_snames_maxquant('Ratio', clean = TRUE)
   file \%>\% load_snames_maxquant('Ratio', clean = TRUE, designify = TRUE)

   file \%>\% load_snames_maxquant('Intensity')
   file \%>\% load_snames_maxquant('Intensity', clean = TRUE)
   file \%>\% load_snames_maxquant('Intensity', clean = TRUE, designify = TRUE)
}
if (require(graumann.lfq)){
   file <- system.file('extdata/proteinGroups.txt', package = 'graumann.lfq')
   file \%>\% load_snames_maxquant('Intensity')
   file \%>\% load_snames_maxquant('Intensity', clean = TRUE)
   file \%>\% load_snames_maxquant('Intensity', clean = TRUE, designify = TRUE)
   file \%>\% load_snames_maxquant('LFQ intensity')
   file \%>\% load_snames_maxquant('LFQ intensity', clean = TRUE)
   file \%>\% load_snames_maxquant('LFQ intensity', clean = TRUE, designify = TRUE)
}
}
