% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_soma.R
\name{load_soma_exprs}
\alias{load_soma_exprs}
\title{Load soma exprs}
\usage{
load_soma_exprs(file)
}
\arguments{
\item{file}{string: path to adat file}
}
\value{
sample dataframe
}
\description{
Load soma exprs
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   file <- system.file('extdata/stemcell.comparison/stemcell.comparison.adat',
                        package = 'autonomics.data')
   file \%>\% autonomics.import::load_soma_exprs()
}
}
\author{
Aditya Bhagwat
}
