% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_soma.R
\name{load_soma_fdata}
\alias{load_soma_fdata}
\title{Load soma fdata}
\usage{
load_soma_fdata(file)
}
\arguments{
\item{file}{string: path to adat file}
}
\value{
feature dataframe
}
\description{
Load soma fdata
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   file <- system.file('extdata/stemcell.comparison/stemcell.comparison.adat',
                        package = 'autonomics.data')
   file \%>\% autonomics.import::load_soma_fdata() \%>\% head()
}
if (require(atkin.2014)){
   file <- system.file('extdata/soma/WCQ-14-130_Set_A_RPT.HybMedNormCal_20140925.adat',
                        package = 'atkin.2014')
   file \%>\% autonomics.import::load_soma_fdata()
}
}
\author{
Aditya Bhagwat
}
