% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutators.R
\name{log2,SummarizedExperiment-method}
\alias{log2,SummarizedExperiment-method}
\title{log2 transform assayData matrices}
\usage{
\S4method{log2}{SummarizedExperiment}(x)
}
\arguments{
\item{x}{SummarizedExperiment}
}
\description{
log2 transform assayData
}
