% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_load_rnaseq.R
\name{logcpm}
\alias{logcpm}
\title{Convert raw counts into log cpm}
\usage{
logcpm(object)
}
\arguments{
\item{object}{eset}
}
\value{
log cpm
}
\description{
Convert raw counts into log cpm in exactly the same way as limma::voom does
}
\examples{
require(magrittr)
if (require(billing.differentiation.data)){
   eset1 <- billing.differentiation.data::rna.counts
   eset1 \%>\% autonomics.import::exprs()    \%>\% magrittr::extract(1:3, 1:3)
   eset1 \%>\% autonomics.import::logcpm() \%>\%
             autonomics.import::exprs() \%>\% magrittr::extract(1:3, 1:3)
}
}
