% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_omics.R
\name{make_gtf_url}
\alias{make_gtf_url}
\title{Make link to GTF file}
\usage{
make_gtf_url(organism, release)
}
\arguments{
\item{organism}{'Homo sapiens', 'Mus musculus', or 'Rattus norvegicus'}

\item{release}{number}
}
\description{
Make link to GTF file
}
\examples{
make_gtf_url(organism = 'Homo sapiens', release = 95)
make_gtf_url(organism = 'Mus musculus', release = 95)
}
