% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_maxquant_evidences.R
\name{matrixify_summarized_evidences}
\alias{matrixify_summarized_evidences}
\title{Matrixify summarized evidences}
\usage{
matrixify_summarized_evidences(summarized_evidences, sample_file,
  var = "Ratio")
}
\arguments{
\item{summarized_evidences}{data.table with summarized evidences}

\item{sample_file}{file with sample annotation}

\item{var}{variable to extract and matrixify}
}
\description{
Convert summarized evidences from long data.table into
feature x sample matrix
}
\examples{
require(magrittr)
\dontrun{
if (require(billing.differentiation.data)){
   evidence_file <- system.file(
      'extdata/maxquant/evidence.txt',
      package = 'billing.differentiation.data')
   sample_file <- system.file(
      'extdata/maxquant/sample_design.txt',
      package = 'billing.differentiation.data')
   summarized_evidences <- evidences_to_proteingroups(evidence_file)
   summarized_evidences \%>\% matrixify_summarized_evidences(sample_file)
}
}
}
