% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_maxquant.R
\name{maxquant_dt_to_sumexp}
\alias{maxquant_dt_to_sumexp}
\title{Convert long data table into eset}
\usage{
maxquant_dt_to_sumexp(DT, entity, quantity, design_file, parameter_file)
}
\arguments{
\item{DT}{long datatable}

\item{entity}{entity}

\item{quantity}{quantity}

\item{design_file}{sample design file}

\item{parameter_file}{parameter file}
}
\description{
Convert long data table into eset
}
\examples{
#require(magrittr)
#if (require(autonomics.data)){
#   maxquant_dir       <- system.file('extdata/stemdiff/maxquant',
#                                      package = 'autonomics.data')
#   proteingroups_file <- paste0(maxquant_dir, '/proteinGroups.txt')
#   phosphosites_file  <- paste0(maxquant_dir, '/phospho (STY)Sites.txt')
#   parameter_file     <- paste0(maxquant_dir, '/parameters.txt')
#   design_file <- tempfile()
#   proteingroups_file \%>\% autonomics.import::write_maxquant_design(infer_from_sampleids = TRUE,
#                                                                   design_file = design_file)
#   DT <- proteingroups_file \%>\% autonomics.import::load_proteingroups_to_long_dt()
#   DT \%>\% autonomics.import::maxquant_dt_to_sumexp(
#             entity = 'proteingroup', quantity = 'normalized.ratio',
#             design_file, parameter_file)
#   DT <- phosphosites_file \%>\% autonomics.import::load_phosphosite_ratios_to_long_dt()
#   DT \%>\% autonomics.import::maxquant_dt_to_sumexp(
#             entity = 'phosphosite', quantity = 'normalized.ratio',
#             design_file, parameter_file)
#   DT <- autonomics.import::load_phosphosite_occupancies_to_long_dt(phosphosites_file)
#   DT \%>\% autonomics.import::maxquant_dt_to_sumexp(
#             entity = 'phosphosite', quantity = 'occupancy',
#             design_file, parameter_file)
#}
#if (require(graumann.lfq)){
#   maxquant_dir <- system.file('extdata', package = 'graumann.lfq')
#   proteingroups_file <- paste0(maxquant_dir, '/proteinGroups.txt')
#   design_file <- paste0(maxquant_dir,    '/sample_design.txt')
#   parameter_file <- paste0(maxquant_dir, '/parameters.txt')
#   DT <- autonomics.import::load_proteingroups_to_long_dt(proteingroups_file)
#   DT \%>\% autonomics.import::maxquant_dt_to_sumexp(
#             entity = 'proteingroup', quantity = 'lfq.intensity', design_file, parameter_file)
#}
}
