% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_omics.R
\docType{data}
\name{maxquant_patterns}
\alias{maxquant_patterns}
\title{maxquant patterns}
\format{An object of class \code{character} of length 7.}
\usage{
maxquant_patterns
}
\description{
maxquant patterns
}
\keyword{datasets}
