% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_load_maxquant.R
\docType{data}
\name{MAXQUANT_VALUE_TYPES}
\alias{MAXQUANT_VALUE_TYPES}
\title{Max Quant value types}
\format{An object of class \code{character} of length 5.}
\usage{
MAXQUANT_VALUE_TYPES
}
\description{
Max Quant value types
}
\keyword{datasets}
