% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_load_rnaseq.R
\name{mean_var_monotonic}
\alias{mean_var_monotonic}
\title{Is mean var trend monotonic}
\usage{
mean_var_monotonic(object, normalize.method)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{normalize.method}{normalization method}
}
\value{
logical
}
\description{
Is mean var trend monotonic
}
