% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{merge_sdata}
\alias{merge_sdata}
\title{Merge sdata into SummarizedExperiment}
\usage{
merge_sdata(object, newdata, by, newdata_first = FALSE)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{newdata}{dataframe}

\item{by}{variable by which to merge}

\item{newdata_first}{logical: whether to have newdata first in sdata}
}
\description{
Merge sdata into SummarizedExperiment
}
\details{
Common variables (other than by) are removed from newdata prior to merge.
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   object <- system.file('extdata/stemcell.comparison/stemcell.comparison.adat',
                          package = 'autonomics.data') \%>\%
             autonomics.import::load_soma()
   newdata <- autonomics.import::sdata(object) \%>\%
              magrittr::extract(, 'SampleId', drop = FALSE) \%>\%
              cbind(Letter = letters[1:nrow(.)])
   object \%>\% autonomics.import::sdata()
   object \%>\% autonomics.import::merge_sdata(newdata, by = 'SampleId') \%>\%
              autonomics.import::sdata()
   object \%>\% autonomics.import::merge_sdata(newdata, by = 'SampleId', newdata_first = TRUE) \%>\%
              autonomics.import::sdata()
}
}
