% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load2_sdata.R
\docType{data}
\name{METABOLONLIPIDS_SHEETS}
\alias{METABOLONLIPIDS_SHEETS}
\title{Possible sheet values for metabolonlipids data}
\format{An object of class \code{character} of length 6.}
\usage{
METABOLONLIPIDS_SHEETS
}
\description{
Possible sheet values for metabolonlipids data
}
\keyword{datasets}
