% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess.R
\name{na_aware_equals}
\alias{na_aware_equals}
\title{Convenient equals operator}
\usage{
na_aware_equals(x, y)
}
\arguments{
\item{x}{numeric vector or scalar}

\item{y}{numeric scalar}
}
\description{
Performs x == y, but returns FALSE rather than NA for NA elements of x.
}
\examples{
require(magrittr)
x <- c(A=1,B=3,C=2,D=3, E=NA)
y <- 3
x \%>\% equals(y)
}
