% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_sdata.R
\name{nameify_strings}
\alias{nameify_strings}
\title{Nameify strings}
\usage{
nameify_strings(x, verbose = TRUE)
}
\arguments{
\item{x}{character vector}

\item{verbose}{character}
}
\value{
validified subgroups
}
\description{
Nameify strings
}
\examples{
require(magrittr)
}
