% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_maxquant.R
\name{old_load_proteingroups}
\alias{old_load_proteingroups}
\title{Load proteingroup ratios/intensities}
\usage{
old_load_proteingroups(proteingroups_file,
  design_file = paste0(dirname(proteingroups_file), "/sample_design.txt"),
  parameter_file = paste0(dirname(proteingroups_file), "/parameters.txt"),
  log2_transform = TRUE,
  value_type = autonomics.import::infer_maxquant_value_type(proteingroups_file))
}
\arguments{
\item{proteingroups_file}{full path to "proteinGroups.txt"}

\item{design_file}{full path to "sample_design.txt"}

\item{parameter_file}{full path to "parameters.txt"}

\item{log2_transform}{logical}

\item{value_type}{any value in autonomics.import::MAXQUANT_VALUE_TYPES}
}
\value{
data.table or ExpressionSet
}
\description{
Load proteingroup ratios/intensities
}
\examples{
#library(magrittr)
#if (require(autonomics.data)){
#
#   # STEM CELL COMPARISON
#   proteingroups_file <- system.file('extdata/stemcomp/maxquant/proteinGroups.txt',
#                                      package = 'autonomics.data')
#   design_file <- tempfile()
#   autonomics.import::write_maxquant_design(proteingroups_file,
#      infer_from_sampleids = TRUE, design_file = design_file)
#   proteingroups_file \%>\% autonomics.import::load_proteingroups(design_file = design_file)
#
#   # STEM CELL DIFFERENTIATION
#   proteingroups_file <- system.file(
#      'extdata/stemdiff/maxquant/proteinGroups.txt',
#      package = 'autonomics.data')
#   design_file <- tempfile()
#   autonomics.import::write_maxquant_design(proteingroups_file,
#      design_file = design_file, infer_from_sampleids = TRUE)
#   proteingroups_file \%>\% autonomics.import::load_proteingroups(design_file = design_file)
#}
#if (require(billing.vesicles)){
#   proteingroups_file <- system.file('extdata/proteinGroups.txt',
#                                      package = 'billing.vesicles')
#   proteingroups_file \%>\% autonomics.import::load_proteingroups()
#}
#if (require(graumann.lfq)){
#   proteingroups_file <- system.file('extdata/proteinGroups.txt',
#                                      package = 'graumann.lfq') \%>\%
#   proteingroups_file \%>\% autonomics.import::load_proteingroups()
#}
#if (require(graumann.zebra)){
#   proteingroups_file <- system.file('extdata/proteinGroups.txt',
#                                      package = 'graumann.zebra')
#   proteingroups_file \%>\% autonomics.import::load_proteingroups()
#}
}
