% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_soma.R
\name{old_load_soma}
\alias{old_load_soma}
\title{Old function to load soma data}
\usage{
old_load_soma(soma_file, sample_file = NULL, log2_transform = TRUE,
  keep_only_passes = TRUE, keep_only_samples = TRUE, infer_design = FALSE,
  rm_na_svars = TRUE, rm_single_value_svars = TRUE)
}
\arguments{
\item{soma_file}{adat file}

\item{sample_file}{NULL or character (file with sample design information)}

\item{log2_transform}{logical: whether to log2 transform (logical)}

\item{keep_only_passes}{logical: whether to keep only features and samples that pass QC}

\item{keep_only_samples}{logical: whether to keep only biological samples (and rm QC, buffer, and calibrator samples).}

\item{infer_design}{logical: whethe to infer design from sample_ids}

\item{rm_na_svars}{logical: whether to rm na svars}

\item{rm_single_value_svars}{logical: whether to rm single value svars}
}
\description{
Loads data from soma file.
Extracts sample_id definitions from 'SampleId'    column
Extracts subgroup definitions  from 'SampleGroup' column when available.
When not available, attempts to extract subgroup definitions from 'sample_id'column.
}
\examples{
require(magrittr)
if (require(autonomics.data)){

   # Loading soma is easy
   soma_file <- system.file('extdata/stemcell.comparison/stemcell.comparison.adat',
                             package = 'autonomics.data')
   soma_file \%>\% autonomics.import::old_load_soma()

   # Three ways to specify sample design
      # Taken from from 'SampleId' column in soma file
      soma_file \%>\% autonomics.import::old_load_soma() \%>\% autonomics.import::sdata() \%>\% head()

      # Inferred from sample ids
      soma_file \%>\% autonomics.import::old_load_soma(infer_design = TRUE) \%>\%
                    autonomics.import::sdata() \%>\% head()

      # Specified through sample file
      sample_file <- tempfile()
      soma_file \%>\% autonomics.import::create_soma_sample_file(sample_file = sample_file, infer_design = TRUE)
      soma_file \%>\% autonomics.import::old_load_soma(sample_file = sample_file) \%>\%
                    autonomics.import::sdata() \%>\% head()
}

}
