% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_proteingroups.R
\name{open_uniprot_connection}
\alias{open_uniprot_connection}
\title{Open uniprot connection}
\usage{
open_uniprot_connection(object)
}
\arguments{
\item{object}{SummarizedExperiment}
}
\value{
uniprot webservice connection
}
\description{
Open uniprot connection
}
\examples{
\dontrun{
   require(magrittr)
   if (require(autonomics.data)){
      object <- autonomics.data::stemcomp.proteinratios
      object \%>\% open_uniprot_connection()
   }
}
}
