% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eset_specific_var.R
\name{oraid_var}
\alias{oraid_var}
\alias{oraid_values}
\title{Get ora id var/values}
\usage{
oraid_var(object)

oraid_values(object)
}
\arguments{
\item{object}{eset}
}
\description{
Get ora id var/values
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   object <- autonomics.data::billing2016
   object \%>\% autonomics.import::oraid_var()
   object \%>\% autonomics.import::oraid_values() \%>\% head(3)
}
if (require(billing.differentiation.data)){
   object <- billing.differentiation.data::rna.voomcounts
   object \%>\% autonomics.import::oraid_var()
   object \%>\% autonomics.import::oraid_values() \%>\% head(3)

   object <- billing.differentiation.data::protein.ratios
   object \%>\% autonomics.import::oraid_var()
   object \%>\% autonomics.import::oraid_values()
}
if (require(atkin.2014)){
   object <- atkin.2014::soma
   object \%>\% autonomics.import::oraid_var()
   object \%>\% autonomics.import::oraid_values() \%>\% head(3)
}
if (require(subramanian.2016)){
   object <- subramanian.2016::metabolon
   object \%>\% autonomics.import::oraid_var()
   object \%>\% autonomics.import::oraid_values() \%>\% head(3)

   object <- subramanian.2016::exiqon
   object \%>\% autonomics.import::oraid_var()
   object \%>\% autonomics.import::oraid_values() \%>\% head(3)
}
}
