% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_load_maxquant.R
\docType{data}
\name{PHOSPHO_VARS_AUTONOMICS}
\alias{PHOSPHO_VARS_AUTONOMICS}
\title{PHOSPHO_VARS_AUTONOMICS}
\format{An object of class \code{character} of length 9.}
\usage{
PHOSPHO_VARS_AUTONOMICS
}
\description{
PHOSPHO_VARS_AUTONOMICS
}
\keyword{datasets}
