% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_omics.R
\docType{data}
\name{phosphosite_fvars}
\alias{phosphosite_fvars}
\title{phosphosites fvars}
\format{An object of class \code{character} of length 4.}
\usage{
phosphosite_fvars
}
\description{
phosphosites fvars
}
\keyword{datasets}
