% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_design.R
\name{prepare_design}
\alias{prepare_design}
\title{Standardize design in sdata}
\usage{
prepare_design(object, sampleid_var, subgroup_var = NULL, infer_design = if
  (is.null(subgroup_var)) TRUE else FALSE)
}
\arguments{
\item{object}{Summarizedexperiment}

\item{sampleid_var}{sampleid svar}

\item{subgroup_var}{subgroup svar}

\item{infer_design}{logical}
}
\value{
SummarizedExperiment
}
\description{
Standardize design in sdata
}
