% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters.R
\docType{methods}
\name{prepro<-}
\alias{prepro<-}
\alias{prepro<-,SummarizedExperiment,list-method}
\alias{prepro<-,SummarizedExperiment,NULL-method}
\alias{prepro<-,eSet,list-method}
\alias{prepro<-,EList,list-method}
\alias{prepro}
\alias{prepro,SummarizedExperiment-method}
\alias{prepro,eSet-method}
\alias{prepro,EList-method}
\title{Get/Set prepro}
\usage{
prepro(object) <- value

\S4method{prepro}{SummarizedExperiment,list}(object) <- value

\S4method{prepro}{SummarizedExperiment,`NULL`}(object) <- value

\S4method{prepro}{eSet,list}(object) <- value

\S4method{prepro}{EList,list}(object) <- value

prepro(object)

\S4method{prepro}{SummarizedExperiment}(object)

\S4method{prepro}{eSet}(object)

\S4method{prepro}{EList}(object)
}
\arguments{
\item{object}{eSet}

\item{value}{list}
}
\value{
preprocessing details (get) or updated eset
}
\description{
Get/Set preprocessing details
}
