% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_exiqon.R
\name{prepro_exiqon}
\alias{prepro_exiqon}
\title{Preprocess Exiqon SumExp}
\usage{
prepro_exiqon(object, filter_features = "`#RefGenes`==0 & `#Spike`   ==0",
  align_sample_means = TRUE, lod = 37, zero_consistent_nas = TRUE,
  plot = TRUE)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{filter_features}{character(1) with feature filter condition}

\item{align_sample_means}{logical: whether to align the sample maens}

\item{lod}{numeric(1): lod Ct value}

\item{zero_consistent_nas}{logical(1)}

\item{plot}{logical: whether to plot sample distributions}
}
\value{
SummarizedExperiment
}
\description{
Preprocess Exiqon SumExp
}
\examples{
require(magrittr)
if (require(subramanian.2016)){
   file <- system.file('extdata/exiqon/subramanian.2016.exiqon.xlsx',
                        package = 'subramanian.2016')
   file \%>\% autonomics.import::load_exiqon(infer_design_from_sampleids = TRUE) \%>\%
            autonomics.import::prepro_exiqon(lod=36)
}
}
