% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rnaseq.R
\name{prepro_rnaseq}
\alias{prepro_rnaseq}
\title{Preprocess RNAseq counts}
\usage{
prepro_rnaseq(object, plot = TRUE)
}
\arguments{
\item{object}{SummarizedExperiment}
}
\description{
Preprocess RNAseq counts
}
\examples{
require(magrittr)
if (require(subramanian.2016)){
   object <- 'extdata/rnaseq/gene_counts.txt' \%>\%
              system.file(package = 'subramanian.2016') \%>\%
              autonomics.import::load_rnaseq(infer_design_from_sampleids = TRUE)
   object \%<>\% prepro_rnaseq()
   object \%>\% autonomics.plot::plot_overlayed_sample_distributions()
}
}
