% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_maxquant.R
\docType{data}
\name{PROTEIN_VARS_AUTONOMICS}
\alias{PROTEIN_VARS_AUTONOMICS}
\title{PROTEIN_VARS_AUTONOMICS}
\format{An object of class \code{character} of length 5.}
\usage{
PROTEIN_VARS_AUTONOMICS
}
\description{
PROTEIN_VARS_AUTONOMICS
}
\keyword{datasets}
