% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_load_maxquant.R
\docType{data}
\name{PROTEIN_VARS_MAXQUANT}
\alias{PROTEIN_VARS_MAXQUANT}
\title{PROTEIN_VARS_MAXQUANT}
\format{An object of class \code{character} of length 5.}
\usage{
PROTEIN_VARS_MAXQUANT
}
\description{
PROTEIN_VARS_MAXQUANT
}
\keyword{datasets}
