% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_omics.R
\name{read_affy}
\alias{read_affy}
\title{Read affymetrix microarray}
\usage{
read_affy(celfiles)
}
\arguments{
\item{celfiles}{string vector: CEL file paths}
}
\description{
Read affymetrix microarray
}
\examples{
\dontrun{ # since it requires to be online
   if (require(autonomics.data)){

      # download cel files
      require(magrittr)
      url <- 'http://www.bioconductor.org/help/publications/2003/Chiaretti/chiaretti2/T33.tgz'
      download.file(url, destfile = file.path(tempdir(), basename(url)))
      untar(file.path(tempdir(), 'T33.gz'), exdir = tempdir())
      unlink('T33.gz')

      # read
      read_affy(celfiles = list.files(file.path(tempdir(), 'T33'), full.names = TRUE))
   }
}
}
