% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_omics.R
\name{read_asis}
\alias{read_asis}
\title{Read omics data as is from file}
\usage{
read_asis(file, sheet)
}
\arguments{
\item{file}{character(1)}

\item{sheet}{numeric(1) or character(1)}
}
\value{
data.table
}
\description{
Read omics data as is from file
}
