% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_omics.R
\name{read_bam}
\alias{read_bam}
\title{Read BAM files}
\usage{
read_bam(bamdir, ispaired = FALSE, gtffile = NULL,
  fvars = c("gene_name", "gene_biotype"), sumexpfile = NULL,
  nthreads = parallel::detectCores(), ...)
}
\arguments{
\item{bamdir}{string: path to SAM or BAM file directory (one SAM or BAM file per sample)}

\item{ispaired}{TRUE or FALSE: paired end reads?}

\item{gtffile}{string: path to GTF file. Alternatively NULL: then Rsubread's inbuilt annotations are used - only for human and mouse.}

\item{fvars}{string vector: feature variables to include in object. Subset of c('Chr', 'Start', 'End', 'Strand', 'Length', 'gene_name', 'gene_biotype')}

\item{sumexpfile}{string: file which object is saved to (with saveRDS)}

\item{nthreads}{number: number of cores to be used by Rsubread::featureCounts()}

\item{...}{passed to Rsubread::featureCounts}
}
\description{
Read BAM files, count reads, and pack into SummarizedExperiment
}
\examples{
\dontrun{ # requires internet: slow and doesn't always work (https://stackoverflow.com/questions/55532102)
          # requires rsubread (which is not available on windows)

   # Download example data and unzip
     url <- "https://bitbucket.org/graumannlab/billing.stemcells/downloads/stemcomp.bamfiles.zip"
     dir.create('~/.autonomics', showWarnings = FALSE)
     destfile <- "~/.autonomics/stemcomp.bamfiles.zip"
     download.file(url, destfile = destfile, )
     utils::unzip(destfile, exdir = '~/.autonomics')
     unlink(destfile)

   # Download gtf file
     gtffile <- '~/.autonomics/human_95.gtf'
     download_gtf('Homo sapiens', 95, gtffile = gtffile)

   # Generate feature counts
     read_bam(bamdir  = "~/.autonomics/stemcomp.bamfiles",
              gtffile = gtffile,
              ispaired  = TRUE)
}
}
