% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_omics.R
\name{read_dataframe}
\alias{read_dataframe}
\title{Read dataframe block from excel or txt file}
\usage{
read_dataframe(file, sheet, range, transpose = FALSE)
}
\arguments{
\item{file}{character(1)}

\item{sheet}{numeric(1), character(1), or NULL}

\item{range}{integer(4): c(row1, rown, col1, coln)}

\item{transpose}{logical(1)}
}
\description{
Read dataframe block from excel or txt file
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   file <- system.file('extdata/glutaminase/glutaminase.xlsx', package = 'autonomics.data')
   file \%>\% read_dataframe(2, c(10, 401,  1, 14))                   \%>\% extract(1:3, 1:3)           # fdata
   file \%>\% read_dataframe(2, c(1,   10, 14, 86), transpose = TRUE) \%>\% extract(1:3, 1:3)           # sdata
}
}
