% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_design.R
\name{read_design}
\alias{read_design}
\title{Read design}
\usage{
read_design(design_file)
}
\arguments{
\item{design_file}{string: path to design file}
}
\value{
sample dataframe
}
\description{
Read design
}
\examples{
require(magrittr)

# METABOLON
if (require(autonomics.data)){
   file <- system.file('extdata/glutaminase/glutaminase.xlsx',
                package = 'autonomics.data')
   design_file <- tempfile()
   autonomics.import::write_design(
      file, 'metabolon', infer_from_sampleids = TRUE, design_file = design_file)
   design_file \%>\% read_design() \%>\% head()
}

# METABOLONLIPIDS
file <- '../../datasets/WCQA-01-18MLCLP-1/WCQA-01-18MLCLP CLP  6-TAB FILE (180710).XLSX'
if (file.exists(file)){
   design_file <- tempfile()
   autonomics.import::write_design(
      file, 'metabolonlipids', design_file = design_file) \%>\% head()
   read_design(design_file) \%>\% head()
}

# SOMA
if (require(autonomics.data)){
   file <- system.file('extdata/stemcell.comparison/stemcell.comparison.adat',
                package = 'autonomics.data')
   design_file <- tempfile()
   autonomics.import::write_design(file, 'soma', infer_from_sampleids = TRUE, design_file = design_file)
   read_design(design_file) \%>\% head()
}
}
