% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_load_maxquant_evidences.R
\name{read_evidences}
\alias{read_evidences}
\title{Read evidence file}
\usage{
read_evidences(evidence_file)
}
\arguments{
\item{evidence_file}{path to evidence file}
}
\description{
Read (relevant variables of) evidence file into a data table
}
\examples{
require(magrittr)
if (require(billing.differentiation.data)){
   system.file('extdata/maxquant/evidence.txt',
                package = 'billing.differentiation.data') \%>\%
   read_evidences()

}
}
