% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_omics.R
\name{read_exiqon}
\alias{read_exiqon}
\title{Read exiqon data}
\usage{
read_exiqon(file)
}
\arguments{
\item{file}{character(1)}
}
\description{
Read exiqon data
}
\examples{
if require(subramanian.2016){
   file <- system.file('extdata/exiqon/subramanian.2016.exiqon.xlsx', package = 'subramanian.2016')
   file \%>\% read_exiqon()
}
}
