% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_omics.R
\name{read_gtf}
\alias{read_gtf}
\title{Read GTF file}
\usage{
read_gtf(gtffile, filter = NULL, return_value = FALSE)
}
\arguments{
\item{gtffile}{string: path to gtf file (which can be downloaded with download_gtf)}

\item{filter}{Filter on feature name or feature id. By default all features are extracted.}

\item{return_value}{Display feature annotation on console. By default FALSE}
}
\description{
Read GTF file
}
\examples{
read_gtf(filter = 'ENSG00000198947', return_value = FALSE)
}
\seealso{
download_gtf
}
