% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_design.R
\name{read_maxquant_design}
\alias{read_maxquant_design}
\title{Read maxquant design file}
\usage{
read_maxquant_design(design_file)
}
\arguments{
\item{design_file}{sample design file}
}
\value{
sample design datatable
}
\description{
Read maxquant design file
}
\examples{
if (require(autonomics.data)){
   design_file <- system.file('extdata/billing2016/sample_design.txt',
                                      package = 'autonomics.data')
   autonomics.import::read_maxquant_design(design_file)
}
if (require(billing.differentiation.data)){
   design_file <- system.file('extdata/maxquant/sample_design.txt',
                                      package = 'billing.differentiation.data')
   read_maxquant_design(design_file)
}
}
