% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_omics_asis.R
\name{read_metabolon_asis}
\alias{read_metabolon_asis}
\title{Read meabolon data}
\usage{
read_metabolon_asis(file, sheet = 2, fid_var = "COMP_ID",
  sid_var = "CLIENT_IDENTIFIER")
}
\arguments{
\item{file}{character(1)}

\item{sheet}{numeric(1) or character(1)}

\item{fid_var}{character(1): feature id variable}

\item{sid_var}{character(1): sample id variable}
}
\value{
SummarizedExperiment
}
\description{
Read meabolon data
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   file <- system.file('extdata/glutaminase/glutaminase.xlsx', package = 'autonomics.data')
   file \%>\% autonomics.import::read_metabolon_asis()
}
}
