% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_design.R
\name{read_metabolon_design}
\alias{read_metabolon_design}
\title{Read metabolon design}
\usage{
read_metabolon_design(design_file)
}
\arguments{
\item{design_file}{string: path to design file}
}
\value{
sample dataframe
}
\description{
Read metabolon design
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   design_file <- tempfile()
   system.file('extdata/glutaminase/glutaminase.xlsx',
                package = 'autonomics.data') \%>\%
   autonomics.import::write_metabolon_design(design_file = design_file, infer_from_sampleids = TRUE)
   design_file \%>\% read_metabolon_design() \%>\% head()
}
file <- '../../datasets/WCQA-01-18MLCLP-1/WCQA-01-18MLCLP CLP  6-TAB FILE (180710).XLSX'
if (file.exists(file)){
   design_file <- tempfile()
   autonomics.import::write_metabolon_design(file, design_file = design_file,
                                                   infer_from_sampleids = TRUE) \%>\% head()
   design_file \%>\% read_metabolon_design() \%>\% head()
}
}
