% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_omics.R
\name{read_omics}
\alias{read_omics}
\title{Read omics data from txt or xls file}
\usage{
read_omics(file, sheet, exprs_range, fid_range = NULL, sid_range = NULL,
  fdata_range = NULL, sdata_range = NULL, transpose = FALSE)
}
\arguments{
\item{file}{character(1)}

\item{sheet}{integer(1) or character(1)}

\item{exprs_range}{numeric(4): c(row1, rown, col1, coln)}

\item{fid_range}{numeric(4): c(row1, rown, col1, coln)}

\item{sid_range}{numeric(4): c(row1, rown, col1, coln)}

\item{fdata_range}{numeric(4): c(row1, rown, col1, coln)}

\item{sdata_range}{numeric(4): c(row1, rown, col1, coln)}

\item{transpose}{logical(1)}
}
\description{
Read omics data from txt or xls file
}
\examples{
require(magrittr)
if (require(autonomics.data)){
  file <- system.file('extdata/glutaminase/glutaminase.xlsx', package = 'autonomics.data')
  file \%>\% autonomics.import::read_omics(sheet       = 2,
                                         exprs_range = c(11, 401, 15, 86),
                                         fid_range   = c(11, 401,  5,  5),
                                         sid_range   = c( 3,   3, 15, 86),
                                         fdata_range = c(10, 401,  1, 14),
                                         sdata_range = c(1,   10, 14, 86),
                                         transpose   = FALSE)
}
if (require(subramanian.2016)){
   file <- system.file('extdata/rnaseq/gene_counts.txt', package = 'subramanian.2016')
   fdata_range <- c(1, 48821, 1, 4)
   sdata_range <- c(1, 1, 5, 86)
   exprs_range <- c(2, 48821, 5, 86)
   transpose <- FALSE
   sid_var <- NULL
   fid_var <- 'gene_id'
}
}
