% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_omics.R
\name{read_phosphosites}
\alias{read_phosphosites}
\title{Read phosphosites}
\usage{
read_phosphosites(
  file,
  proteingroups_file = dirname(file) \%>\% paste0("/phospho (STY)Sites.txt"),
  quantity = guess_maxquant_quantity(file),
  fvars = phosphosite_fvars,
  standardize_snames = TRUE,
  demultiplex_snames = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{file}{string: phosphosites filepath}

\item{proteingroups_file}{string: proteingroups filepath}

\item{quantity}{NULL or value in names(maxquant_patterns)}

\item{fvars}{string vector}

\item{standardize_snames}{logical}

\item{demultiplex_snames}{logical}

\item{verbose}{logical}
}
\description{
Read phosphosites
}
\examples{
\dontrun{
if (require(autonomics.data)){
   require(magrittr)
   file <- 'extdata/stemdiff/maxquant/phospho (STY)Sites.txt' \%>\%
            system.file(package = 'autonomics.data')
   file \%>\% read_phosphosites()
}
}
}
