% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_proteingroups.R
\name{read_phosphosites_asis}
\alias{read_phosphosites_asis}
\title{Read phosphosites file}
\usage{
read_phosphosites_asis(file, proteingroups_file = file \%>\%
  stringi::stri_replace_first_fixed("phospho (STY)Sites.txt",
  "proteinGroups.txt"), quantity = NULL, fvars = c("id",
  "Protein group IDs", "Positions within proteins", "Localization prob"),
  min_localization_prob = 0.75, clean_snames = TRUE,
  uniplex_snames = TRUE)
}
\arguments{
\item{file}{character(1)}

\item{proteingroups_file}{character(1)}

\item{quantity}{NULL or value in names(autonomics.import::maxquant_patterns)}

\item{fvars}{character(n)}

\item{min_localization_prob}{numeric(1)}

\item{clean_snames}{logical(1)}

\item{uniplex_snames}{logical(1)}

\item{add_occupancies}{logical(1): whether to add occupancies(object) = phosphosite - proteingroups}

\item{verbose}{logical(1)}
}
\description{
Read phosphosites file
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   file <- system.file('extdata/stemdiff/maxquant/phospho (STY)Sites.txt', package = 'autonomics.data')
   file \%>\% autonomics.import::read_phosphosites_asis()
}
}
