% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_omics.R
\name{read_proteingroups}
\alias{read_proteingroups}
\title{Read proteingroups}
\usage{
read_proteingroups(
  file,
  quantity = guess_maxquant_quantity(file),
  fvars = proteingroups_fvars,
  standardize_snames = TRUE,
  demultiplex_snames = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{file}{character(1). Path to 'proteinGroups.txt'}

\item{quantity}{character(1). Expression columns to extract from proteinGroups file: 'Ratio normalized', 'Ratio', 'LFQ intensity',
'LFQ intensity', 'Reporter intensity corrected', 'Reporter intensity', 'Intensity labeled', or 'Intensity'.}

\item{fvars}{character(n). Annotation columns to extract from proteinGroups file.}

\item{standardize_snames}{logical(1).   Standardize maxquant snames: Ratio normalized H/L WT(L).KD(H).R1 -> WT(L).KD(H).R1{H/L} ?}

\item{demultiplex_snames}{logical(1).   Demultiplex (standardized) maxquant snames: WT(L).KD(H).R1{H/L} -> KD(H)_WT(L).R1 ?}

\item{verbose}{logical(1).   Message progress?}
}
\description{
Read proteingroups
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   file <- 'extdata/stemcomp/maxquant/proteinGroups.txt' \%>\%
            system.file(package = 'autonomics.data')
   file \%>\% read_proteingroups()
}
}
