% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_omics_asis.R
\name{read_proteingroups_asis}
\alias{read_proteingroups_asis}
\title{Read proteingroups}
\usage{
read_proteingroups_asis(file, quantity = NULL,
  fvars = autonomics.import::proteingroups_fvars, verbose = TRUE)
}
\arguments{
\item{file}{character(1):      proteingroups file}

\item{quantity}{NULL/character(1): any value in names(maxquant_patterns)}

\item{fvars}{character(n)}

\item{verbose}{logical(1)}
}
\description{
Read proteingroups
}
