% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_omics.R
\name{read_rnaseq}
\alias{read_rnaseq}
\title{Read data from rnaseq file}
\usage{
read_rnaseq(file, fid_var = "gene_id")
}
\arguments{
\item{file}{character(1)}

\item{fid_var}{character(1): feature id variable}
}
\description{
Read data from rnaseq file
}
\examples{
#file <- system.file('extdata/rnaseq/gene_counts.txt', package = 'subramanian.2016')
}
