% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_omics_asis.R
\name{read_rnaseq_asis}
\alias{read_rnaseq_asis}
\title{Read rnaseq data}
\usage{
read_rnaseq_asis(file, fid_var = "gene_id")
}
\arguments{
\item{file}{character(1)}

\item{fid_var}{character(1): feature id variable}
}
\description{
Read rnaseq data
}
\examples{
require(magrittr)
if (require(subramanian.2016)){
   file <- system.file('extdata/rnaseq/gene_counts.txt', package = 'subramanian.2016')
   file \%>\% read_rnaseq_asis()
}
}
