% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_design.R
\name{read_soma_design}
\alias{read_soma_design}
\title{Read soma design}
\usage{
read_soma_design(design_file)
}
\arguments{
\item{design_file}{string: path to design file}
}
\value{
sample dataframe
}
\description{
Read soma design
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   design_file <- tempfile()
   system.file('extdata/stemcell.comparison/stemcell.comparison.adat',
                package = 'autonomics.data') \%>\%
   autonomics.import::write_soma_design(design_file, infer_from_sampleids = TRUE)
   design_file \%>\% read_soma_design()
}
}
