% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_omics.R
\name{read_somascan}
\alias{read_somascan}
\title{Read SOMAScan data}
\usage{
read_somascan(file, fid_var = "SeqId", sid_var = "SampleId")
}
\arguments{
\item{file}{character(1)}

\item{fid_var}{character(1): feature id variable}

\item{sid_var}{character(1): sample id variable}
}
\value{
Summarizedexperiment
}
\description{
Read SOMAScan data
}
\examples{
require(magrittr)
if (require(atkin.2014)){
   file <- system.file('extdata/soma/WCQ-18-007.hybNorm.plateScale.medNorm.calibrate.20181004.adat',
                        package = 'atkin.2014')
   file \%>\% autonomics.import::read_somascan()
}
}
