% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_omics_asis.R
\name{read_somascan_asis}
\alias{read_somascan_asis}
\title{Read somascan data}
\usage{
read_somascan_asis(file, fid_var = "SeqId", sid_var = "SampleId")
}
\arguments{
\item{file}{character(1): adat file name}

\item{fid_var}{character(1): feature id variable}

\item{sid_var}{character(1): sample id variable}
}
\description{
Read somascan data
}
