% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_load_maxquant.R
\name{rename_maxquant_annotation_columns}
\alias{rename_maxquant_annotation_columns}
\title{Rename annotation columns}
\usage{
rename_maxquant_annotation_columns(DT)
}
\arguments{
\item{DT}{maxquant datatable}
}
\description{
Rename annotation columns
}
