% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutators.R
\name{rename_subgroups}
\alias{rename_subgroups}
\title{Rename subgroup levels}
\usage{
rename_subgroups(object, mapping)
}
\arguments{
\item{object}{eSet}

\item{mapping}{named character vector (names = old, values = new)}
}
\description{
Rename subgroup levels
}
