% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutators.R
\name{replace_nas_with_zeros}
\alias{replace_nas_with_zeros}
\title{Replace NAs with zeros}
\usage{
replace_nas_with_zeros(object, verbose = TRUE)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{verbose}{TRUE or FALSE}
}
\value{
eset
}
\description{
Replace NAs with zeros
}
\examples{
if (require(autonomics.data)){
   require(magrittr)
   object <- autonomics.data::stemcomp.proteinratios
   replace_nas_with_zeros(object)
}
}
