% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rnaseq.R
\docType{data}
\name{REQUIRED_FVARS_RNASEQ}
\alias{REQUIRED_FVARS_RNASEQ}
\title{Required feature variables}
\format{An object of class \code{character} of length 2.}
\usage{
REQUIRED_FVARS_RNASEQ
}
\description{
Required variables in annotated count file
}
\keyword{datasets}
