% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_load_rnaseq.R
\docType{data}
\name{REQUIRED_SVARS_RNASEQ}
\alias{REQUIRED_SVARS_RNASEQ}
\title{Required variables in annotated count file}
\format{An object of class \code{character} of length 3.}
\usage{
REQUIRED_SVARS_RNASEQ
}
\description{
Required variables in annotated count file
}
\keyword{datasets}
