% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xxx_values.R
\name{reshape_values}
\alias{reshape_values}
\title{Reshape composite values}
\usage{
reshape_values(x, fill, fun.aggregate, sep = guess_sep(x))
}
\arguments{
\item{x}{character(n) with composite values}

\item{fill}{missing value placeholder (passed to data.table::dcast)}

\item{fun.aggregate}{'unique' (for layout purposes) or length (for tabulate purposes)}

\item{sep}{component separator}
}
\description{
Reshapes a vector with composite values into a (2d) matrix:
   - rownames = values of component 1:(n-1)
   - colnames = values of component n
   - values can be either composite levels themselves   (fill = '', fun.aggregate = unique)
                or frequency counts of composite levels (fill = 0,  fun.aggregate = length)
}
\examples{
if (require(autonomics.data)){
   # GLUTAMINASE
      require(magrittr)
      x <- autonomics.data::glutaminase \%>\% subgroup_values()
      x \%>\% reshape_values(fill = '', fun.aggregate = unique)
      x \%>\% reshape_values(fill = 0,  fun.aggregate = length)
}

}
