% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_load_maxquant.R
\name{rm_missing_id_values}
\alias{rm_missing_id_values}
\title{Rm missing id values}
\usage{
rm_missing_id_values(PG)
}
\arguments{
\item{PG}{maxquant dataframe}
}
\description{
Rm missing id values
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   PG <- system.file('extdata/stemdiff/maxquant/proteinGroups.txt',
                      package = 'autonomics.data') \%>\%
         data.table::fread()
   PG \%>\% dim()
   PG \%>\% autonomics.import::rm_missing_id_values() \%>\% dim()
}
}
