% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_omics.R
\name{sampleid_varname}
\alias{sampleid_varname}
\title{Get sampleid svar name}
\usage{
sampleid_varname(platform)
}
\arguments{
\item{platform}{'metabolonlipids', 'metabolon', 'soma'}
}
\value{
string
}
\description{
Get sampleid svar name
}
\examples{
sampleid_varname('metabolonlipids')
sampleid_varname('metabolon')
sampleid_varname('soma')
}
