% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scomponents.R
\name{scomponents}
\alias{scomponents}
\alias{subgroup_components}
\title{Get subgroup components}
\usage{
scomponents(object, svar)

subgroup_components(object)
}
\arguments{
\item{object}{SummarizedExperiment, eSet, or EList}

\item{svar}{sample var}
}
\value{
data.table with decomposed subgroup components
}
\description{
Get data.table in which subgroup components have been decomposed
}
\examples{
require(magrittr)
if (require(billing.differentiation.data)){
   billing.differentiation.data::protein.ratios \%>\% scomponents('subgroup')
   billing.differentiation.data::protein.ratios \%>\% subgroup_components()
}
if (require(halama.2016)){
   halama.2016::cell.metabolites  \%>\% scomponents('subgroup')
   halama.2016::cell.metabolites  \%>\% subgroup_components()
}
if (require(subramanian.2016)){
   subramanian.2016::metabolon    \%>\% scomponents('subgroup')
   subramanian.2016::metabolon    \%>\% subgroup_components()
}
if (require(graumann.lfq)){
   graumann.lfq::lfq.intensities  \%>\% scomponents('subgroup')
   graumann.lfq::lfq.intensities  \%>\% subgroup_components()
}
}
