% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters.R
\docType{methods}
\name{sdata}
\alias{sdata}
\alias{sdata,eSet-method}
\alias{sdata,EList-method}
\alias{sdata,SummarizedExperiment-method}
\alias{sdata<-}
\alias{sdata<-,SummarizedExperiment,data.frame-method}
\alias{sdata<-,eSet,data.frame-method}
\alias{sdata<-,EList,data.frame-method}
\title{Get/Set sdata}
\usage{
sdata(object)

\S4method{sdata}{eSet}(object)

\S4method{sdata}{EList}(object)

\S4method{sdata}{SummarizedExperiment}(object)

sdata(object) <- value

\S4method{sdata}{SummarizedExperiment,data.frame}(object) <- value

\S4method{sdata}{eSet,data.frame}(object) <- value

\S4method{sdata}{EList,data.frame}(object) <- value
}
\arguments{
\item{object}{SummarizedExperiment, eSet, or EList}

\item{value}{dataframe}
}
\value{
sample dataframe (get) or updated object (set)
}
\description{
Get/Set sample data
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   object <- autonomics.data::stemcomp.proteinratios
   sdata(object) \%<>\% cbind(z=1)
   object
}
}
