% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_omics.R
\name{select_gtf_features}
\alias{select_gtf_features}
\title{Select GTF feature}
\usage{
select_gtf_features(gtffile, filter = NULL, return_feature = FALSE)
}
\arguments{
\item{gtffile}{string: path to gtf file (which can be downloaded with download_gtf)}

\item{filter}{Filter on feature name or feature id. By default all features are extracted.}

\item{return_feature}{Display feature annotation on console. By default FALSE}
}
\description{
Select GTF feature
}
\examples{
\dontrun{ # requires internet connection
   gtffile <- download_gtf(organism = 'Homo sapiens', release = 95)
   select_gtf_features(gtffile = gtffile, filter = 'ENSG00000198947', return_feature = FALSE)
}
}
\seealso{
download_gtf
}
