% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eset_specific_var.R
\name{sep}
\alias{sep}
\alias{get_sep}
\title{Get separator of collapsed fields}
\usage{
sep(object)

get_sep(object)
}
\arguments{
\item{object}{eset}
}
\value{
string with separator
}
\description{
Get separator of collapsed fields
}
\examples{
if (require(billing.differentiation.data)){
   require(magrittr)
   object <- rna.voomcounts
   object \%>\% autonomics.import::sep()
}
}
