% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_load_maxquant.R
\name{set_summary_attr}
\alias{set_summary_attr}
\title{Set autonomics summary attributes}
\usage{
set_summary_attr(x, y)
}
\arguments{
\item{x}{object}

\item{y}{autonomics summary attribute}
}
\value{
updated object
}
\description{
Set autonomics summary attributes
}
