% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters.R
\docType{methods}
\name{sign,SummarizedExperiment-method}
\alias{sign,SummarizedExperiment-method}
\title{Get (exprs) sign matrix}
\usage{
\S4method{sign}{SummarizedExperiment}(x)
}
\arguments{
\item{x}{SummarizedExperiment}
}
\description{
Get (exprs) sign matrix
}
