% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters.R
\name{slevels}
\alias{slevels}
\alias{subgroup_levels}
\title{Get slevels}
\usage{
slevels(object, svar)

subgroup_levels(object)
}
\arguments{
\item{object}{SummarizedExperiment, eSet, or eList}

\item{svar}{sample var (character)}
}
\value{
svar values (character)
}
\description{
Get svar levels
}
\examples{
require(magrittr)
if (require(subramanian.2016)){
   subramanian.2016::metabolon \%>\% slevels('subgroup')
   subramanian.2016::metabolon \%>\% subgroup_levels()
}
}
