% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters.R
\docType{methods}
\name{snames}
\alias{snames}
\alias{snames,SummarizedExperiment-method}
\alias{snames,eSet-method}
\alias{snames,EList-method}
\alias{snames<-}
\alias{snames<-,SummarizedExperiment,character-method}
\alias{snames<-,eSet,character-method}
\alias{snames<-,EList,character-method}
\title{Get/Set snames}
\usage{
snames(object)

\S4method{snames}{SummarizedExperiment}(object)

\S4method{snames}{eSet}(object)

\S4method{snames}{EList}(object)

snames(object) <- value

\S4method{snames}{SummarizedExperiment,character}(object) <- value

\S4method{snames}{eSet,character}(object) <- value

\S4method{snames}{EList,character}(object) <- value
}
\arguments{
\item{object}{SummarizedExperiment, eSet, or EList}

\item{value}{string vector with sample names}
}
\value{
sample names vector (get) or updated eSet (set)
}
\description{
Get/Set sample names
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   object <- autonomics.data::stemcomp.proteinratios
   snames(object) \%<>\% paste0('SAMPLE', .)
   object
}
}
