% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_soma.R
\name{soma_to_sumexp}
\alias{soma_to_sumexp}
\title{Convert WideSomaLogic object to SummarizedExperiment}
\usage{
soma_to_sumexp(soma, log2_transform = TRUE)
}
\arguments{
\item{soma}{WideSomaLogic object}

\item{log2_transform}{logical}
}
\description{
Convert WideSomaLogic object to SummarizedExperiment
}
