% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{split_by_svar}
\alias{split_by_svar}
\title{Split SummarizedExperiment by svar}
\usage{
split_by_svar(object, svar = "subgroup")
}
\arguments{
\item{object}{SummarizedExperiment}

\item{svar}{'subgroup'}
}
\value{
list of SummarizedExperiments
}
\description{
Split SummarizedExperiment by svar
}
\examples{
if (require(autonomics.data)){
   require(magrittr)
   object <- autonomics.data::stemcomp.proteinratios
   object \%>\% split_by_svar()
}
}
