% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scomponents.R
\name{split_components}
\alias{split_components}
\title{Split (subgroup) components}
\usage{
split_components(values, sep = autonomics.import::infer_design_sep(values))
}
\arguments{
\item{values}{character vector with subgroup values}
}
\value{
data.table with decomposed subgroup components
}
\description{
Get data.table in which subgroup components have been decomposed
}
\examples{
require(magrittr)
if (require(billing.differentiation.data)){
   object <- billing.differentiation.data::protein.ratios
   object \%>\% subgroup_values() \%>\% split_components()
}
if (require(autonomics.data)){
   autonomics.data::glutaminase \%>\% subgroup_values()  \%>\% split_components()
}
if (require(subramanian.2016)){
   subramanian.2016::metabolon \%>\% subgroup_values()   \%>\% split_components()
}
if (require(graumann.lfq)){
   graumann.lfq::lfq.intensities \%>\% subgroup_values() \%>\% split_components()
}
}
