% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutators.R
\name{split_on_svar}
\alias{split_on_svar}
\title{Split on svar}
\usage{
split_on_svar(object, svar = NULL)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{svar}{sample var}
}
\value{
list of SummarizedExperiments
}
\description{
Split on svar
}
\examples{
require(magrittr)
if (require(subramanian.2016)){
   object <- subramanian.2016::metabolon
   svar = 'condition'
   object \%>\% autonomics.import::split_on_svar(svar)
}
}
