% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xxx_composite_values.R
\name{split_values}
\alias{split_values}
\alias{split_values.character}
\alias{split_values.factor}
\alias{split_values.SummarizedExperiment}
\alias{scomponents}
\alias{subgroup_components}
\title{Split composite (s)values}
\usage{
split_values(x, ...)

\method{split_values}{character}(x, sep = autonomics.import::guess_sep(x,
  verbose = FALSE), keep = FALSE)

\method{split_values}{factor}(x, sep = autonomics.import::guess_sep(x, verbose
  = FALSE), keep = FALSE)

\method{split_values}{SummarizedExperiment}(object, svar = "subgroup",
  keep = FALSE)

scomponents(...)

subgroup_components(...)
}
\arguments{
\item{x}{character(n), factor(n), or SummarizedExperiment}

\item{sep}{character(1):  separator}

\item{keep}{logical(1_: whether to keep original values (in first column)}

\item{svar}{character(1)}
}
\value{
data.table(x, x1, x2, ...) or data.table(x1, x2, ...)
}
\description{
Get data.table in which composite (s)values have been split into components
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   # STEM CELL DIFFERENTIATION
      x <- autonomics.data::stemdiff.proteinratios
      x \%>\% autonomics.import::subgroup_values() \%>\% autonomics.import::split_values()
      x \%>\%                                          autonomics.import::split_values()
      x \%>\%                                          autonomics.import::split_values(keep = TRUE)
   # GLUTAMINASE
      x <- autonomics.data::glutaminase
      x \%>\% autonomics.import::subgroup_values() \%>\% autonomics.import::split_values()
      x \%>\%                                          autonomics.import::split_values()
}
if (require(subramanian.2016)){
   x <- subramanian.2016::metabolon
   x \%>\% autonomics.import::subgroup_values()    \%>\% autonomics.import::split_values()
   x                                             \%>\% autonomics.import::split_values()
}
if (require(graumann.lfq)){
   x <- graumann.lfq::lfq.intensities
   x \%>\% autonomics.import::subgroup_values()    \%>\% autonomics.import::split_values()
   x                                             \%>\% autonomics.import::split_values()
}
if (require(atkin.2014)){
   x <- atkin.2014::soma.2018
   x \%>\% autonomics.import::subgroup_values()    \%>\% autonomics.import::split_values()
   x                                             \%>\% autonomics.import::split_values()
}
}
