% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssep.R
\name{ssep}
\alias{ssep}
\alias{subgroup_sep}
\title{Get svar separator}
\usage{
ssep(object, svar = "subgroup")

subgroup_sep(object)
}
\arguments{
\item{object}{SummarizedExperiment, eSet, or EList}

\item{svar}{sample variable}
}
\value{
string or NULL (if no separator found)
}
\description{
Get svar separator
}
\examples{
require(magrittr)
if (require(halama.2016)){
   halama.2016::cell.metabolites \%>\% autonomics.import::ssep()
}
if (require(subramanian.2016)){
   subramanian.2016::metabolon   \%>\% autonomics.import::ssep()
}
if (require(graumann.lfq)){
   graumann.lfq::lfq.intensities \%>\% autonomics.import::ssep()
}
}
