% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_omics.R
\name{standardize_maxquant_snames}
\alias{standardize_maxquant_snames}
\alias{standardize_maxquant_snames.character}
\alias{standardize_maxquant_snames.SummarizedExperiment}
\title{Standardize maxquant snames}
\usage{
standardize_maxquant_snames(x, ...)

\method{standardize_maxquant_snames}{character}(
  x,
  quantity = guess_maxquant_quantity(x),
  verbose = FALSE,
  ...
)

\method{standardize_maxquant_snames}{SummarizedExperiment}(
  x,
  quantity = guess_maxquant_quantity(x),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{character(.) or SummarizedExperiment}

\item{...}{allow for proper S3 method dispatch}

\item{quantity}{maxquant quantity}

\item{verbose}{logical(1)}
}
\description{
For charactervector or SummarizedExperiment
}
\details{
Drop "Ratio normalized", "LFQ intensity" etc from maxquant snames & sample_id values
}
\examples{
require(magrittr)

# character vector
    standardize_maxquant_snames("Ratio M/L normalized STD(L)_EM00(M)_EM01(H)_R1")
    standardize_maxquant_snames("Ratio M/L STD(L)_EM00(M)_EM01(H)_R1")
    standardize_maxquant_snames('LFQ intensity STD_R1')
    standardize_maxquant_snames('LFQ intensity L STD(L)_EM00(M)_EM01(H)_R1')
    standardize_maxquant_snames('Reporter intensity 0 A(0)_B(1)_C(2)_D(3)_E(4)_F(5)_R1')
    standardize_maxquant_snames('Reporter intensity corrected 0 A(0)_B(1)_C(2)_D(3)_E(4)_F(5)_R1')

# SummarizedExperiment
if (require(autonomics.data)){
     x <- 'extdata/stemdiff/maxquant/proteinGroups.txt' \%>\%
           system.file(package = 'autonomics.data')     \%>\%
           read_proteingroups(standardize_snames = FALSE,
                              demultiplex_snames = FALSE)
     x \%>\% standardize_maxquant_snames(verbose = TRUE)
}
}
