% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scomponents.R
\name{subgroup_components}
\alias{subgroup_components}
\alias{scomponents}
\title{Split (subgroup) components}
\usage{
subgroup_components(object)

scomponents(object, svar)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{svar}{character}
}
\value{
data.table with decomposed subgroup components
}
\description{
Get data.table in which subgroup components have been decomposed
}
\examples{
require(magrittr)
if (require(billing.differentiation.data)){
   object <- billing.differentiation.data::protein.ratios
   object \%>\% subgroup_components()
   object \%>\% scomponents('subgroup')
}
if (require(autonomics.data)){
   object <- autonomics.data::glutaminase
   object \%>\% subgroup_components()
   object \%>\% scomponents('subgroup')
}
if (require(subramanian.2016)){
   object <- subramanian.2016::metabolon
   object \%>\% subgroup_components()
   object \%>\% scomponents('subgroup')
}
if (require(graumann.lfq)){
   object <- graumann.lfq::lfq.intensities
   object \%>\% subgroup_components()
   object \%>\% scomponents('subgroup')
}
}
