% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_has_is.R
\name{subgroup_defs_complete}
\alias{subgroup_defs_complete}
\title{Are subgroup definitions complete?}
\usage{
subgroup_defs_complete(object)
}
\arguments{
\item{object}{SummarizedExperiment}
}
\value{
logical
}
\description{
Are subgroup definitions complete?
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   object <- autonomics.data::glutaminase
   object \%>\% subgroup_defs_complete()
}
}
