% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_design.R
\name{subgroup_varname}
\alias{subgroup_varname}
\title{Get subgroup svar name}
\usage{
subgroup_varname(platform)
}
\arguments{
\item{platform}{'metabolonlipids', 'metabolon', 'soma'}
}
\value{
string
}
\description{
Get subgroup svar name
}
\examples{
subgroup_varname('metabolonlipids')
subgroup_varname('metabolon')
subgroup_varname('soma')
}
