% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumexp_to_table.R
\name{sumexp_to_wide_dt}
\alias{sumexp_to_wide_dt}
\alias{eset_to_wide_table}
\title{Convert sumexp into wide datatable}
\usage{
sumexp_to_wide_dt(object, fid = autonomics.import::fid_var(object),
  fvars = character(0))

eset_to_wide_table(...)
}
\arguments{
\item{object}{sumexp}

\item{fid}{fvar carrying feature id}

\item{fvars}{additional fvars to include in table}

\item{...}{(backward compatibility)}
}
\value{
wide data.table
}
\description{
Convert sumexp into wide datatable
}
