% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_prepro_list.R
\docType{data}
\name{SUPPORTED_ASSAYS}
\alias{SUPPORTED_ASSAYS}
\title{Supported assays}
\format{
An object of class \code{character} of length 5.
}
\usage{
SUPPORTED_ASSAYS
}
\description{
Supported assays
}
\keyword{datasets}
