% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_prepro_list.R
\docType{data}
\name{SUPPORTED_ENTITIES}
\alias{SUPPORTED_ENTITIES}
\title{Supported entities}
\format{An object of class \code{character} of length 5.}
\usage{
SUPPORTED_ENTITIES
}
\description{
Supported entities
}
\keyword{datasets}
