% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_prepro_list.R
\docType{data}
\name{SUPPORTED_QUANTITIES}
\alias{SUPPORTED_QUANTITIES}
\title{Supported quantities}
\format{An object of class \code{character} of length 9.}
\usage{
SUPPORTED_QUANTITIES
}
\description{
Supported quantities
}
\keyword{datasets}
