% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_prepro_list.R
\docType{data}
\name{SUPPORTED_SOFTWARES}
\alias{SUPPORTED_SOFTWARES}
\title{Supported softwares}
\format{An object of class \code{character} of length 5.}
\usage{
SUPPORTED_SOFTWARES
}
\description{
Supported softwares
}
\keyword{datasets}
