% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters.R
\docType{methods}
\name{svars}
\alias{svars}
\alias{svars,SummarizedExperiment-method}
\alias{svars,eSet-method}
\alias{svars,EList-method}
\alias{svars<-}
\alias{svars<-,SummarizedExperiment,character-method}
\alias{svars<-,eSet,character-method}
\alias{svars<-,EList,character-method}
\title{Get/Set svars}
\usage{
svars(object)

\S4method{svars}{SummarizedExperiment}(object)

\S4method{svars}{eSet}(object)

\S4method{svars}{EList}(object)

svars(object) <- value

\S4method{svars}{SummarizedExperiment,character}(object) <- value

\S4method{svars}{eSet,character}(object) <- value

\S4method{svars}{EList,character}(object) <- value
}
\arguments{
\item{object}{SummarizedExperiment}

\item{value}{string fector with variable names}
}
\value{
sample variable names (get) or updated eSet (set)
}
\description{
Get/Set sample variables
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   object <- autonomics.data::stemcomp.proteinratios
   svars(object)[1] \%<>\% paste0('1')
   object
}
}
