% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters.R
\name{tmpplot}
\alias{tmpplot}
\alias{tmpplot,SummarizedExperiment-method}
\alias{tmpplot<-}
\alias{tmpplot<-,SummarizedExperiment,list-method}
\title{Get/set tmpplot}
\usage{
tmpplot(object)

\S4method{tmpplot}{SummarizedExperiment}(object)

tmpplot(object) <- value

\S4method{tmpplot}{SummarizedExperiment,list}(object) <- value
}
\arguments{
\item{object}{SummarizedExperiment}

\item{value}{list}
}
\value{
(temporary) plot (get) or updated object (set)
}
\description{
Get/set tmpplot
}
