% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutators.R
\name{uncollapse_n_pick_first}
\alias{uncollapse_n_pick_first}
\title{Extract first from collapsed fvalues}
\usage{
uncollapse_n_pick_first(object, fvar, sep)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{fvar}{fvar}

\item{sep}{separator}
}
\value{
updated object
}
\description{
Extract first from collapsed fvalues
}
\examples{
require(magrittr)
if (require(subramanian.2016)){
   subramanian.2016::metabolon \%>\%
      autonomics.import::uncollapse_n_pick_first(
         fvar = 'PUBCHEM',
         sep = ';')
}
}
