% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eset_specific_var.R
\name{uniprot_values}
\alias{uniprot_values}
\title{Get uniprot values}
\usage{
uniprot_values(object, first_only = FALSE)
}
\arguments{
\item{object}{eset}

\item{first_only}{logical(1)}
}
\value{
character vector
}
\description{
Get uniprot values
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   object <- autonomics.data::stemcomp.proteinratios
   object \%>\% uniprot_values() \%>\% head()
   object \%>\% uniprot_values(first_only = TRUE) \%>\% head()
}
}
