% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eset_specific_var.R
\name{uniprot_var}
\alias{uniprot_var}
\alias{get_uniprot_var}
\title{Get uniprot var}
\usage{
uniprot_var(object)

get_uniprot_var(object)
}
\arguments{
\item{object}{eset}
}
\value{
fvar with uniprot accessions
}
\description{
Get uniprot var
}
