% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load2_sdata.R
\name{validify_sample_ids}
\alias{validify_sample_ids}
\title{Validify sample/feature names}
\usage{
validify_sample_ids(x)
}
\arguments{
\item{x}{character vector with sample ids}
}
\value{
character vector
}
\description{
Validify sample/feature names
}
