% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rnaseq.R
\name{voom_filter_n_transform}
\alias{voom_filter_n_transform}
\title{Voom filter features}
\usage{
voom_filter_n_transform(object, normalize.method = "none", file = NULL,
  width = 10, height = 5)
}
\arguments{
\item{object}{exprs object}

\item{normalize.method}{passed to limma::voom}

\item{file}{file to print mean var curve to}

\item{width}{fig width in inches}

\item{height}{fig height in inches}
}
\value{
filtered eset
}
\description{
Filter low count features (for all samples) to achieve monotonic mean variance trend,
as is recommended by Gordon Smyth & co. (see references)
}
\references{
https://support.bioconductor.org/p/64484/
https://support.bioconductor.org/p/80331/
}
