% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_load_rnaseq.R
\name{voom_plot}
\alias{voom_plot}
\title{Plot voom mean variance trend}
\usage{
voom_plot(object, normalize.method, title)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{normalize.method}{passed to limma::voom}

\item{title}{plot title}
}
\value{
plot object
}
\description{
Plot voom mean variance trend
}
