% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_load_rnaseq.R
\name{voom_transform}
\alias{voom_transform}
\title{Voom transform eset}
\usage{
voom_transform(object, normalize.method = "none", plot = TRUE,
  verbose = TRUE)
}
\arguments{
\item{object}{eset}

\item{normalize.method}{normalization method}

\item{plot}{whether to plot mean variance trend}

\item{verbose}{whether to report progress}
}
\description{
Voom transform eset
}
\examples{
require(magrittr)
if (require(billing.differentiation.data)){

   # eset with counts
      object <- billing.differentiation.data::rna.counts
      object \%>\% autonomics.import::exprs() \%>\% extract(1:3, 1:3)

   # voom transformation
         object \%>\% autonomics.import::voom_transform() \%>\% autonomics.import::exprs() \%>\%
                    magrittr::extract(1:3, 1:3)
      # containing logcpm exprs
         object \%>\% autonomics.import::logcpm() \%>\% autonomics.import::exprs() \%>\%
                    extract(1:3, 1:3)
         object \%>\% autonomics.import::exprs() \%>\% magrittr::extract(1:3, 1:3)
      # and precision weights
         object \%>\% voom_transform() \%>\% autonomics.import::weights() \%>\%
                    extract(1:3, 1:3)
}
if (require(subramanian.2016)){
   dir <- system.file('extdata/rnaseq', package = 'subramanian.2016')
   object <- autonomics.import::load_rnaseq_counts(dir, sample_file = NULL)
   object \%>\% autonomics.import::voom_transform()
}
}
