% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters.R
\name{weights}
\alias{weights}
\alias{weights,SummarizedExperiment-method}
\alias{weights,ExpressionSet-method}
\alias{weights,EList-method}
\alias{weights<-}
\alias{weights<-,SummarizedExperiment,matrix-method}
\alias{weights<-,SummarizedExperiment,numeric-method}
\alias{weights<-,ExpressionSet,matrix-method}
\alias{weights<-,EList,matrix-method}
\alias{weights<-,EList,numeric-method}
\title{Get/Set weights}
\usage{
weights(object, ...)

\S4method{weights}{SummarizedExperiment}(object)

\S4method{weights}{ExpressionSet}(object)

\S4method{weights}{EList}(object)

weights(object) <- value

\S4method{weights}{SummarizedExperiment,matrix}(object) <- value

\S4method{weights}{SummarizedExperiment,numeric}(object) <- value

\S4method{weights}{ExpressionSet,matrix}(object) <- value

\S4method{weights}{EList,matrix}(object) <- value

\S4method{weights}{EList,numeric}(object) <- value
}
\arguments{
\item{object}{SummarizedExperiment}

\item{...}{addtional params}

\item{value}{ratio matrix (features x samples)}
}
\value{
weight matrix (get) or updated object (set)
}
\description{
Get/Set weight matrix
}
\examples{
if (require(autonomics.data)){
   require(magrittr)
   object <- autonomics.data::stemcomp.proteinratios
   weights(object)
   weights(object) <- 1; weights(object) \%>\% str()
}
}
