% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_design.R
\name{write_design}
\alias{write_design}
\title{Write design for metabolon, metabolonlipids, or soma data}
\usage{
write_design(file, platform, infer_from_sampleids = FALSE,
  design_file = NULL, sheet = 2)
}
\arguments{
\item{file}{string: path to metabolon file}

\item{infer_from_sampleids}{logical: whether to infer design from CLIENT_IDENTIFIER}

\item{design_file}{string: path to design file}

\item{sheet}{character(1) or numeric(1): xls sheet}
}
\value{
design dataframe
}
\description{
Write design for metabolon, metabolonlipids, or soma data
}
\examples{
require(magrittr)

# METABOLON
if (require(autonomics.data)){
   file <- system.file('extdata/glutaminase/glutaminase.xlsx',
                                  package = 'autonomics.data')
   file \%>\% write_design('metabolon') \%>\% head()
   file \%>\% write_design('metabolon') \%>\% head()
   file \%>\% write_design('metabolon', infer_from_sampleids = TRUE) \%>\% head()
}

# METABOLONLIPIDS
file <- '../../datasets/WCQA-01-18MLCLP-1/WCQA-01-18MLCLP CLP  6-TAB FILE (180710).XLSX'
if (file.exists(file)){
   file \%>\% write_design('metabolonlipids') \%>\% head(3)
   file \%>\% write_design('metabolonlipids', infer_from_sampleids = TRUE) \%>\% head(3)
}

# SOMA
if (require(autonomics.data)){
   soma_file <- system.file('extdata/stemcell.comparison/stemcell.comparison.adat',
                             package = 'autonomics.data')
   soma_file \%>\% autonomics.import::write_design('soma')
   soma_file \%>\% autonomics.import::write_design('soma', infer_from_sampleids = TRUE)
}
}
