% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_design.R
\name{write_design_file}
\alias{write_design_file}
\title{Write sample df to file}
\usage{
write_design_file(design_df, design_file)
}
\arguments{
\item{design_df}{sample dataframe}

\item{design_file}{sample file}
}
\value{
path to sample file
}
\description{
Write sample df to file
}
