% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_design.R
\name{write_exiqon_design}
\alias{write_exiqon_design}
\title{Write exiqon design}
\usage{
write_exiqon_design(file, infer_from_sampleids = FALSE, design_file = NULL)
}
\arguments{
\item{file}{string: path to exiqon file}

\item{infer_from_sampleids}{logical: whether to infer design from sampleids}

\item{design_file}{string: path to sample file}
}
\value{
string: path to design file
}
\description{
Write exiqon design
}
\examples{
require(magrittr)
if (require(subramanian.2016)){
   file <- system.file('extdata/exiqon/subramanian.2016.exiqon.xlsx',
                               package = 'subramanian.2016')
   file \%>\% autonomics.import::write_exiqon_design()
   file \%>\% autonomics.import::write_exiqon_design(infer_from_sampleids = TRUE)
}
}
