% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_fdata_to_file}
\alias{write_fdata_to_file}
\title{Write fdata to file}
\usage{
write_fdata_to_file(object, file = "")
}
\arguments{
\item{object}{eset}

\item{file}{file}
}
\description{
Write fdata to file
}
\examples{
if (require(autonomics.data)){
   object <- ALL
   file <- tempfile()
   write_fdata_to_file(object, file)
   file.exists(file)
   unlink(file)
}
}
