% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_fvar_to_file}
\alias{write_fvar_to_file}
\title{Write fvar to file}
\usage{
write_fvar_to_file(object, fvar, file = "")
}
\arguments{
\item{object}{eset}

\item{fvar}{fvar}

\item{file}{file}
}
\description{
Write fvar to file
}
\examples{
if (require(autonomics.data)){
   object <- ALL
   fvar <- 'gene_symbols'
   file <- tempfile()
   write_fvar_to_file(object, fvar, file)
   file.exists(file)
   unlink(file)
}
}
