% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sample_file.R
\name{write_sample_file}
\alias{write_sample_file}
\title{Write sample df to file}
\usage{
write_sample_file(sample_df, sample_file)
}
\arguments{
\item{sample_df}{sample dataframe}

\item{sample_file}{sample file}
}
\value{
path to sample file
}
\description{
Write sample df to file
}
