% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_design.R
\name{write_soma_design}
\alias{write_soma_design}
\title{Write soma design}
\usage{
write_soma_design(soma_file, infer_from_sampleids = FALSE,
  design_file = NULL)
}
\arguments{
\item{soma_file}{string: path to soma file}

\item{infer_from_sampleids}{logical: whether to infer design from sampleids}

\item{design_file}{string: path of sample file to which design is written}
}
\value{
sample design dataframe
}
\description{
Write soma design
}
\examples{
if (require(autonomics.data)){
   soma_file <- system.file('extdata/stemcell.comparison/stemcell.comparison.adat',
                             package = 'autonomics.data')
   soma_file \%>\% autonomics.import::write_soma_design()
   soma_file \%>\% autonomics.import::write_soma_design(infer_from_sampleids = TRUE)
}
}
