% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_to_na.R
\name{zero_to_na}
\alias{zero_to_na}
\alias{nan_to_na}
\alias{na_to_zero}
\alias{minusinf_to_na}
\title{Switch between nondetect representations}
\usage{
zero_to_na(object, verbose = FALSE)

nan_to_na(object, verbose = FALSE)

na_to_zero(object, verbose = FALSE)

minusinf_to_na(object, verbose = FALSE)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{verbose}{logical(1)}
}
\value{
Updated SummarizedExperiment
}
\description{
Switch between nondetect representations
}
\examples{
require(magrittr)

# 0 -> NA (proteinGroups LFQ intensities)
#----------------------------------------

# NaN -> NA (proteinGroups ratios)
#---------------------------------
   if (require(autonomics.data)){
      object <- 'extdata/stemcomp/maxquant/proteinGroups.txt' \%>\%
                 system.file(package='autonomics.data')       \%>\%
                 read_proteingroups()
      object \%>\% nan_to_na(verbose = TRUE)
   }

# -Inf -> NA (log2 transformed proteinGroups LFQ intensity)
#----------------------------------------------------------

# NA -> 0
#--------
   if (require(autonomics.data)){
      object <- 'extdata/glutaminase/glutaminase.xlsx' \%>\%
                 system.file(package = 'autonomics.data') \%>\%
                 read_metabolon()
      object \%>\% na_to_zero(verbose = TRUE)
   }


}
