% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{add_kegg_pathways_to_fdata}
\alias{add_kegg_pathways_to_fdata}
\title{Map a vector of KEGG IDs to KEGG Pathways}
\usage{
add_kegg_pathways_to_fdata(object)
}
\arguments{
\item{object}{SummarizedExperiment}
}
\value{
vector with pathway IDs (collapsed)
}
\description{
Map a vector of KEGG IDs to KEGG Pathways
}
\examples{
require(magrittr)
\dontrun{
   if (require(subramanian.2016)){
      subramanian.2016::metabolon \%>\%
      autonomics.import::add_kegg_pathways_to_fdata()
   }
   if (require(autonomics.data)){
      file <- system.file('extdata/glutaminase/glutaminase.xlsx',
                           package = 'autonomics.data')
      object <- autonomics.import::load_metabolon(file)
      autonomics.import::add_kegg_pathways_to_fdata(object)
   }
}
}
