% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_maxquant.R
\name{annotate_proteingroups}
\alias{annotate_proteingroups}
\title{Annotate proteingroups through uniprot.ws}
\usage{
annotate_proteingroups(object, connection = object \%>\%
  autonomics.import::open_uniprot_connection(),
  columns = c("SUBCELLULAR-LOCATIONS", "INTERPRO", "GO-ID"))
}
\arguments{
\item{object}{SummarizedExperiment}

\item{connection}{uniprot webservice connection}

\item{columns}{uniprot webservice columns}
}
\value{
Annotated SummarizedExperiment
}
\description{
Annotate proteingroups through uniprot.ws
}
\examples{
require(magrittr)
\dontrun{
   if (require(autonomics.data)){
      object <- autonomics.data::stemcomp.proteinratios
      connection <- autonomics.import::open_uniprot_connection(object)
      object[1:10, ] \%>\% annotate_proteingroups(connection, c('SUBCELLULAR-LOCATIONS', 'GO-ID'))
   }
}
}
