% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_has_is.R
\name{contains_ratios}
\alias{contains_ratios}
\title{Does object contain ratio values?}
\usage{
contains_ratios(object)
}
\arguments{
\item{object}{SummarizedExperiment}
}
\value{
logical
}
\description{
Does object contain ratio values?
}
\examples{

if (require(autonomics.data)){

   # STEM CELL COMPARISON
   require(magrittr)
   object <- autonomics.data::stemcomp.proteinratios
   object \%>\% contains_ratios()

   # GLUTAMINASE
   object <- autonomics.data::glutaminase
   object \%>\% contains_ratios()
}

}
