% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters.R
\name{contrastdefs}
\alias{contrastdefs}
\alias{contrastdefs,SummarizedExperiment-method}
\alias{contrastdefs<-}
\alias{contrastdefs<-,SummarizedExperiment,character-method}
\alias{contrastdefs<-,SummarizedExperiment,NULL-method}
\alias{set_contrastdefs}
\title{Get/Set contrast definitions}
\usage{
contrastdefs(object)

\S4method{contrastdefs}{SummarizedExperiment}(object)

contrastdefs(object) <- value

\S4method{contrastdefs}{SummarizedExperiment,character}(object) <- value

\S4method{contrastdefs}{SummarizedExperiment,`NULL`}(object) <- value

set_contrastdefs(object, value)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{value}{named string vector (see examples)}
}
\value{
updated SummarizedExperiment
}
\description{
Get/Set contrast definitions
}
\examples{
if (require(autonomics.data)){
   require(magrittr)
   object <- autonomics.data::stemcomp.proteinratios
   contrastdefs1 <- c(EM_E = 'EM_E', BM_E = 'BM_E', BM_EM = 'BM_EM')
   contrastdefs(object) <-     contrastdefs1   # conventional setter
   object \%>\% set_contrastdefs(contrastdefs1)  # piping       setter
   contrastdefs(object)                        # getter
}
}
