% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xxx_values.R
\name{count_values}
\alias{count_values}
\alias{count_values.character}
\alias{count_values.factor}
\alias{count_values.SummarizedExperiment}
\title{Count composite (s)values}
\usage{
count_values(x, ...)

\method{count_values}{character}(x, ...)

\method{count_values}{factor}(x, ...)

\method{count_values}{SummarizedExperiment}(x, svar = "subgroup", ...)
}
\arguments{
\item{x}{character(n) factor(n), or SummarizedExperiment with composite (s)values.}

\item{...}{used for S3 method dispatch}

\item{svar}{character(1)}
}
\value{
integer
}
\description{
Reshapes a vector with composite values into a 2d matrix:
   + rownames = values of component 1:(n-1)
   + colnames = values of component n
   + values = frequency counts of composite levels
}
\examples{
require(magrittr)

# STEM CELL DIFFERENTIATION
if (require(autonomics.data)){
   x <- autonomics.data::stemdiff.proteinratios
   x \%>\% autonomics.import::subgroup_values() \%>\% autonomics.import::count_values()
   x \%>\%                                          autonomics.import::count_values()
}

# ATKIN
if (require(atkin.2014)){
   x <- atkin.2014::soma.2018
   x \%>\% autonomics.import::subgroup_values() \%>\% autonomics.import::count_values()
   x \%>\%                                          autonomics.import::count_values()
}
}
