% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters.R
\name{counts}
\alias{counts}
\alias{counts,SummarizedExperiment-method}
\alias{counts<-}
\alias{counts<-,SummarizedExperiment,matrix-method}
\alias{counts<-,SummarizedExperiment,numeric-method}
\title{Get/Set counts}
\usage{
counts(object)

\S4method{counts}{SummarizedExperiment}(object)

counts(object) <- value

\S4method{counts}{SummarizedExperiment,matrix}(object) <- value

\S4method{counts}{SummarizedExperiment,numeric}(object) <- value
}
\arguments{
\item{object}{SummarizedExperiment}

\item{value}{count matrix (features x samples)}
}
\value{
count matrix (get) or updated object (set)
}
\description{
Get / Set counts matrix
}
