% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rnaseq.R
\name{libsizes}
\alias{libsizes}
\alias{counts_to_cpm}
\alias{cpm_to_counts}
\title{Convert between counts and cpm (counts per million scaled reads)}
\usage{
libsizes(counts)

counts_to_cpm(counts, lib.size = libsizes(counts))

cpm_to_counts(cpm, lib.size)
}
\arguments{
\item{counts}{count matrix}

\item{lib.size}{scaled library sizes (vector)}
}
\value{
cpm matrix
}
\description{
Convert between counts and cpm (counts per million scaled reads)
}
\examples{
require(magrittr)
if (require(subramanian.2016)){
   counts <- 'extdata/rnaseq/gene_counts.txt' \%>\%
              system.file(package = 'subramanian.2016') \%>\%
              autonomics.import::load_rnaseq(infer_design_from_sampleids = TRUE) \%>\%
              autonomics.import::exprs()
   lib.size <- counts \%>\% autonomics.import::libsizes()
   cpm      <- counts \%>\% autonomics.import::counts_to_cpm(lib.size)
   counts2  <- cpm \%>\% autonomics.import::cpm_to_counts(lib.size)
   sum(counts-counts2)
}
}
