% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_design_matrix.R
\name{create_design_matrix}
\alias{create_design_matrix}
\title{Create design matrix for statistical analysis}
\usage{
create_design_matrix(object,
  intercept = length(unique(autonomics.import::sdata(object)$subgroup)) == 1,
  confounders = character(0))
}
\arguments{
\item{object}{eSet or eList}

\item{confounders}{confounder svars (character)}
}
\value{
design matrix
}
\description{
Create design matrix for statistical analysis
}
\examples{
require(magrittr)

# STEM CELL COMPARISON
if (require(autonomics.data)){
   object <- autonomics.data::stemdiff.proteinratios
   object \%>\% autonomics.import::create_design_matrix()
   object \%>\% autonomics.import::create_design_matrix(intercept = TRUE)
}
if (require(billing.differentiation.data)){
   billing.differentiation.data::rna.voomcounts \%>\%
   autonomics.import::create_design_matrix()      \%>\%
   magrittr::extract(1:3,)
}
}
