% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_design_file.R
\name{create_maxquant_design_df}
\alias{create_maxquant_design_df}
\alias{create_sample_design_df}
\title{Create maxquant design df}
\usage{
create_maxquant_design_df(proteingroups_file,
  value_type = autonomics.import::infer_maxquant_value_type(proteingroups_file),
  infer_design = FALSE)

create_sample_design_df(...)
}
\arguments{
\item{proteingroups_file}{string: full path to protein groups file}

\item{value_type}{string: any value in autonomics.import::MAXQUANT_VALUE_TYPES}

\item{infer_design}{logical: should design be infered from sample ids (see details)}

\item{...}{backward compatibility to deprecated functions}

\item{sample_file}{string: full path to sample design file}
}
\value{
sample design dataframe
}
\description{
For automated infer of design, use the following sample naming
scheme (before running max quant): WT(L).KD(M).OE(H).R1
}
\examples{
require(magrittr)

# LABELED RATIOS and INTENSITIES
if (require(autonomics.data)){
   file <- system.file(
      'extdata/stemcell.differentiation/maxquant/proteinGroups.txt',
       package = 'autonomics.data')
   file \%>\% create_maxquant_design_df()
   file \%>\% create_maxquant_design_df(infer_design = TRUE)
   file \%>\% create_maxquant_design_df(infer_design = TRUE, value_type = 'raw.intensity')
   file \%>\% create_maxquant_design_df(infer_design = TRUE, value_type = 'raw.ratio')
}

# LFQ INTENSITIES
if (require(graumann.lfq)){
   file <- system.file('extdata/proteinGroups.txt', package = 'graumann.lfq')
   file \%>\% autonomics.import::create_maxquant_design_df()
   file \%>\% autonomics.import::create_maxquant_design_df(infer_design = TRUE)
}

# UNLABELED INTENSITIES
if (require(alnoubi.2017)){
   file <- system.file('extdata/proteinGroups.txt', package = 'alnoubi.2017')
   file \%>\% autonomics.import::create_maxquant_design_df()
   file \%>\% autonomics.import::create_maxquant_design_df(infer_design = TRUE)
}

# REPORTER INTENSITIES
if (require(billing.vesicles)){
   file <- system.file('extdata/proteinGroups.txt', package = 'billing.vesicles')
   file \%>\% autonomics.import::create_maxquant_design_df()
   file \%>\% autonomics.import::create_maxquant_design_df(infer_design = TRUE)
}

}
\author{
Aditya Bhagwat
}
