% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_omics.R
\name{deconvolute_proteingroups}
\alias{deconvolute_proteingroups}
\title{Deconvolute proteingroups}
\usage{
deconvolute_proteingroups(object, fastafile, fastafields = c("GENES",
  "PROTEIN-NAMES", "EXISTENCE", "REVIEWED"), drop_isoform_info = FALSE)
}
\arguments{
\item{object}{SummerizedExperiment with proteinGroups data}

\item{fastafile}{path to fastafile}

\item{fastafields}{character vector: fields to load from fastafile}

\item{drop_isoform_info}{logical: whether to drop isoform info}
}
\value{
deconvoluted and annotated SummarizedExperiment
}
\description{
Deconvolute proteingroups
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   fastafile <- '../data/uniprot_hsa_20140515.fasta'
   if (file.exists(fastafile)){
      object <- 'extdata/stemcomp/maxquant/proteinGroups.txt'            \%>\%
                 system.file(package='autonomics.data')                  \%>\%
                 load_proteingroups(infer_design_from_sampleids = TRUE)
      autonomics.import::fdata(object) \%>\% head()

      object \%<>\% magrittr::extract(1:100, )                             \%>\%
                  deconvolute_proteingroups(fastafile = fastafile)
      autonomics.import::fdata(object) \%>\% head()
   }
}
}
