% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_omics.R
\name{download_gtf}
\alias{download_gtf}
\title{Download feature annotations}
\usage{
download_gtf(organism, release = 95,
  gtffile = sprintf("~/.autonomics/gtf/\%s",
  basename(make_gtf_link(organism, release))))
}
\arguments{
\item{organism}{'Homo sapiens', 'Mus musculus' or 'Rattus norvegicus'}

\item{release}{GTF release. By default release 95 selected}

\item{gtffile}{string: local GTF file path}
}
\description{
Download feature annotations in GTF format
}
\examples{
\dontrun{
   download_gtf(organism = 'Homo sapiens')
   download_gtf(organism = 'Mus musculus')
   download_gtf(organism = 'Rattus norvegicus')
}
}
