% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eset_specific_var.R
\name{ensg_var}
\alias{ensg_var}
\alias{ensg_values}
\alias{get_ensg_var}
\alias{get_ensg_values}
\title{Get ensg var/values}
\usage{
ensg_var(object)

ensg_values(object)

get_ensg_var(object)

get_ensg_values(object)
}
\arguments{
\item{object}{eset}
}
\description{
Get ensg var/values
}
\examples{
require(magrittr)
if (require(billing.differentiation.data)){
   object <- billing.differentiation.data::rna.voomcounts
   object \%>\% autonomics.import::ensg_var()
   object \%>\% autonomics.import::ensg_values()
}
}
