% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters.R
\docType{methods}
\name{exprs}
\alias{exprs}
\alias{exprs,SummarizedExperiment-method}
\alias{exprs,ExpressionSet-method}
\alias{exprs,EList-method}
\alias{exprs<-}
\alias{exprs<-,SummarizedExperiment,matrix-method}
\alias{exprs<-,SummarizedExperiment,numeric-method}
\alias{exprs<-,ExpressionSet,matrix-method}
\alias{exprs<-,EList,matrix-method}
\alias{exprs<-,EList,numeric-method}
\title{Get/Set exprs}
\usage{
exprs(object)

\S4method{exprs}{SummarizedExperiment}(object)

\S4method{exprs}{ExpressionSet}(object)

\S4method{exprs}{EList}(object)

exprs(object) <- value

\S4method{exprs}{SummarizedExperiment,matrix}(object) <- value

\S4method{exprs}{SummarizedExperiment,numeric}(object) <- value

\S4method{exprs}{ExpressionSet,matrix}(object) <- value

\S4method{exprs}{EList,matrix}(object) <- value

\S4method{exprs}{EList,numeric}(object) <- value
}
\arguments{
\item{object}{SummarizedExperiment, ExpressionSet, EList}

\item{value}{ratio matrix (features x samples)}
}
\value{
exprs matrix (get) or updated object (set)
}
\description{
Get / Set exprs matrix
}
\examples{
if (require(autonomics.data)){
   object <- autonomics.data::stemcomp.proteinratios
   exprs(object) <- 0
   object
}
}
