% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_limma.R
\name{extract_limma_dt}
\alias{extract_limma_dt}
\title{Extract limma datatable}
\usage{
extract_limma_dt(object)
}
\arguments{
\item{object}{SummarizedExperiment}
}
\value{
melted data.table
}
\description{
Extract limma datatable
}
\examples{
if (require(autonomics.data)){
   require(magrittr)
   autonomics.data::stemcomp.proteinratios \%>\% extract_limma_dt()
}
}
