% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_limma.R
\name{extract_limma_matrix}
\alias{extract_limma_matrix}
\alias{effect}
\alias{p}
\alias{fdr}
\title{Extract limma matrix from object}
\usage{
extract_limma_matrix(object, quantity)

effect(object)

p(object)

fdr(object)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{quantity}{'value', p', 'fdr', 'bonf'}
}
\value{
matrix (nfeature x ncontrast)
}
\description{
Extract limma matrix from object
}
\examples{
require(magrittr)
if (require(subramanian.2016)){
   object <- subramanian.2016::metabolon
   object \%>\% extract_limma_matrix(quantity = 'effect') \%>\% extract(1:3, 1:3)
   object \%>\% effect() \%>\% extract(1:3, 1:3)
   object \%>\% p()      \%>\% extract(1:3, 1:3)
   object \%>\% fdr()    \%>\% extract(1:3, 1:3)
}
}
