% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_limma.R
\name{extract_limma_matrix}
\alias{extract_limma_matrix}
\alias{coef}
\alias{p}
\alias{fdr}
\title{Extract coef matrix from object}
\usage{
extract_limma_matrix(object, quantity)

coef(object)

p(object)

fdr(object)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{quantity}{'coef', p', 'fdr', 'bonf'}
}
\value{
matrix (nfeature x ncontrast)
}
\description{
Extract coef matrix from object
}
\examples{
require(magrittr)
if (require(subramanian.2016)){
   object <- subramanian.2016::metabolon
   object \%>\% autonomics.import::coef()
   object \%>\% autonomics.import::p()
}
}
