% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_maxquant2.R
\name{extract_maxquant_exprs}
\alias{extract_maxquant_exprs}
\title{Extract maxquant exprs}
\usage{
extract_maxquant_exprs(file, what)
}
\arguments{
\item{file}{full path to proteinGroups.txt}

\item{what}{'ratios', 'normalized ratios', 'intensities', 'lfq intensities', or 'reporter intensities'}
}
\value{
character vector with sample names
}
\description{
Extract maxquant exprs
}
\examples{
 require(magrittr)
 if (require(autonomics.data)){
   file <- 'extdata/stemcell.comparison/maxquant/proteinGroups.txt' \%>\%
            system.file(package = 'autonomics.data')
   file \%>\% autonomics.import::extract_maxquant_exprs('intensities') \%>\% magrittr::extract(1:3, 1:3)
}
}
