% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_maxquant2.R
\name{extract_maxquant_intensity_colnames}
\alias{extract_maxquant_intensity_colnames}
\title{Extract maxquant intensity colnames}
\usage{
extract_maxquant_intensity_colnames(file, quantity = "Intensity")
}
\arguments{
\item{file}{full path to proteinGroups.txt}

\item{quantity}{'Intensity', 'LFQ intensity', 'Reporter intensity'}
}
\value{
character vector with intensity column names
}
\description{
Extract maxquant intensity colnames
}
\examples{
 require(magrittr)
 if (require(autonomics.data)){
   file <- 'extdata/stemcell.comparison/maxquant/proteinGroups.txt' \%>\%
            system.file(package = 'autonomics.data')
   file \%>\% autonomics.import::extract_maxquant_intensity_colnames() \%>\% head(3)
   file \%>\%
}
if (require(graumann.lfq)){
   file <- system.file('extdata/proteinGroups.txt', package = 'graumann.lfq')
   file \%>\% autonomics.import::extract_maxquant_intensity_colnames() \%>\% head(3)
   file \%>\% autonomics.import::extract_maxquant_intensity_colnames('LFQ intensity') \%>\% head(3)
}
if (require(billing.vesicles)){
   file <- system.file('extdata/proteinGroups.txt', package = 'billing.vesicles')
   file \%>\% autonomics.import::extract_maxquant_intensity_colnames('Reporter intensity') \%>\% head(3)
}
}
