% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_omics.R
\name{extract_rectangle}
\alias{extract_rectangle}
\title{Extract rectangle from omics datatable}
\usage{
extract_rectangle(dt, rows, cols, transpose = FALSE, drop = FALSE)
}
\arguments{
\item{dt}{datatable with raw omics data}

\item{rows}{row selector: logical(n), numeric(n), or character(n)}

\item{cols}{col selector: logical(n), numeric(n), or character(n)}

\item{transpose}{logical(1)}

\item{drop}{logical(1)}
}
\description{
Extract rectangle from omics datatable
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   file <- system.file('extdata/glutaminase/glutaminase.xlsx', package = 'autonomics.data')
   dt <- file \%>\% read_file(sheet = 2)
   dt \%>\% extract_rectangle(rows = 11:401, cols = 15:86)                    \%>\% extract(1:3, 1:3)           # exprs
   dt \%>\% extract_rectangle(rows = 11:401, cols = 5    )                    \%>\% extract(1:3, , drop=FALSE)  # fids
   dt \%>\% extract_rectangle(rows = 2,      cols = 15:86)                    \%>\% extract(, 1:3, drop=FALSE)  # sids
   dt \%>\% extract_rectangle(rows = 10:401, cols = 1:14 )                    \%>\% extract(1:3, 1:3)           # fdata
   dt \%>\% extract_rectangle(rows = 1:10,   cols = 14:86, transpose = TRUE)  \%>\% extract(1:3, 1:3)           # sdata
}
}
