% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eset_specific_var.R
\name{fid_var}
\alias{fid_var}
\alias{fid_values}
\title{Get feature id variable}
\usage{
fid_var(object)

fid_values(object)
}
\arguments{
\item{object}{SummarizedExperiment}
}
\description{
Get feature id variable
}
