% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_features_}
\alias{filter_features_}
\alias{filter_features}
\title{Filter features on condition}
\usage{
filter_features_(object, condition, verbose = FALSE)

filter_features(object, condition, verbose = FALSE)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{condition}{filter condition}

\item{verbose}{logical}
}
\value{
filtered eSet
}
\description{
Filter features on condition
}
\examples{
require(magrittr)
\dontrun{
if (require(autonomics.data)){
   autonomics.data::ALL \%>\%
   autonomics.import::filter_features(gene_symbols \%in\% c('LIG4', 'MAPK12', 'MAPK1') )

   autonomics.data::ALL \%>\%
   filter_features(gene_symbols \%in\% c('LIG4', 'MAPK12', 'MAPK1'), verbose = TRUE)

   autonomics.data::ALL \%>\%
   filter_features_("gene_symbols \%in\% c('LIG4', 'MAPK12', 'MAPK1')")

   autonomics.data::ALL \%>\%
   filter_features_("gene_symbols \%in\% c('LIG4', 'MAPK12', 'MAPK1')", verbose = TRUE)
}
}
}
