% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_features_}
\alias{filter_features_}
\alias{filter_features}
\title{Filter features on condition}
\usage{
filter_features_(x, condition)

filter_features(x, condition)
}
\arguments{
\item{x}{eSet}

\item{condition}{filter condition}
}
\value{
filtered eSet
}
\description{
Filter features on condition
}
\examples{
if (require(autonomics.data)){
   filter_features( ALL,  gene_symbols \%in\% c('LIG4', 'MAPK12', 'MAPK1') )
   filter_features_(ALL, "gene_symbols \%in\% c('LIG4', 'MAPK12', 'MAPK1')")
}
}
