% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_filter.R
\name{filter_features_min_expr}
\alias{filter_features_min_expr}
\title{Filter features with min expr for all samples}
\usage{
filter_features_min_expr(object, min_expr)
}
\arguments{
\item{object}{exprs object}

\item{min_expr}{minimum expression}
}
\description{
Filter features with min expr for all samples
}
