% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_samples_}
\alias{filter_samples_}
\alias{filter_samples}
\title{Filter samples on condition}
\usage{
filter_samples_(object, condition, verbose = FALSE)

filter_samples(object, condition, verbose = FALSE)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{condition}{filter condition}

\item{verbose}{logical}
}
\value{
filtered SummarizedExperiment
}
\description{
Filter samples on condition
}
\examples{
if (require(autonomics.data)){
   autonomics.import::filter_samples( autonomics.data::ALL,  sex == 'M')
   autonomics.import::filter_samples_(autonomics.data::ALL, "sex == 'M'")
}
}
