% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_samples_}
\alias{filter_samples_}
\alias{filter_samples}
\title{Filter samples on condition}
\usage{
filter_samples_(x, condition, verbose = FALSE)

filter_samples(x, condition, verbose = FALSE)
}
\arguments{
\item{x}{eSet}

\item{condition}{filter condition}

\item{verbose}{logical}
}
\value{
filtered eSet
}
\description{
Filter samples on condition
}
\examples{
if (require(autonomics.data)){
   filter_samples( ALL,  sex == 'M')
   filter_samples_(ALL, "sex == 'M'")
}
}
