% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters.R
\name{flevels}
\alias{flevels}
\title{Get fvar levels}
\usage{
flevels(object, fvar)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{fvar}{feature variable}
}
\value{
fvar values
}
\description{
Get fvar levels
}
\examples{
if (require(autonomics.data)){
   require(magrittr)
   autonomics.data::stemcomp.proteinratios \%>\% flevels('Gene names') \%>\% head()
}
}
