% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters.R
\name{fvalues}
\alias{fvalues}
\title{Get fvalues}
\usage{
fvalues(object, fvar)
}
\arguments{
\item{object}{SummarizedExperiment, eSet, or EList}

\item{fvar}{feature variable}
}
\value{
fvar values
}
\description{
Get fvar values
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   object <-autonomics.data::stemcomp.proteinratios
   object \%>\% fvalues('Gene names') \%>\% head()
   object \%>\% fvalues(NULL)
}
}
