% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_omics.R
\name{get_feature_counts}
\alias{get_feature_counts}
\title{Generates feature counts text file}
\usage{
get_feature_counts(filedir, gtffile, paired_end = FALSE, ...)
}
\arguments{
\item{filedir}{string: path to SAM or BAM file directory (one SAM or BAM file per sample)}

\item{gtffile}{string: path to (organism specific) gtffile}

\item{paired_end}{TRUE or FALSE: paired end reads?}

\item{...}{passed to Rsubread::featureCounts}
}
\description{
Generates feature counts text file
}
\examples{
download .zip file from "https://bitbucket.org/graumannlab/billing.stemcells/downloads/rnaseq_example_data.zip" and unzip it under ~/.autonomics/
get_feature_counts( filedir = "~/.autonomics/rnaseq_example_data/comparison",
                    gtffile = "~/.autonomics/gtf/Homo_sapiens.GRCh38.95.gtf",
                    paired_end = TRUE)
}
