% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_omics.R
\name{get_feature_counts}
\alias{get_feature_counts}
\title{Generates feature counts text file}
\usage{
get_feature_counts(filedir, gtffile, paired_end = FALSE, ...)
}
\arguments{
\item{filedir}{string: path to SAM or BAM file directory (one SAM or BAM file per sample)}

\item{gtffile}{string: path to (organism specific) gtffile}

\item{paired_end}{TRUE or FALSE: paired end reads?}

\item{...}{passed to Rsubread::featureCounts}
}
\description{
Generates feature counts text file
}
\examples{
\dontrun{ # requires internet and does not always work: https://stackoverflow.com/questions/55532102

   # Download example data and unzip
     url <- "https://bitbucket.org/graumannlab/billing.stemcells/downloads/stemcomp.bamfiles.zip"
     destfile <- "~/.autonomics/stemcomp.bamfiles.zip"
     download.file(url, destfile = destfile)
     utils::unzip(destfile, exdir = '~/.autonomics')
     unlink(destfile)

   # Download gtf file
     gtffile <- download_gtf('Homo sapiens', 95)

   # Generate feature counts
     get_feature_counts(filedir = "~/.autonomics/rnaseq_example_data/comparison",
                        gtffile = gtffile,
                        paired_end = TRUE)
}
}
