% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_consistent_nas.R
\name{impute_around_zero}
\alias{impute_around_zero}
\title{Impute around zero}
\usage{
impute_around_zero(x)
}
\arguments{
\item{x}{exprs matrix}
}
\value{
exprs matrix
}
\description{
Impute around zero
}
\examples{
require(magrittr)
x <- matrix(c(20, 21, 22,
              30, 31, 32,
              NA, NA, NA,
              40, NA, 41,
              NA, NA, NA,
              NA, 50, NA), ncol=3, byrow=TRUE)
x \%>\% impute_around_zero()
object <- SummarizedExperiment::SummarizedExperiment(
  assays = list(exprs = x))
object \%>\% autonomics.import::impute_common_nas(
  imputefun = impute_around_zero) \%>\%
  autonomics.import::exprs()
object$subgroup <- rep('EV', 3)
object \%>\% autonomics.import::impute_consistent_nas(
  imputefun = impute_around_zero) \%>\%
  autonomics.import::exprs()
}
