% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_design.R
\name{infer_design_from_sampleids}
\alias{infer_design_from_sampleids}
\title{Convert sample IDs into sample design}
\usage{
infer_design_from_sampleids(sample_ids)
}
\arguments{
\item{sample_ids}{character vector with sample ids}
}
\description{
Convert sample IDs into sample design
}
\examples{
require(magrittr)
sample_ids <- c('PERM_NON.R1[H/L]', 'PERM_NON.R2[H/L]', 'PERM_NON.R3[H/L]', 'PERM_NON.R4[H/L]')
sample_ids \%>\% infer_design_from_sampleids()
if (require(subramanian.2016)){
   subramanian.2016::metabolon$sample_id \%>\%
   autonomics.import::infer_design_from_sampleids()
}
}
