% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_design.R
\name{infer_design_from_sampleids}
\alias{infer_design_from_sampleids}
\title{Infer design from sample ids}
\usage{
infer_design_from_sampleids(sample_ids, sep = sample_ids \%>\%
  autonomics.import::infer_design_sep(c(".", " ", "_")), maxquant = FALSE)
}
\arguments{
\item{sample_ids}{character vector}

\item{sep}{string: design separator}

\item{maxquant}{logical: maxquant sample ids?}
}
\description{
Infer design from sample ids
}
\examples{
require(magrittr)
sample_ids <- c("UT_10h_R1", "UT_10h_R2", "UT_10h_R3", "UT_10h_R4")
sample_ids \%>\% autonomics.import::infer_design_from_sampleids()

sample_ids <- c("E(L).EM(M).BM(H).R1[M/L]", "E(L).EM(M).BM(H).R1[H/L]")
sample_ids \%>\% infer_design_from_sampleids(maxquant=TRUE)
}
