% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rnaseq.R
\name{libsizes}
\alias{libsizes}
\title{Compute effective library sizes (using TMM)}
\usage{
libsizes(counts)
}
\arguments{
\item{counts}{counts matrix}
}
\value{
vector (nsample)
}
\description{
Compute effective library sizes (using TMM)
}
