% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_exiqon.R
\name{load_exiqon}
\alias{load_exiqon}
\title{Load exiqon data}
\usage{
load_exiqon(file, design_file = NULL, infer_design_from_sampleids = FALSE,
  design_sep = NULL)
}
\arguments{
\item{file}{exiqon xlsx file}

\item{design_file}{NULL or character (sample design file)}

\item{infer_design_from_sampleids}{logical: whether to infer design from sample ids}

\item{design_sep}{string: sample id separator}
}
\value{
SummarizedExperiment
}
\description{
Load exiqon data
}
\examples{
require(magrittr)
if (require(subramanian.2016)){

   # Load exiqon and infer design from sampleids
   #--------------------------------------------
   file <- system.file('extdata/exiqon/subramanian.2016.exiqon.xlsx',
                        package = 'subramanian.2016')
   file \%>\% autonomics.import::load_exiqon(infer_design_from_sampleids = TRUE)

   # Load exiqon with design file
   #-----------------------------
   design_file <- tempfile()
   file \%>\% write_design('exiqon', infer_design_from_sampleids = TRUE,
                                   design_file = design_file)
   file \%>\% load_exiqon(design_file = design_file) \%>\%
            autonomics.import::sdata() \%>\% magrittr::extract(1:3, 1:5)
}
}
