% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_exiqon.R
\name{load_exiqon_sdata}
\alias{load_exiqon_sdata}
\alias{load_exiqon_fdata}
\alias{load_exiqon}
\title{Load exiqon file}
\usage{
load_exiqon_sdata(file)

load_exiqon_fdata(file)

load_exiqon(file, rm_ref_genes = TRUE, rm_spike = TRUE)
}
\arguments{
\item{file}{path to exiqon txt file}

\item{rm_ref_genes}{whether to remove reference genes (logical)}

\item{rm_spike}{whether to remove spike-ins (logical)}
}
\value{
eset
}
\description{
Load exiqon file
}
\examples{
require(magrittr)
if (require(subramanian.2016)){
   file <- system.file('extdata/exiqon/subramanian.2016.exiqon.xlsx',
                        package = 'subramanian.2016')
   file \%>\% autonomics.import::load_exiqon()
   file \%>\% autonomics.import::load_exiqon_sdata() \%>\% head(1)
   file \%>\% autonomics.import::load_exiqon_fdata() \%>\% head(1)
}
}
