% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_omics.R
\name{load_exiqon}
\alias{load_exiqon}
\title{Load exiqon data}
\usage{
load_exiqon(file, design_file = NULL, log2_transform = TRUE,
  log2_offset = 0, infer_design_from_sampleids = FALSE, design_sep = NULL,
  rm_ref_features = TRUE, rm_spike_features = TRUE)
}
\arguments{
\item{file}{exiqon xlsx file}

\item{design_file}{NULL or character (sample design file)}

\item{log2_transform}{logical: whether to log2 transform}

\item{log2_offset}{offset in mapping x -> log2(offset + x)}

\item{infer_design_from_sampleids}{logical: whether to infer design from sample ids}

\item{design_sep}{string: sample id separator}

\item{rm_ref_features}{logical}

\item{rm_spike_features}{logical}
}
\value{
SummarizedExperiment
}
\description{
Load exiqon data
}
\examples{
require(magrittr)
if (require(subramanian.2016)){

   # Load exiqon
   file <- system.file('extdata/exiqon/subramanian.2016.exiqon.xlsx',
                        package = 'subramanian.2016')
   file \%>\% load_exiqon(infer_design_from_sampleids = TRUE)

   # Load exiqon with design file
   design_file <- tempfile()
   file \%>\% write_design('exiqon', infer_design_from_sampleids = TRUE,
                                   design_file = design_file)
   file \%>\% load_exiqon(design_file = design_file) \%>\%
            autonomics.import::sdata() \%>\% magrittr::extract(1:3, 1:5)
}
}
