% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load2_exprs.R
\name{load_exprs}
\alias{load_exprs}
\title{Load exprs}
\usage{
load_exprs(file, platform, sheet = NULL, quantity = NULL)
}
\arguments{
\item{file}{path to omics data file}

\item{platform}{'exiqon', 'maxquant', 'metabolon', 'metabolonlipids', 'soma'}

\item{sheet}{excel sheet number or name (if applicable)}

\item{quantity}{string: quantity to be loaded as exprs (if applicable)}
}
\value{
sample dataframe
}
\description{
Load exprs
}
\examples{

# EXIQON
require(magrittr)
if (require(subramanian.2016)){
   file <- system.file('extdata/exiqon/subramanian.2016.exiqon.xlsx',
                        package = 'subramanian.2016')
   file \%>\% autonomics.import::load_exprs('exiqon')
}

# METABOLON
require(magrittr)
if (require(autonomics.data)){
   file <- system.file('extdata/glutaminase/glutaminase.xlsx', package = 'autonomics.data')
   file \%>\% load_exprs('metabolon', sheet = 2) \%>\% extract(1:3, 1:3)
}

# METABOLON LIPIDS
file <- '../../datasets/WCQA-01-18MLCLP-1/WCQA-01-18MLCLP CLP  6-TAB FILE (180710).XLSX'
if (file.exists(file)){
   file \%>\% load_exprs('metabolonlipids', sheet = 'Lipid Class Concentrations') \%>\%
            extract(1:3, 1:3)
}

# RNASEQ
 if (require(subramanian.2016)){
    file <- system.file('extdata/rnaseq/gene_counts.txt', package = 'subramanian.2016')
    file \%>\% autonomics.import::load_exprs('rnaseq')
 }
}
