% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load___fdata.R
\name{load_fdata}
\alias{load_fdata}
\title{Load fdata}
\usage{
load_fdata(file, platform, sheet = NULL)
}
\arguments{
\item{file}{path to omics data file}

\item{platform}{'exiqon', 'maxquant', 'metabolon', 'metabolonlipids', 'soma'}

\item{sheet}{excel sheet number or name if applicable}
}
\value{
sample dataframe
}
\description{
Load fdata
}
\examples{
 require(magrittr)

# EXIQON
if (require(subramanian.2016)){
   file <- system.file('extdata/exiqon/subramanian.2016.exiqon.xlsx',
                        package = 'subramanian.2016')
   file \%>\% autonomics.import::load_fdata('exiqon')
}

# METABOLON
if (require(autonomics.data)){
   file <- system.file('extdata/glutaminase/glutaminase.xlsx', package = 'autonomics.data')
   file \%>\% load_fdata(platform = 'metabolon', sheet = 2) \%>\% extract(1:3, 1:3)
}

# METABOLONLIPIDS
file <- '../../datasets/WCQA-01-18MLCLP-1/WCQA-01-18MLCLP CLP  6-TAB FILE (180710).XLSX'
if (file.exists(file)){
   file \%>\% load_fdata(platform = 'metabolonlipids',
                       sheet = 'Lipid Class Concentrations') \%>\% head()
}

# RNASEQ
 if (require(subramanian.2016)){
    file <- system.file('extdata/rnaseq/gene_counts.txt', package = 'subramanian.2016')
    file \%>\% autonomics.import::load_fdata('rnaseq')
 }
}
