% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load___fdata.R
\name{load_fdata_rnaseq}
\alias{load_fdata_rnaseq}
\title{Load rnaseq fdata}
\usage{
load_fdata_rnaseq(file)
}
\arguments{
\item{file}{path to exiqon xls file}
}
\value{
sample dataframe
}
\description{
Load rnaseq fdata
}
\examples{
 require(magrittr)
 if (require(subramanian.2016)){
    file <- system.file('extdata/rnaseq/gene_counts.txt', package = 'subramanian.2016')
    file \%>\% autonomics.import::load_fdata_rnaseq() \%>\% head()
 }
}
