% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load2_fdata.R, R/load_rnaseq.R
\name{load_fdata_rnaseq}
\alias{load_fdata_rnaseq}
\alias{load_fdata_rnaseq}
\title{Load rnaseq fdata}
\usage{
load_fdata_rnaseq(count_file, map_entrezg = FALSE, verbose = TRUE)

load_fdata_rnaseq(count_file, map_entrezg = FALSE, verbose = TRUE)
}
\arguments{
\item{count_file}{count file}

\item{map_entrezg}{whether to map features to their entrezg (logical)}

\item{verbose}{whether to report messages (logical)}

\item{file}{path to rnaseq data file}
}
\value{
feature dataframe
}
\description{
Load rnaseq fdata

Load fdata
}
\examples{
require(magrittr)
if (require(subramanian.2016)){
   file <- system.file('extdata/rnaseq/gene_counts.txt',
                        package = 'subramanian.2016')
   file \%>\% autonomics.import::load_fdata_exiqon()
}
if (require(billing.differentiation.data)){
   count_file <- system.file('extdata/rnaseq/gene_counts.txt',
                              package = 'billing.differentiation.data')
   load_fdata_rnaseq(count_file)
}
if (require(subramanian.2016)){
   count_file <- system.file('extdata/rnaseq/gene_counts.txt',
                              package = 'subramanian.2016')
}
}
