% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_metabolon.R
\name{load_metabolon}
\alias{load_metabolon}
\alias{load_metabolon_file}
\title{Load metabolon data}
\usage{
load_metabolon(file, sheet = 2, design_file = NULL, log2_transform = TRUE,
  infer_design_from_sampleids = FALSE, add_kegg_pathways = FALSE,
  add_smiles = FALSE)

load_metabolon_file(...)
}
\arguments{
\item{file}{metabolon xlsx file}

\item{sheet}{xls sheet name  or number}

\item{design_file}{NULL or character (sample design file)}

\item{log2_transform}{logical: whether to log2 transform}

\item{infer_design_from_sampleids}{logical: whether to infer design from sample ids}

\item{add_kegg_pathways}{logical: whether to add KEGG pathways to fdata}

\item{add_smiles}{logical: whether to add SMILES to fdata}

\item{...}{(backward compatibility)}
}
\value{
SummarizedExperiment (load_metabolon) or dataframe (load_metabolon_sdata, load_metabolon_fdata)
}
\description{
Load metabolon data
}
\examples{
require(magrittr)
if (require(autonomics.data)){

   # Loading metabolon file is easy
   file <- system.file('extdata/glutaminase/glutaminase.xlsx', package = 'autonomics.data')
   file \%>\% autonomics.import::load_metabolon()

   # Three ways to specify sample design
      # Use Group definition in metabolon file
      file \%>\% autonomics.import::load_metabolon() \%>\%
                         autonomics.import::sdata() \%>\% magrittr::extract(1:3, 1:5)
      # Infer from sample id values
      file \%>\% autonomics.import::load_metabolon(infer_design_from_sampleids = TRUE) \%>\%
                         autonomics.import::sdata() \%>\% magrittr::extract(1:3, 1:5)
      # Merge in from sample file
      design_file <- tempfile()
      file \%>\% autonomics.import::write_metabolon_design(design_file = design_file, infer_from_sampleids = TRUE)
      file \%>\% autonomics.import::load_metabolon(design_file = design_file) \%>\%
                         autonomics.import::sdata() \%>\% magrittr::extract(1:3, 1:5)
}
}
