% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_omics.R
\name{load_omics}
\alias{load_omics}
\title{Load omics}
\usage{
load_omics(file, sheet = 2, platform, log2_transform = TRUE,
  design_file = NULL, infer_design_from_sampleids = FALSE)
}
\arguments{
\item{file}{path to omics data file}

\item{sheet}{excel sheet number or name if applicable}

\item{platform}{'metabolon', 'metabolonlipids'}

\item{log2_transform}{logical}

\item{design_file}{path to design file}

\item{infer_design_from_sampleids}{logical}
}
\value{
sample dataframe
}
\description{
Load omics
}
\examples{
 require(magrittr)
 if (require(autonomics.data)){
    file <- system.file('extdata/glutaminase/glutaminase.xlsx', package = 'autonomics.data')
    file \%>\% load_omics(sheet=2, platform = 'metabolon')
 }
file <- '../../datasets/WCQA-01-18MLCLP-1/WCQA-01-18MLCLP CLP  6-TAB FILE (180710).XLSX'
if (file.exists(file)){
   file \%>\% load_omics(sheet = 'Lipid Class Concentrations', platform = 'metabolonlipids')
}
}
