% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load__omics.R
\name{load_omics}
\alias{load_omics}
\title{Load omics}
\usage{
load_omics(object, platform, sheet = 2, quantity = NULL,
  log2_transform = TRUE, log2_offset = 0, design_file = NULL,
  infer_design_from_sampleids = FALSE, design_sep = NULL)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{platform}{'exiqon', 'maxquant', 'metabolon', 'soma'}

\item{sheet}{excel sheet number or name if applicable}

\item{quantity}{string: which quantity to extract into exprs}

\item{log2_transform}{logical}

\item{log2_offset}{offset in mapping x -> log2(x+offset)}

\item{design_file}{path to design file}

\item{infer_design_from_sampleids}{logical}

\item{design_sep}{string: design separator}
}
\value{
sample dataframe
}
\description{
Load omics
}
