% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_load_maxquant.R
\name{load_phosphosite_occupancies}
\alias{load_phosphosite_occupancies}
\title{Load phosphosite occupancies}
\usage{
load_phosphosite_occupancies(phosphosites_file,
  proteingroups_file = paste0(dirname(phosphosites_file),
  "/proteinGroups.txt"), design_file = paste0(dirname(phosphosites_file),
  "/sample_design.txt"), parameter_file = paste0(dirname(phosphosites_file),
  "/parameters.txt"), log2_transform = TRUE,
  value_type = autonomics.import::infer_maxquant_value_type(phosphosites_file))
}
\arguments{
\item{phosphosites_file}{full path to "Phospho (STY)Sites.txt"}

\item{proteingroups_file}{full path to "proteinGroups.txt"}

\item{design_file}{full path to "sample_design.txt"}

\item{parameter_file}{full path to "parameters.txt"}

\item{log2_transform}{logical}

\item{value_type}{any value in autonomics.import::MAXQUANT_VALUE_TYPES}
}
\description{
Load phosphosite occupancies
}
\examples{
require(magrittr)
if (require(billing.differentiation.data)){
   'extdata/maxquant/Phospho (STY)Sites.txt'              \%>\%
    system.file(package = 'billing.differentiation.data') \%>\%
    autonomics.import::load_phosphosite_occupancies()
}
}
