% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_maxquant.R
\name{load_proteingroups}
\alias{load_proteingroups}
\title{Load proteingroup ratios/intensities}
\usage{
load_proteingroups(proteingroups_file,
  sample_file = paste0(dirname(proteingroups_file),
  "/sample_design.txt"),
  parameter_file = paste0(dirname(proteingroups_file),
  "/parameters.txt"), log2_transform = TRUE,
  value_type = autonomics.import::infer_maxquant_value_type(proteingroups_file))
}
\arguments{
\item{proteingroups_file}{full path to "proteinGroups.txt"}

\item{sample_file}{full path to "sample_design.txt"}

\item{parameter_file}{full path to "parameters.txt"}

\item{log2_transform}{logical}

\item{value_type}{any value in autonomics.import::MAXQUANT_VALUE_TYPES}
}
\value{
data.table or ExpressionSet
}
\description{
Load proteingroup ratios/intensities
}
\examples{
library(magrittr)
if (require(autonomics.data)){
   system.file('extdata/billing2016/proteinGroups.txt', package = 'autonomics.data') \%>\%
   autonomics.import::load_proteingroups()
}
if (require(billing.differentiation.data)){
   system.file('extdata/maxquant/proteinGroups.txt',
                package = 'billing.differentiation.data') \%>\%
   autonomics.import::load_proteingroups()
}
if (require(billing.vesicles)){
   system.file('extdata/proteinGroups.txt', package = 'billing.vesicles') \%>\%
   autonomics.import::load_proteingroups()
}
if (require(graumann.lfq)){
   system.file('extdata/proteinGroups.txt', package = 'graumann.lfq') \%>\%
   autonomics.import::load_proteingroups()
}
if (require(graumann.zebra)){
   system.file('extdata/proteinGroups.txt', package = 'graumann.zebra') \%>\%
   autonomics.import::load_proteingroups()
}
if (require(alnoubi.2017)){
   system.file('extdata/proteinGroups.txt', package = 'alnoubi.2017') \%>\%
   autonomics.import::load_proteingroups(value_type = 'raw.intensity')
}
}
