% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_maxquant2.R
\name{load_proteingroups2}
\alias{load_proteingroups2}
\title{Load proteingroups}
\usage{
load_proteingroups2(file,
  quantity = autonomics.import::infer_maxquant_quantity(file),
  infer_design_from_sampleids = FALSE, design_file = NULL,
  fasta_file = NULL, log2_transform = TRUE, log2_offset = 0)
}
\arguments{
\item{file}{full path to proteinGroups.txt}

\item{quantity}{'Ratio normalized', 'Ratio', 'Intensity', 'LFQ intensity', 'Reporter intensity'}

\item{design_file}{full path to design file (created with write_maxquant_design)}
}
\description{
Load proteingroups
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   file <- system.file('extdata/stemcell.comparison/maxquant/proteinGroups.txt',
                        package = 'autonomics.data')
   object <- file \%>\% autonomics.import::load_proteingroups2()
}
}
