% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_maxquant2.R
\name{load_proteingroups2}
\alias{load_proteingroups2}
\title{Load proteingroups}
\usage{
load_proteingroups2(file, quantity)
}
\arguments{
\item{file}{full path to proteinGroups.txt}

\item{quantity}{'Ratio normalized', 'Ratio', 'Intensity', 'LFQ intensity', 'Reporter intensity'}
}
\description{
Load proteingroups
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   file <- system.file('extdata/stemcell.comparison/maxquant/proteinGroups.txt',
                        package = 'autonomics.data')
   quantity <- 'Ratio normalized'
   file \%>\%
}
}
