% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_maxquant.R
\name{load_proteingroups_to_wide_dt}
\alias{load_proteingroups_to_wide_dt}
\title{Load proteingroups file into wide data.table}
\usage{
load_proteingroups_to_wide_dt(proteingroups_file,
  value_type = infer_maxquant_value_type(proteingroups_file))
}
\arguments{
\item{proteingroups_file}{path to proteinGroups file}

\item{value_type}{any value in names(VALUE_COLNAME_PATTERNS)}
}
\description{
Load proteingroups file into wide data.table
}
\examples{
require(magrittr)

# NORMALIZED RATIOS
if (require(billing.differentiation.data)){
   system.file('extdata/maxquant/proteinGroups.txt', package = 'billing.differentiation.data') \%>\%
      load_proteingroups_to_wide_dt() \%>\%
      str()
}

# REPORTER INTENSITIES
if (require(billing.vesicles)){
   system.file('extdata/proteinGroups.txt', package = 'billing.vesicles') \%>\%
      load_proteingroups_to_wide_dt() \%>\%
      str()
}

# RAW RATIOS
if (require(billing.differentiation.data)){
   system.file('extdata/maxquant/proteinGroups.txt', package = 'billing.differentiation.data') \%>\%
      load_proteingroups_to_wide_dt(value_type = 'raw.ratio') \%>\%
      str()
}

# LFQ INTENSITIES
if (require(graumann.lfq)){
   system.file('extdata/proteinGroups.txt', package = 'graumann.lfq') \%>\%
      load_proteingroups_to_wide_dt() \%>\%
      str()
}

# RAW INTENSITIES
if (require(alnoubi.2017)){
   system.file('extdata/proteinGroups.txt', package = 'alnoubi.2017') \%>\%
      load_proteingroups_to_wide_dt(value_type = 'raw.intensity') \%>\%
      str()
   system.file('extdata/proteinGroups.txt', package = 'alnoubi.2017') \%>\%
      load_proteingroups_to_wide_dt() \%>\%
      str()
}
}
