% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_omics.R
\name{load_rnaseq}
\alias{load_rnaseq}
\title{Load RNAseq}
\usage{
load_rnaseq(file, design_file = NULL, infer_design_from_sampleids = FALSE,
  design_sep = NULL, log2_transform = TRUE)
}
\arguments{
\item{file}{rnaseq counts file}

\item{design_file}{sample design file}

\item{infer_design_from_sampleids}{logical}

\item{design_sep}{character}

\item{log2_transform}{logical}
}
\description{
Load RNAseq
}
\examples{
require(magrittr)
if (require(autonomics.data){
   file <- system.file('extdata/stemdiff/rnaseq/gene_counts.txt', package = 'autonomics.data')
})
if (require(subramanian.2016)){
   file <- system.file('extdata/rnaseq/gene_counts.txt', package = 'subramanian.2016')
   infer_design_from_sampleids <- TRUE

}
}
