% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_omics.R
\name{load_sdata}
\alias{load_sdata}
\title{Load sdata}
\usage{
load_sdata(file, sheet = NULL, platform)
}
\arguments{
\item{file}{path to omics data file}

\item{sheet}{excel sheet number or name if applicable}

\item{platform}{'metabolon', 'metabolonlipids'}
}
\value{
sample dataframe
}
\description{
Load sdata
}
\examples{
 require(magrittr)
 if (require(autonomics.data)){
    file <- system.file('extdata/glutaminase/glutaminase.xlsx', package = 'autonomics.data')
    file \%>\% load_sdata(2, 'metabolon') \%>\% extract(1:3, 1:3)
 }
file <- '../../datasets/WCQA-01-18MLCLP-1/WCQA-01-18MLCLP CLP  6-TAB FILE (180710).XLSX'
if (file.exists(file)){
   file \%>\% load_sdata('Lipid Class Concentrations', 'metabolonlipids') \%>\% head()
}
}
