% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load2_sdata.R
\name{load_sdata}
\alias{load_sdata}
\title{Load sdata}
\usage{
load_sdata(file, platform, sheet = NULL, quantity = NULL)
}
\arguments{
\item{file}{path to omics data file}

\item{platform}{'exiqon', 'maxquant', 'metabolon', 'metabolonlipids', 'soma'}

\item{sheet}{excel sheet number or name if applicable}

\item{quantity}{string: which quantity should be extracted (only applicable for maxquant platform)}
}
\value{
sample dataframe
}
\description{
Load sdata
}
\examples{
require(magrittr)

# EXIQON
 if (require(subramanian.2016)){
    file <- system.file('extdata/exiqon/subramanian.2016.exiqon.xlsx',
                         package = 'subramanian.2016')
    file \%>\% autonomics.import::load_sdata('exiqon')
 }

# METABOLON
if (require(autonomics.data)){
   file <- system.file('extdata/glutaminase/glutaminase.xlsx', package = 'autonomics.data')
   file \%>\% load_sdata(platform = 'metabolon', sheet = 2) \%>\% extract(1:3, 1:3)
}

# METABOLONLIPIDS
file <- '../../datasets/WCQA-01-18MLCLP-1/WCQA-01-18MLCLP CLP  6-TAB FILE (180710).XLSX'
if (file.exists(file)){
   file \%>\% load_sdata(platform = 'metabolonlipids',
                       sheet = 'Lipid Class Concentrations') \%>\% head()
}
}
