% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load2_sdata.R
\name{load_sdata_rnaseq}
\alias{load_sdata_rnaseq}
\title{Load rnaseq sdata}
\usage{
load_sdata_rnaseq(file)
}
\arguments{
\item{file}{path to exiqon xls file}
}
\value{
sample dataframe
}
\description{
Load rnaseq sdata
}
\examples{
 require(magrittr)
 if (require(subramanian.2016)){
    file <- system.file('extdata/rnaseq/gene_counts.txt', package = 'subramanian.2016')
    file \%>\% autonomics.import::load_sdata_rnaseq()
 }
}
