% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load2_sdata.R
\name{load_snames_maxquant}
\alias{load_snames_maxquant}
\title{Load maxquant snames}
\usage{
load_snames_maxquant(file,
  quantity = autonomics.import::infer_maxquant_quantity(file),
  rm_quantity_from_snames = TRUE, infer_design_from_sampleids = FALSE)
}
\arguments{
\item{file}{path to maxquant file}

\item{quantity}{string: which quantity from maxquant table to use as exprs}

\item{rm_quantity_from_snames}{logical: whether to rm quantity from snames}

\item{infer_design_from_sampleids}{logical: whether to simplify snames for design inferral}
}
\description{
Load maxquant snames
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   file <- 'extdata/stemcell.comparison/maxquant/proteinGroups.txt' \%>\%
            system.file(package = 'autonomics.data')

   file \%>\% load_snames_maxquant()
   file \%>\% load_snames_maxquant(rm_quantity_from_snames = FALSE)
   file \%>\% load_snames_maxquant(infer_design_from_sampleids = TRUE)

   file \%>\% load_snames_maxquant('Intensity')
   file \%>\% load_snames_maxquant('Intensity', rm_quantity_from_snames = FALSE)
   file \%>\% load_snames_maxquant('Intensity', infer_design_from_sampleids = TRUE)
}
if (require(graumann.lfq)){
   file <- system.file('extdata/proteinGroups.txt', package = 'graumann.lfq')
   file \%>\% load_snames_maxquant('Intensity')
   file \%>\% load_snames_maxquant('Intensity', clean = TRUE)
   file \%>\% load_snames_maxquant('Intensity', clean = TRUE, designify = TRUE)
   file \%>\% load_snames_maxquant('LFQ intensity')
   file \%>\% load_snames_maxquant('LFQ intensity', clean = TRUE)
   file \%>\% load_snames_maxquant('LFQ intensity', clean = TRUE, designify = TRUE)
}
}
