% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_soma.R
\name{load_soma}
\alias{load_soma}
\title{Load soma data}
\usage{
load_soma(file, keepOnlyPasses = TRUE, keepOnlySamples = TRUE,
  verbose = FALSE)
}
\arguments{
\item{file}{adat file}

\item{keepOnlyPasses}{A logical value indicating whether or not to keep only the
rows and columns where the data quality was considered to be passable.}

\item{keepOnlySamples}{A logical value indicating whether or not to keep only the rows
containing actual samples (as opposed to QC, buffer, and calibrator
samples).}

\item{verbose}{Logical value indicating whether (lots of) diagnostic messages should be shown.}
}
\description{
Load soma data
}
\examples{
require(magrittr)
if (require(atkin.2014)){
   object <- system.file(
                'extdata/soma/WCQ-14-130_Set_A_RPT.HybMedNormCal_20140925.adat',
                 package = 'atkin.2014') \%>\%
             load_soma()
}
}
