% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_soma.R
\name{load_soma}
\alias{load_soma}
\title{Load soma data}
\usage{
load_soma(file, log2_transform = TRUE, keep_only_passes = TRUE,
  keep_only_samples = TRUE, infer_design = FALSE, rm_na_svars = TRUE,
  rm_single_value_svars = TRUE)
}
\arguments{
\item{file}{adat file}

\item{log2_transform}{logical: whether to log2 transform (logical)}

\item{keep_only_passes}{logical: whether to keep only features and samples that pass QC}

\item{keep_only_samples}{logical: whether to keep only biological samples (and rm QC, buffer, and calibrator samples).}

\item{infer_design}{logical: whethe to infer design from sample_ids}

\item{rm_na_svars}{logical: whether to rm na svars}

\item{rm_single_value_svars}{logical: whether to rm single value svars}
}
\description{
Loads data from soma file.
Extracts sample_id definitions from 'SampleId'    column
Extracts subgroup definitions  from 'SampleGroup' column when available.
When not available, attempts to extract subgroup definitions from 'sample_id'column.
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   file <- system.file('extdata/stemcell.comparison/stemcell.comparison.adat',
                        package = 'autonomics.data')
   file \%>\% autonomics.import::load_soma()
   file \%>\% autonomics.import::load_soma(infer_design = TRUE)
}

}
