% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_omics.R
\name{load_soma}
\alias{load_soma}
\title{Load soma}
\usage{
load_soma(file, design_file = NULL, infer_design_from_sampleids = FALSE,
  design_sep = NULL, log2_transform = TRUE, rm_sample_type = character(0),
  rm_feature_type = character(0), rm_sample_quality = character(0),
  rm_feature_quality = character(0), rm_na_svars = TRUE,
  rm_single_value_svars = TRUE)
}
\arguments{
\item{file}{string: path to adat file}

\item{design_file}{NULL or string:  path to sample file}

\item{infer_design_from_sampleids}{logical: whether to infer design from SampleId values}

\item{design_sep}{string: sample id separator}

\item{log2_transform}{logical: whether to log2 transform}

\item{rm_sample_type}{character vector: sample  types to be removed. Probably a subset of c('Sample', 'QC', 'Buffer', 'Calibrator').}

\item{rm_feature_type}{character vector: feature types to be removed. Probably a subset of c('Protein', 'Hybridization Control Elution', 'Rat Protein').}

\item{rm_sample_quality}{character vector: sample  qualities to be removed. Probably a subset of c('PASS', 'FLAG', 'FAIL')}

\item{rm_feature_quality}{character vector: feature qualities to be removed. Probably a subset of c('PASS', 'FLAG', 'FAIL')}

\item{rm_na_svars}{logical: whether to rm NA svars}

\item{rm_single_value_svars}{logical: whether to rm single value svars}
}
\value{
SummarizedExperiment
}
\description{
Loads data from soma file.
Extracts sample_id definitions from 'SampleId'    column
Extracts subgroup definitions  from 'SampleGroup' column when available.
When not available, attempts to extract subgroup definitions from 'sample_id'column.
}
\examples{
require(magrittr)
if (require(autonomics.data)){

   # Loading soma file is simple
   file <- system.file('extdata/stemcomp/soma/stemcomp.adat',
                        package = 'autonomics.data')
   file \%>\% autonomics.import::load_soma()

   # Three ways to specify sample design
      # Taken from 'SampleId' column in soma file
      file \%>\% autonomics.import::load_soma() \%>\% autonomics.import::sdata() \%>\% head()

      # Inferred from sample ids
      file \%>\% autonomics.import::load_soma(infer_design = TRUE) \%>\%
               autonomics.import::sdata() \%>\% head()

      # Specified through sample file
      design_file <- tempfile()
      file \%>\% write_design('soma', infer_design_from_sampleids = TRUE, design_file = design_file)
      file \%>\% load_soma(design_file = design_file) \%>\%
               autonomics.import::sdata() \%>\% head()
}
if (require(atkin.2014)){
   file <- system.file('extdata/soma/WCQ-14-130_Set_A_RPT.HybMedNormCal_20140925.adat',
                        package = 'atkin.2014')
   file \%>\% autonomics.import::load_soma()
   file \%>\% autonomics.import::load_soma(rm_feature_quality = 'FAIL')
}
}
\author{
Aditya Bhagwat
}
