% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_omics.R
\name{make_gtf_link}
\alias{make_gtf_link}
\title{Make link to GTF file}
\usage{
make_gtf_link(organism, release)
}
\description{
Make link to GTF file
}
\examples{
make_gtf_link(organism = 'Homo sapiens', release = 95)
make_gtf_link(organism = 'Mus musculus', release = 95)
}
