% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_maxquant.R
\name{melt_wide_maxquant_dt}
\alias{melt_wide_maxquant_dt}
\title{Tidy max quant table}
\usage{
melt_wide_maxquant_dt(DT, log2_transform = TRUE, log2_offset = 0)
}
\arguments{
\item{DT}{maxquant datatable}

\item{log2_transform}{logical}

\item{log2_offset}{offset added to value before log2 transformation ()}
}
\description{
Tidy max quant table
}
\examples{
require(magrittr)
if (require(billing.differentiation.data)){
   system.file('extdata/maxquant/proteinGroups.txt', package = 'billing.differentiation.data') \%>\%
      load_proteingroups_to_wide_dt() \%>\% melt_wide_maxquant_dt() \%>\% print()
}
if (require(graumann.lfq)){
  system.file('extdata/proteinGroups.txt', package = 'graumann.lfq') \%>\%
     load_proteingroups_to_wide_dt() \%>\% melt_wide_maxquant_dt() \%>\% print()
}
if (require(billing.vesicles)){
  system.file('extdata/proteinGroups.txt', package = 'billing.vesicles') \%>\%
     load_proteingroups_to_wide_dt() \%>\% melt_wide_maxquant_dt() \%>\% print()
}
}
