% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_maxquant.R
\name{melt_wide_maxquant_dt}
\alias{melt_wide_maxquant_dt}
\title{Melt wide maxquant datatable}
\usage{
melt_wide_maxquant_dt(DT, log2_transform = TRUE, log2_offset = 0)
}
\arguments{
\item{DT}{maxquant datatable}

\item{log2_transform}{logical}

\item{log2_offset}{offset added to value before log2 transformation ()}
}
\description{
Melt wide maxquant datatable
}
\examples{
#require(magrittr)
#if (require(autonomics.data)){
#   DT <- system.file('extdata/stemdiff/maxquant/proteinGroups.txt',
#                      package = 'autonomics.data') \%>\%
#         autonomics.import::load_proteingroups_to_wide_dt()
#   DT \%>\% melt_wide_maxquant_dt() \%>\% print()
#}
#if (require(graumann.lfq)){
#  system.file('extdata/proteinGroups.txt', package = 'graumann.lfq') \%>\%
#     load_proteingroups_to_wide_dt() \%>\% melt_wide_maxquant_dt() \%>\% print()
#}
#if (require(billing.vesicles)){
#  system.file('extdata/proteinGroups.txt', package = 'billing.vesicles') \%>\%
#     load_proteingroups_to_wide_dt() \%>\% melt_wide_maxquant_dt() \%>\% print()
#}
}
