% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eset_specific_var.R
\name{oraid_var}
\alias{oraid_var}
\alias{oraid_values}
\title{Get ora id var/values}
\usage{
oraid_var(object)

oraid_values(object)
}
\arguments{
\item{object}{eset}
}
\description{
Get ora id var/values
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   object <- 'extdata/stemdiff/rnaseq/gene_counts.txt' \%>\%
              system.file(package='autonomics.data')   \%>\%
              read_rnaseq(fid_var = 'gene_id')
   object \%>\% autonomics.import::oraid_var()
   object \%>\% autonomics.import::oraid_values() \%>\% head(3)
}
}
