% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_omics.R
\name{read_exiqon}
\alias{read_exiqon}
\title{Read exiqon}
\usage{
read_exiqon(file)
}
\arguments{
\item{file}{string: path to exiqon genex file}
}
\description{
Read exiqon
}
\examples{
if (require(subramanian.2016)){
   require(magrittr)
   file <- system.file('extdata/exiqon/subramanian.2016.exiqon.xlsx', package = 'subramanian.2016')
   file \%>\% read_exiqon()
}
}
