% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_omics.R
\name{read_gtf}
\alias{read_gtf}
\title{Read GTF file into data.table}
\usage{
read_gtf(gtffile, var = "gene_id", values = NULL, writefile = NULL)
}
\arguments{
\item{gtffile}{string: path to gtffile}

\item{var}{string: variable on which to filter}

\item{values}{filter variable for these values. If NULL, not filtering.}

\item{writefile}{string: file to write gtf table to}

<<<<<<< HEAD
\item{writefile}{data.table}
=======
\item{gtf}{data.table}
>>>>>>> 00976d4a30ac3c8ed217e690bbb7b4083800a493
}
\description{
Read GTF file into a data.table. Filter for particular values of variable.
}
\examples{
\dontrun{ # requires internet connection
   require(magrittr)
   gtffile <- download_gtf(organism = 'Homo sapiens', release = 95)
   gtfdt <- read_gtf(gtffile, var = 'gene_id', values = 'ENSG00000198947')
}
}
