% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_omics.R
\name{read_metabolon}
\alias{read_metabolon}
\title{Read data from metabolon file}
\usage{
read_metabolon(file, sheet = 2, fid_var = "COMP_ID",
  sid_var = "CLIENT_IDENTIFIER")
}
\arguments{
\item{file}{character(1)}

\item{sheet}{numeric(1) or character(1)}

\item{fid_var}{character(1): feature id variable}

\item{sid_var}{character(1): sample id variable}
}
\value{
SummarizedExperiment
}
\description{
Read data from metabolon file
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   file <- system.file('extdata/glutaminase/glutaminase.xlsx', package = 'autonomics.data')
   file \%>\% autonomics.import::read_metabolon()
}
}
