% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_omics.R
\name{read_metabolon}
\alias{read_metabolon}
\alias{load_metabolon}
\title{Read metabolon}
\usage{
read_metabolon(
  file,
  sheet = find_origscale_sheet(file),
  fid_var = "(COMP|COMP_ID)",
  sid_var = "(CLIENT_IDENTIFIER|Client ID)",
  subgroup_var = "Group",
  fname_var = "BIOCHEMICAL"
)

load_metabolon(file, sheet = 2, ...)
}
\arguments{
\item{file}{string: path to metabolon xlsx file}

\item{sheet}{number/string: xls sheet number or name}

\item{fid_var}{string: feature_id variable (ideally transcends dataset)}

\item{sid_var}{string: sample_id variable}

\item{subgroup_var}{string: subgroup variable (human comprehensible)}

\item{fname_var}{string: feature_name variable}

\item{...}{enable backward compatibility to deprecated load_metabolon}
}
\description{
Read metabolon
}
\examples{
if (require(autonomics.data)){
   require(magrittr)
   file <- 'extdata/glutaminase/glutaminase.xlsx' \%>\%
            system.file(package = 'autonomics.data')
   file \%>\% read_metabolon()
}
}
