% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_design.R
\name{read_metabolon_design}
\alias{read_metabolon_design}
\title{Read metabolon design}
\usage{
read_metabolon_design(design_file)
}
\arguments{
\item{design_file}{string: path to design file}
}
\value{
sample dataframe
}
\description{
Read metabolon design
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   design_file <- tempfile()
   system.file('extdata/glutaminase/glutaminase.xlsx',
                package = 'autonomics.data') \%>\%
   autonomics.import::write_metabolon_design(design_file, infer_from_sampleids = TRUE)
   design_file \%>\% read_metabolon_design()
}
}
