% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_omics.R
\name{read_rnaseq}
\alias{read_rnaseq}
\title{Read rnaseq counts}
\usage{
read_rnaseq(file, fid_var = "gene_id")
}
\arguments{
\item{file}{string: path to rnaseq counts file}

\item{fid_var}{string: feature id variable}
}
\description{
Read rnaseq counts
}
\examples{
if (require(autonomics.data)){
   require(magrittr)
   file <- 'extdata/stemdiff/rnaseq/gene_counts.txt' \%>\%
            system.file(package = 'autonomics.data')
   file \%>\% read_rnaseq()
}
}
\seealso{
merge_sdata, merge_fdata
}
