% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_omics.R
\name{read_somascan}
\alias{read_somascan}
\alias{load_soma}
\title{Read somascan}
\usage{
read_somascan(
  file,
  fid_var = "SeqId",
  sid_var = "SampleId",
  subgroup_var = "SampleGroup",
  fname_var = "EntrezGeneSymbol"
)

load_soma(file, ...)
}
\arguments{
\item{file}{string: path to *.adat file}

\item{fid_var}{string: feature_id   variable}

\item{sid_var}{string: sample_id    variable}

\item{subgroup_var}{string: subgroup     variable}

\item{fname_var}{string: feature_name variable}

\item{...}{provide backward compatibility to deprecated function load_soma}
}
\value{
Summarizedexperiment
}
\description{
Read data from somascan adat file
}
\examples{
if (require(autonomics.data)){
   require(magrittr)
   file <- system.file('extdata/stemcomp/soma/stemcomp.adat', package = 'autonomics.data')
   file \%>\% read_somascan()
}
}
\seealso{
prepare_somascan
}
