% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutators.R
\name{replace_nas_with_zeros}
\alias{replace_nas_with_zeros}
\title{Replace NAs with zeros}
\usage{
replace_nas_with_zeros(object, verbose = TRUE)
}
\arguments{
\item{object}{eSet}

\item{verbose}{TRUE or FALSE}
}
\value{
eset
}
\description{
Replace NAs with zeros
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   object <- autonomics.data::billing2016
   replace_nas_with_zeros(object)
}
if (require(billing.differentiation.data)){
   object <- billing.differentiation.data::rna.voomcounts
   object \%>\% autonomics.import::replace_nas_with_zeros()
}
}
