% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eset_specific_var.R
\name{sep}
\alias{sep}
\title{Get separator of collapsed fields}
\usage{
sep(object)
}
\arguments{
\item{object}{eset}
}
\value{
string with separator
}
\description{
Get separator of collapsed fields
}
\examples{
if (require(autonomics.data)){
   require(magrittr)
   object <- autonomics.data::stemcomp.proteinratios
   object \%>\% sep()
}
}
