% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters.R
\name{slevels}
\alias{slevels}
\alias{subgroup_levels}
\title{Get slevels}
\usage{
slevels(object, svar)

subgroup_levels(object)
}
\arguments{
\item{object}{SummarizedExperiment, eSet, or eList}

\item{svar}{sample var (character)}
}
\value{
svar values (character)
}
\description{
Get svar levels
}
\examples{
if (require(autonomics.data)){
   require(magrittr)
   autonomics.data::stemcomp.proteinratios \%>\% slevels('subgroup')
   autonomics.data::stemcomp.proteinratios \%>\% subgroup_levels()
}
}
