% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{split_by_svar}
\alias{split_by_svar}
\title{Split SummarizedExperiment by svar}
\usage{
split_by_svar(object, svar = "subgroup")
}
\arguments{
\item{object}{SummarizedExperiment}

\item{svar}{'subgroup'}
}
\value{
list
}
\description{
Split SummarizedExperiment by svar
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   autonomics.data::stemdiff.proteinratios \%>\%
   split_by_svar()
}
}
