% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xxx_components.R
\name{split_components}
\alias{split_components}
\alias{split_components.character}
\alias{subgroup_components}
\alias{scomponents}
\title{Split components}
\usage{
split_components(x, ...)

\method{split_components}{character}(values,
  sep = autonomics.import::guess_component_sep(values, verbose = FALSE))

subgroup_components(...)

scomponents(...)
}
\arguments{
\item{x}{character(n) or SummarizedExperiment}

\item{sep}{character(1): separator}

\item{svar}{character}
}
\value{
data.table with decomposed subgroup components
}
\description{
Get data.table in which subgroup components have been decomposed
}
\examples{
require(magrittr)
if (require(billing.differentiation.data)){
   billing.differentiation.data::protein.ratios \%>\% subgroup_values() \%>\% split_components()
   billing.differentiation.data::protein.ratios \%>\%                       split_components()
}
if (require(autonomics.data)){
   autonomics.data::glutaminase \%>\% subgroup_values()  \%>\% split_components()
   autonomics.data::glutaminase                            split_components()
}
if (require(subramanian.2016)){
   subramanian.2016::metabolon \%>\% subgroup_values()   \%>\% split_components()
   subramanian.2016::metabolon \%>\%                         split_components()
}
if (require(graumann.lfq)){
   graumann.lfq::lfq.intensities \%>\% subgroup_values() \%>\% split_components()
   graumann.lfq::lfq.intensities \%>\%                       split_components()
}
if (require(atkin.2014)){
   object <- 'extdata/soma/WCQ-18-007.hybNorm.plateScale.medNorm.calibrate.20181004.adat' \%>\%
              system.file(package='atkin.2014') \%>\%
              autonomics.import::load_soma()
   subgroup_values <- object \%>\% subgroup_values() \%>\% as.character() \%>\% unique() \%>\% split_components()
   subgroup_values <- object \%>\%                                                       split_components()
}
}
