% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xxx_values.R
\name{split_values}
\alias{split_values}
\alias{split_values.character}
\alias{split_values.factor}
\alias{split_values.SummarizedExperiment}
\alias{scomponents}
\alias{subgroup_components}
\title{Split composite (s)values}
\usage{
split_values(x, ...)

\method{split_values}{character}(x, sep = guess_sep(x, verbose = FALSE),
  keep = FALSE, ...)

\method{split_values}{factor}(x, sep = guess_sep(x, verbose = FALSE),
  keep = FALSE, ...)

\method{split_values}{SummarizedExperiment}(x, svar = "subgroup",
  keep = FALSE, ...)

scomponents(...)

subgroup_components(...)
}
\arguments{
\item{x}{character(n), factor(n), or SummarizedExperiment}

\item{...}{used for proper S3 method dispatch}

\item{sep}{character(1):  separator}

\item{keep}{logical(1_: whether to keep original values (in first column)}

\item{svar}{character(1)}
}
\value{
data.table(x, x1, x2, ...) or data.table(x1, x2, ...)
}
\description{
Get data.table in which composite (s)values have been split into components
}
\examples{
if (require(autonomics.data)){
   # GLUTAMINASE
      require(magrittr)
      x <- autonomics.data::glutaminase
      x \%>\% subgroup_values() \%>\% split_values()
      x \%>\%                       split_values()
}
}
