% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eset_to_table.R
\name{sumexp_to_long_dt}
\alias{sumexp_to_long_dt}
\alias{eset_to_long_table}
\title{Convert sumexp into long datatable}
\usage{
sumexp_to_long_dt(object, fid = autonomics.import::fid_var(object),
  fvars = character(0), sid = "sample_id", svars = character(0))

eset_to_long_table(...)
}
\arguments{
\item{object}{sumexp}

\item{fid}{fvar carrying feature id}

\item{fvars}{additional fvars to include in table}

\item{sid}{svar carrying sample id}

\item{svars}{additional svars to include in table}

\item{...}{(backward compatibility)}
}
\value{
long data.table
}
\description{
Convert sumexp into long datatable
}
