% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters.R
\docType{methods}
\name{svalues}
\alias{svalues}
\alias{subgroup_values}
\alias{sampleid_values}
\alias{svalues<-}
\alias{svalues<-,SummarizedExperiment,character-method}
\title{Get/Set svalues}
\usage{
svalues(object, svar)

subgroup_values(object)

sampleid_values(object)

svalues(object, svar) <- value

\S4method{svalues}{SummarizedExperiment,character}(object, svar) <- value
}
\arguments{
\item{object}{SummarizedExperiment, eSet, or eList}

\item{svar}{sample var (character)}

\item{value}{value vector}
}
\description{
Get/Set svar values
}
\examples{
if (require(autonomics.data)){
   require(magrittr)
   object <- autonomics.data::stemcomp.proteinratios
   object \%>\% svalues('subgroup')
   object \%>\% subgroup_values()
}
}
