% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eset_specific_var.R
\name{uniprot_var}
\alias{uniprot_var}
\title{Get uniprot var}
\usage{
uniprot_var(object)
}
\arguments{
\item{object}{eset}
}
\value{
fvar with uniprot accessions
}
\description{
Get uniprot var
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   object <- 'extdata/stemcomp/maxquant/proteinGroups.txt' \%>\%
              system.file(package = 'autonomics.data')     \%>\%
              read_proteingroups()
   autonomics.import::fvars(object)


   object <- autonomics.data::stemcomp.proteinratios
   object \%>\% autonomics.import::uniprot_var()
}
}
