% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_design.R
\name{write_design}
\alias{write_design}
\title{Write design}
\usage{
write_design(object, subgroup_var = NULL, sep = object \%>\%
  autonomics.import::guess_sep(verbose = TRUE), file = NULL)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{subgroup_var}{character(1)}

\item{sep}{design separator}

\item{file}{character(1)}
}
\value{
design dataframe
}
\description{
Write design
}
\examples{
require(magrittr)

# METABOLON
if (require(autonomics.data)){
   object <- 'extdata/glutaminase/glutaminase.xlsx'     \%>\%
              system.file(package = 'autonomics.data')  \%>\%
              autonomics.import::read_metabolon()
   object \%>\% autonomics.import::write_design() \%>\% head()
   object \%>\% autonomics.import::write_design(subgroup_var = "Group   HMDB_ID") \%>\% head()
}

# SOMASCAN
if (require(autonomics.data)){
   object <- 'extdata/stemcomp/soma/stemcomp.adat'     \%>\%
              system.file(package = 'autonomics.data') \%>\%
              autonomics.import::read_somascan()
   object \%>\% autonomics.import::write_design()
   object \%>\% autonomics.import::write_design(subgroup_var = 'SampleGroup')
}

# EXIQON
if (require(subramanian.2016)){
   object <- 'extdata/exiqon/subramanian.2016.exiqon.xlsx'  \%>\%
              system.file(package = 'subramanian.2016')     \%>\%
              autonomics.import::read_exiqon()
   object \%>\% write_design() \%>\% head()
}

# PROTEINGROUPS
if (require(autonomics.data)){
   object <- 'extdata/stemcomp/maxquant/proteinGroups.txt' \%>\%
              system.file(package = 'autonomics.data')     \%>\%
              autonomics.import::read_proteingroups(simplify_snames = TRUE)
   object \%>\% write_design()
}

}
