% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_features}
\alias{write_features}
\title{Write features to file}
\usage{
write_features(object, file)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{file}{txt file where to write results to}
}
\description{
Write features to file
}
\examples{
require(magrittr)

# STEM CELL COMPARISON
if (require(autonomics.data)){
   object  <- autonomics.data::stemcomp.proteinratios
   file <- tempdir() \%>\% paste0('/features.txt') \%T>\% message()
   object \%>\% write_features(file)
}
}
