% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_design.R
\name{write_maxquant_design}
\alias{write_maxquant_design}
\alias{create_maxquant_design_df}
\alias{create_sample_design_df}
\alias{create_maxquant_design_file}
\alias{create_sample_design_file}
\title{Write maxquant design}
\usage{
write_maxquant_design(file,
  value_type = autonomics.import::infer_maxquant_value_type(file),
  infer_design_from_sampleids = FALSE, design_file = NULL)

create_maxquant_design_df(...)

create_sample_design_df(...)

create_maxquant_design_file(...)

create_sample_design_file(...)
}
\arguments{
\item{file}{string: full path to protein groups file}

\item{value_type}{string: any value in autonomics.import::MAXQUANT_VALUE_TYPES}

\item{infer_design_from_sampleids}{logical: should design be infered from sample ids (see details)}

\item{design_file}{string: full path to sample design file}

\item{...}{backward compatibility to deprecated functions}
}
\value{
sample design dataframe
}
\description{
For automated infer from sampleids, use the following sample naming
scheme (before running max quant): WT(L).KD(M).OE(H).R1
}
\examples{
require(magrittr)

# LABELED RATIOS and INTENSITIES
if (require(autonomics.data)){
   file <- system.file(
      'extdata/stemcell.differentiation/maxquant/proteinGroups.txt',
       package = 'autonomics.data')
   file \%>\% write_maxquant_design()
   file \%>\% write_maxquant_design(infer_design_from_sampleids = TRUE)
   file \%>\% write_maxquant_design(infer_design_from_sampleids = TRUE, value_type = 'raw.intensity')
   file \%>\% write_maxquant_design(infer_design_from_sampleids = TRUE, value_type = 'raw.ratio')
}

# LFQ INTENSITIES
if (require(graumann.lfq)){
   file <- system.file('extdata/proteinGroups.txt', package = 'graumann.lfq')
   file \%>\% autonomics.import::write_maxquant_design()
   file \%>\% autonomics.import::write_maxquant_design(infer_design_from_sampleids = TRUE)
}

# UNLABELED INTENSITIES
if (require(alnoubi.2017)){
   file <- system.file('extdata/proteinGroups.txt', package = 'alnoubi.2017')
   file \%>\% autonomics.import::write_maxquant_design()
   file \%>\% autonomics.import::write_maxquant_design(infer_design_from_sampleids = TRUE)
}

# REPORTER INTENSITIES
if (require(billing.vesicles)){
   file <- system.file('extdata/proteinGroups.txt', package = 'billing.vesicles')
   file \%>\% autonomics.import::write_maxquant_design()
   file \%>\% autonomics.import::write_maxquant_design(infer_design_from_sampleids = TRUE)
}

}
\author{
Aditya Bhagwat
}
